% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{AIC.varlse}
\alias{AIC.varlse}
\alias{AIC.vharlse}
\title{Akaike's Information Criterion of Multivariate Time Series Model}
\usage{
\method{AIC}{varlse}(object, ...)

\method{AIC}{vharlse}(object, ...)
}
\arguments{
\item{object}{Model fit}

\item{...}{not used}
}
\description{
Compute AIC of VAR(p), VHAR, BVAR(p), and BVHAR
}
\details{
Let \eqn{\tilde{\Sigma}_e} be the MLE
and let \eqn{\hat{\Sigma}_e} be the unbiased estimator (from \code{\link[=compute_cov]{compute_cov()}} and the member named \code{covmat}) for \eqn{\Sigma_e}.
Note that

\deqn{\tilde{\Sigma}_e = \frac{s - k}{n} \hat{\Sigma}_e}

Then

\deqn{AIC(p) = \log \det \Sigma_e + \frac{2}{s}(\text{number of freely estimated parameters})}

where the number of freely estimated parameters is \eqn{pm^2}.
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
