% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-forecast.R
\name{alpl}
\alias{alpl}
\alias{alpl.bvharcv}
\title{Evaluate the Density Forecast Based on Average Log Predictive Likelihood (APLP)}
\usage{
alpl(x, ...)

\method{alpl}{bvharcv}(x, ...)
}
\arguments{
\item{x}{Out-of-sample forecasting object to use}

\item{...}{Not used}
}
\description{
This function computes ALPL given forecasting of Bayesian models.
}
