% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.bvarmn}
\alias{autoplot.bvarmn}
\title{Residual Plot for Minnesota Prior VAR Model}
\usage{
\method{autoplot}{bvarmn}(object, hcol = "grey", hsize = 1.5, ...)
}
\arguments{
\item{object}{\code{bvarmn} object}

\item{hcol}{color of horizontal line = 0 (By default, grey)}

\item{hsize}{size of horizontal line = 0 (By default, 1.5)}

\item{...}{additional options for geom_point}
}
\description{
This function draww residual plot for covariance matrix of Minnesota prior VAR model.
}
