% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.bvharsp}
\alias{autoplot.bvharsp}
\title{Plot the Result of BVAR and BVHAR MCMC}
\usage{
\method{autoplot}{bvharsp}(
  object,
  type = c("coef", "trace", "dens", "area"),
  pars = character(),
  regex_pars = character(),
  ...
)
}
\arguments{
\item{object}{\code{bvharsp} object}

\item{type}{The type of the plot. Posterior coefficient (\code{"coef"}), Trace plot (\code{"trace"}), kernel density plot (\code{"dens"}), and interval estimates plot (\code{"area"}).}

\item{pars}{Parameter names to draw.}

\item{regex_pars}{Regular expression parameter names to draw.}

\item{...}{Other options for each \code{\link[bayesplot:MCMC-traces]{bayesplot::mcmc_trace()}}, \code{\link[bayesplot:MCMC-distributions]{bayesplot::mcmc_dens()}}, and \code{\link[bayesplot:MCMC-intervals]{bayesplot::mcmc_areas()}}.}
}
\value{
A ggplot object
}
\description{
Draw BVAR and BVHAR MCMC plots.
}
