% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-bayes.R
\name{autoplot.minnesota}
\alias{autoplot.minnesota}
\title{Density Plot for \code{minnesota} Object}
\usage{
\method{autoplot}{minnesota}(
  object,
  type = c("coef", "variance"),
  var_name = NULL,
  NROW = NULL,
  NCOL = NULL,
  ...
)
}
\arguments{
\item{object}{\code{minnesota} object}

\item{type}{Plot mean or variance. \code{"coef"} indicates VAR coefficients and \code{"variance"} for diagonal elements for Sigma (By default, coefficients).}

\item{var_name}{variable name (for coefficients)}

\item{NROW}{Numer of facet row}

\item{NCOL}{Numer of facet col}

\item{...}{not used}
}
\description{
Density Plot for \code{minnesota} Object
}
