% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.normaliw}
\alias{autoplot.normaliw}
\title{Residual Plot for Minnesota Prior VAR Model}
\usage{
\method{autoplot}{normaliw}(object, hcol = "grey", hsize = 1.5, ...)
}
\arguments{
\item{object}{A \code{normaliw} object}

\item{hcol}{color of horizontal line = 0 (By default, grey)}

\item{hsize}{size of horizontal line = 0 (By default, 1.5)}

\item{...}{additional options for geom_point}
}
\value{
A ggplot object
}
\description{
This function draws residual plot for covariance matrix of Minnesota prior VAR model.
}
