% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.summary.bvarmn}
\alias{autoplot.summary.bvarmn}
\title{Density Plot for Minnesota Prior VAR Model}
\usage{
\method{autoplot}{summary.bvarmn}(
  object,
  type = c("coef", "variance"),
  var_name = NULL,
  NROW = NULL,
  NCOL = NULL,
  ...
)
}
\arguments{
\item{object}{\code{summary.bvarmn} object}

\item{type}{Plot mean or variance. \code{"coef"} indicates VAR coefficients and \code{"variance"} for diagonal elements for Sigma (By default, coefficients).}

\item{var_name}{variable name (for coefficients)}

\item{NROW}{Numer of facet row}

\item{NCOL}{Numer of facet col}

\item{...}{not used}
}
\description{
This function draws density plot for coefficient matrices of Minnesota prior VAR model.
}
