% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.summary.bvharsp}
\alias{autoplot.summary.bvharsp}
\title{Plot the Heatmap of SSVS Coefficients}
\usage{
\method{autoplot}{summary.bvharsp}(object, ...)
}
\arguments{
\item{object}{\code{summary.bvharsp} object}

\item{...}{Other arguments passed on the \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}.}
}
\value{
A ggplot object
}
\description{
Draw heatmap for SSVS prior coefficients.
}
