% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.summary.normaliw}
\alias{autoplot.summary.normaliw}
\title{Density Plot for Minnesota Prior VAR Model}
\usage{
\method{autoplot}{summary.normaliw}(
  object,
  type = c("coef", "variance"),
  var_name,
  true_est = NULL,
  NROW = NULL,
  NCOL = NULL,
  ...
)
}
\arguments{
\item{object}{\code{summary.normaliw} object}

\item{type}{Plot mean or variance. \code{"coef"} indicates VAR coefficients and \code{"variance"} for diagonal elements for Sigma (By default, coefficients).}

\item{var_name}{Variable name (for coefficients)}

\item{true_est}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} True estimates to draw}

\item{NROW}{Numer of facet row}

\item{NCOL}{Numer of facet col}

\item{...}{not used}
}
\description{
This function draws density plot for coefficient matrices of Minnesota prior VAR model.
}
