% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{BIC.varlse}
\alias{BIC.varlse}
\alias{BIC.vharlse}
\title{Bayesian Information Criterion of Multivariate Time Series Model}
\usage{
\method{BIC}{varlse}(object, ...)

\method{BIC}{vharlse}(object, ...)
}
\arguments{
\item{object}{Model fit}

\item{...}{not used}
}
\description{
Compute BIC of VAR(p), VHAR, BVAR(p), and BVHAR
}
\details{
Let \eqn{\tilde{\Sigma}_e} be the MLE
and let \eqn{\hat{\Sigma}_e} be the unbiased estimator (from \code{\link[=compute_cov]{compute_cov()}} and the member named \code{covmat}) for \eqn{\Sigma_e}.
Note that

\deqn{\tilde{\Sigma}_e = \frac{s - k}{n} \hat{\Sigma}_e}

Then

\deqn{BIC(p) = \log \det \Sigma_e + \frac{\log s}{s}(\text{number of freely estimated parameters})}

where the number of freely estimated parameters is \eqn{pm^2}.
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}

Gideon Schwarz. (1978). \emph{Estimating the Dimension of a Model}. Ann. Statist. 6 (2) 461 - 464. \url{https://doi.org/10.1214/aos/1176344136}
}
