% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R, R/print-tuning.R
\name{bound_bvhar}
\alias{bound_bvhar}
\alias{print.boundbvharemp}
\alias{knit_print.boundbvharemp}
\title{Setting Empirical Bayes Optimization Bounds}
\usage{
bound_bvhar(
  bayes_spec = set_bvhar(),
  lower_sigma = rep(0.01, length(bayes_spec$sigma)),
  upper_sigma = rep(10, length(bayes_spec$sigma)),
  lower_lambda = 0.01,
  upper_lambda = Inf,
  lower_delta = rep(0.01, length(bayes_spec$delta)),
  upper_delta = rep(1, length(bayes_spec$delta)),
  lower_daily = NULL,
  upper_daily = NULL,
  lower_weekly = NULL,
  upper_weekly = NULL,
  lower_monthly = NULL,
  upper_monthly = NULL
)

\method{print}{boundbvharemp}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.boundbvharemp(x, ...)
}
\arguments{
\item{bayes_spec}{Bayes model specification}

\item{lower_sigma}{lower bound for sigma}

\item{upper_sigma}{upper bound for sigma}

\item{lower_lambda}{lower bound for lambda}

\item{upper_lambda}{upper bound for lambda}

\item{lower_delta}{lower bound for delta}

\item{upper_delta}{upper bound for delta}

\item{lower_daily}{lower bound for daily, default = \code{NULL} (BVHAR-L)}

\item{upper_daily}{upper bound for daily, default = \code{NULL} (BVHAR-L)}

\item{lower_weekly}{lower bound for weekly, default = \code{NULL} (BVHAR-L)}

\item{upper_weekly}{upper bound for weekly, default = \code{NULL} (BVHAR-L)}

\item{lower_monthly}{lower bound for monthly, default = \code{NULL} (BVHAR-L)}

\item{upper_monthly}{upper bound for monthly, default = \code{NULL} (BVHAR-L)}

\item{x}{\code{boundbvharemp} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\description{
This function sets lower and upper bounds for \code{\link[=set_bvar]{set_bvar()}}, \code{\link[=set_bvhar]{set_bvhar()}}, or \code{\link[=set_weight_bvhar]{set_weight_bvhar()}}.
}
