% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_design}
\alias{build_design}
\title{Build Design Matrix of VAR(p)}
\usage{
build_design(y, var_lag)
}
\arguments{
\item{y}{Matrix, time series data}

\item{var_lag}{VAR lag}
}
\description{
This function constructs design matrix of multivariate regression model formulation of VAR(p).
}
\details{
X0 is
\deqn{X_0 = [Y_p, \ldots, Y_1, 1]}
i.e. (n - p) x (mp + 1) matrix
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
