% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_xdummy}
\alias{build_xdummy}
\title{Construct Dummy design matrix for Minnesota Prior}
\usage{
build_xdummy(p, lambda, sigma, eps)
}
\arguments{
\item{p}{Integer, VAR lag. For VHAR, put 3.}

\item{lambda}{Double, tightness of the prior around a random walk or white noise}

\item{sigma}{Vector, standard error of each variable}

\item{eps}{Double, very small number}
}
\description{
Define dummy X observation to add for Minnesota moments.
}
\details{
Bańbura et al. (2010) defines dummy observation and augment to the original data matrix to construct Litterman (1986) prior.
}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}
}
