% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_y0}
\alias{build_y0}
\title{Build Y0 matrix in VAR(p)}
\usage{
build_y0(x, p, t)
}
\arguments{
\item{x}{Matrix, time series data}

\item{p}{VAR lag}

\item{t}{starting index to extract}
}
\description{
Build Y0 matrix in VAR(p)
}
\details{
Given data Y,
\deqn{Y0 = [y_t^T, y_{t + 1}^T, \ldots, y_{t + n - p - 1}^T]^T}
is the (n - p) x m matrix.

In case of Y0, t = p + 1.
This function is used when constructing X0.
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
