% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-ghosh.R, R/print-bvarghosh.R
\name{bvar_ghosh}
\alias{bvar_ghosh}
\alias{print.bvarghosh}
\alias{knit_print.bvarghosh}
\title{Fit Bayesian VAR(p) of Minnesota Prior}
\usage{
bvar_ghosh(y, p, U)

\method{print}{bvarghosh}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvarghosh(x, ...)
}
\arguments{
\item{y}{matrix, Time series data of which columns indicate the variables}

\item{p}{VAR lag}

\item{U}{Positive definite matrix. By default, identity matrix of dimension ncol(X0)}

\item{x}{\code{bvarghosh} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{bvarmn} object with
\item{\code{design}}{\eqn{X_0}}
\item{\code{y0}}{\eqn{Y_0}}
\item{\code{y}}{raw input}
\item{\code{p}}{lag of VAR: p}
\item{\code{m}}{Dimension of the data}
\item{\code{obs}}{Sample size used when training = \code{totobs} - \code{p}}
\item{\code{totobs}}{Total number of the observation}
\item{\code{process}}{Process: VAR}
\item{\code{call}}{Matched call}
\item{\code{mn_mean}}{Location of posterior matrix normal distribution}
\item{\code{mn_scale}}{First scale matrix of posterior matrix normal distribution}
\item{\code{iw_scale}}{Scale matrix of posterior inverse-wishart distribution}
\item{\code{iw_shape}}{Shape of posterior inverse-wishart distribution}
}
\description{
This function fits BVAR(p) with Ghosh et al. (2018) nonhierarchical prior.
}
\details{
In Ghosh et al. (2018), there are many models for BVAR such as hierarchical or non-hierarchical.
Among these, this function chooses the most simple non-hierarchical matrix normal prior in Section 3.1.

\deqn{B \mid \Sigma_e \sim MN(0, U^{-1}, \Sigma_e)}
\eqn{\Sigma_e \sim} flat
where U: precision matrix.

Then in VAR design equation (Y0 = X0 B + Z),
MN mean can be derived by
\deqn{\hat{B} = (X_0^T X_0 + U)^{-1} X_0^T Y_0}
and the MN scale matrix can be derived by
\deqn{\hat\Sigma_e = Y_0^T (I_s - X_0(X_0^T X_0 + U)^{-1} X_0^T) Y_0}
and IW shape by \eqn{s - m - 1}.

(MN: \href{https://en.wikipedia.org/wiki/Matrix_normal_distribution}{matrix normal}, IW: \href{https://en.wikipedia.org/wiki/Inverse-Wishart_distribution}{inverse-wishart}).
}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}

Ghosh, S., Khare, K., & Michailidis, G. (2018). \emph{High-Dimensional Posterior Consistency in Bayesian Vector Autoregressive Models}. Journal of the American Statistical Association, 114(526). \url{https://doi:10.1080/01621459.2018.1437043}
}
