% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-hierarchical.R
\name{bvar_mixture}
\alias{bvar_mixture}
\title{Fit Bayesian VAR(p) of Hierarchical Normal-Mixture Prior}
\usage{
bvar_mixture(
  y,
  p,
  iter = 1000L,
  type = c("wishart", "lasso", "t"),
  Psi = NULL,
  d = NULL
)
}
\arguments{
\item{y}{Time series data of which columns indicate the variables}

\item{p}{VAR lag}

\item{iter}{The number of iterations for Gibbs sampler}

\item{type}{Prior types ("wishart": default, "lasso": Bayesian Group Lasso, "t": Multivariate t-distribution). See the detail.}

\item{Psi}{Scale matrix for Wishart prior when \code{type = "wishart"}}

\item{d}{Degrees of freedom of Wishart prior becomes d + k when \code{type = "wishart"}. d > -(k + 1)}
}
\value{
\code{bvar_mixture} returns an object \code{bvarghosh} \link{class}.

It is a list with the following components:

\describe{
\item{design}{\eqn{X_0}}
\item{y0}{\eqn{Y_0}}
\item{y}{Raw input}
\item{p}{Lag of VAR}
\item{m}{Dimension of the data}
\item{obs}{Sample size used when training = \code{totobs} - \code{p}}
\item{totobs}{Total number of the observation}
\item{process}{Process: Ghosh}
\item{call}{Matched call}
\item{mn_mean}{Location of posterior matrix normal distribution}
\item{fitted.values}{Fitted values}
\item{residuals}{Residuals}
\item{mn_prec}{Precision matrix of posterior matrix normal distribution}
\item{iw_scale}{Scale matrix of posterior inverse-wishart distribution}
\item{iw_shape}{Shape of posterior inverse-wishart distribution}
}
}
\description{
This function fits BVAR(p) with Ghosh et al. (2018) hierarchical priors.
}
\details{
Ghosh et al. (2018) gives flat prior for residual matrix in BVAR.

Under this setting, there are many models such as hierarchical or non-hierarchical.
This function chooses hierarchical prior in Section 3.3.

Non-hierarchical prior should pre-specify the precision matrix of matrix normal,
while it has prior distribution in hierarchical model.

As in the paper, this function provides the three types of hyper-prior,

\enumerate{
\item Wishart prior: \eqn{U \sim W(\Psi, d + k)}, Psi and d prespecified
\item Bayesian Group Lasso
\item Multivariate t-Distribution
}
}
\references{
Ghosh, S., Khare, K., & Michailidis, G. (2018). \emph{High-Dimensional Posterior Consistency in Bayesian Vector Autoregressive Models}. Journal of the American Statistical Association, 114(526). \url{https://doi:10.1080/01621459.2018.1437043}
}
