% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-hierarchical.R
\name{bvar_niwhm}
\alias{bvar_niwhm}
\alias{print.bvarhm}
\alias{knit_print.bvarhm}
\title{Fitting Hierarchical Bayesian VAR(p)}
\usage{
bvar_niwhm(
  y,
  p,
  num_iter = 1000,
  num_warm = floor(num_iter/2),
  thinning = 1,
  bayes_spec = set_bvar(sigma = set_psi(), lambda = set_lambda()),
  scale_variance = 0.05,
  include_mean = TRUE,
  parallel = list(),
  verbose = FALSE
)

\method{print}{bvarhm}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvarhm(x, ...)
}
\arguments{
\item{y}{Time series data of which columns indicate the variables}

\item{p}{VAR lag}

\item{num_iter}{MCMC iteration number}

\item{num_warm}{Number of warm-up (burn-in). Half of the iteration is the default choice.}

\item{thinning}{Thinning every thinning-th iteration}

\item{bayes_spec}{A BVAR model specification by \code{\link[=set_bvar]{set_bvar()}}.}

\item{scale_variance}{Proposal distribution scaling constant to adjust an acceptance rate}

\item{include_mean}{Add constant term (Default: \code{TRUE}) or not (\code{FALSE})}

\item{parallel}{List the same argument of \code{\link[optimParallel:optimParallel]{optimParallel::optimParallel()}}. By default, this is empty, and the function does not execute parallel computation.}

\item{verbose}{Print the progress bar in the console. By default, \code{FALSE}.}

\item{x}{\code{bvarhm} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{bvar_niwhm} returns an object named \code{bvarhm} \link{class}.
It is a list with the following components:

\describe{
\item{coefficients}{Coefficient Matrix}
\item{p}{Lag of VAR}
\item{m}{Dimension of the data}
\item{obs}{Sample size used when training = \code{totobs} - \code{p}}
\item{totobs}{Total number of the observation}
\item{call}{Matched call}
\item{type}{include constant term (\code{"const"}) or not (\code{"none"})}
\item{y0}{\eqn{Y_0}}
\item{design}{\eqn{X_0}}
\item{y}{Raw input}
}
}
\description{
This function fits hierarchical BVAR(p) with general Minnesota prior.
}
\references{
Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). doi:\href{https://doi:10.1002/jae.1137}{10.1002/jae.1137}

Giannone, D., Lenza, M., & Primiceri, G. E. (2015). \emph{Prior Selection for Vector Autoregressions}. Review of Economics and Statistics, 97(2). doi:\href{https://doi.org/10.1162/REST_a_00483}{10.1162/REST_a_00483}

Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. doi:\href{https://doi.org/10.2307/1391384}{10.2307/1391384}
}
