% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-sv.R
\name{bvar_sv}
\alias{bvar_sv}
\alias{print.bvarsv}
\alias{knit_print.bvarsv}
\title{Fitting Bayesian VAR-SV of Minnesota Belief}
\usage{
bvar_sv(
  y,
  p,
  num_iter = 1000,
  num_burn = floor(num_iter/2),
  thinning = 1,
  bayes_spec = set_bvar(),
  include_mean = TRUE,
  verbose = FALSE
)

\method{print}{bvarsv}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvarsv(x, ...)
}
\arguments{
\item{y}{Time series data of which columns indicate the variables}

\item{p}{VAR lag}

\item{num_iter}{MCMC iteration number}

\item{num_burn}{Number of burn-in (warm-up). Half of the iteration is the default choice.}

\item{thinning}{Thinning every thinning-th iteration}

\item{bayes_spec}{A BVAR model specification by \code{\link[=set_bvar]{set_bvar()}}.}

\item{include_mean}{Add constant term (Default: \code{TRUE}) or not (\code{FALSE})}

\item{verbose}{Print the progress bar in the console. By default, \code{FALSE}.}

\item{x}{\code{bvarsv} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} This function fits VAR-SV with Minnesota belief.
}
\details{
Cholesky stochastic volatility modeling for VAR based on
\deqn{\Sigma_t = L^T D_t^{-1} L}
}
\references{
Chan, J., Koop, G., Poirier, D., & Tobias, J. (2019). \emph{Bayesian Econometric Methods (2nd ed., Econometric Exercises)}. Cambridge: Cambridge University Press. doi:\href{https://doi.org/10.1017/9781108525947}{10.1017/9781108525947}

Cogley, T., & Sargent, T. J. (2005). \emph{Drifts and volatilities: monetary policies and outcomes in the post WWII US}. Review of Economic Dynamics, 8(2), 262–302. doi:\href{https://doi.org/10.1016/j.red.2004.10.009}{10.1016/j.red.2004.10.009}
}
