% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhar-package.R
\docType{package}
\name{bvhar-package}
\alias{bvhar-package}
\alias{_PACKAGE}
\alias{bvhar}
\title{bvhar: Bayesian VHAR model}
\description{
bvhar package models VAR, VHAR, BVAR, and Bayesian VHAR.
See introduction for `browseVignettes(package = "bvhar")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ygeunkim/bvhar}
  \item Report bugs at \url{https://github.com/ygeunkim/bvhar/issues}
}

}
\author{
\strong{Maintainer}: Young Geun Kim \email{dudrms33@g.skku.edu}

Other contributors:
\itemize{
  \item Changryong Baek [contributor]
}

}
\keyword{internal}
