% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhar-package.R
\docType{package}
\name{bvhar}
\alias{bvhar}
\alias{_PACKAGE}
\alias{bvhar-package}
\title{bvhar: Bayesian VHAR model}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Beta package to
        model VAR, VHAR, BVAR, and Bayesian VHAR.
}
\details{
bvhar package models VAR, VHAR, BVAR, and Bayesian VHAR.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ygeunkim/bvhar}
  \item Report bugs at \url{https://github.com/ygeunkim/bvhar/issues}
}

}
\author{
\strong{Maintainer}: Young Geun Kim \email{dudrms33@g.skku.edu}

}
\keyword{internal}
