% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhar-horseshoe.R, R/print-bvharsp.R
\name{bvhar_horseshoe}
\alias{bvhar_horseshoe}
\alias{print.bvharhs}
\alias{knit_print.bvharhs}
\title{Fitting Bayesian VHAR of Horseshoe Prior}
\usage{
bvhar_horseshoe(
  y,
  har = c(5, 22),
  num_iter = 1000,
  num_burn = floor(num_iter/2),
  thinning = 1,
  bayes_spec = set_horseshoe(),
  include_mean = TRUE,
  minnesota = c("no", "short", "longrun"),
  algo = c("block", "gibbs"),
  verbose = FALSE
)

\method{print}{bvharhs}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvharhs(x, ...)
}
\arguments{
\item{y}{Time series data of which columns indicate the variables}

\item{har}{Numeric vector for weekly and monthly order. By default, \code{c(5, 22)}.}

\item{num_iter}{MCMC iteration number}

\item{num_burn}{Number of burn-in (warm-up). Half of the iteration is the default choice.}

\item{thinning}{Thinning every thinning-th iteration}

\item{bayes_spec}{Horseshoe initialization specification by \code{\link[=set_horseshoe]{set_horseshoe()}}.}

\item{include_mean}{Add constant term (Default: \code{TRUE}) or not (\code{FALSE})}

\item{minnesota}{Minnesota type}

\item{algo}{Ordinary gibbs sampling (\code{"gibbs"}) or blocked gibbs (Default: \code{"block"}).}

\item{verbose}{Print the progress bar in the console. By default, \code{FALSE}.}

\item{x}{\code{bvharhs} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{bvhar_horseshoe} returns an object named \code{bvarhs} \link{class}.
It is a list with the following components:

\describe{
\item{phi_record}{MCMC trace for vectorized coefficients (alpha \eqn{\phi}) with \link[posterior:draws_df]{posterior::draws_df} format.}
\item{lambda_record}{MCMC trace for local shrinkage level (lambda \eqn{\lambda}) with \link[posterior:draws_df]{posterior::draws_df} format.}
\item{tau_record}{MCMC trace for global shrinkage level (tau \eqn{\tau}) with \link[posterior:draws_df]{posterior::draws_df} format.}
\item{psi_record}{MCMC trace for precision matrix (psi \eqn{\Psi}) with \link{list} format.}
\item{chain}{The numer of chains}
\item{coefficients}{Posterior mean of VHAR coefficients.}
\item{psi_posterior}{Posterior mean of precision matrix \eqn{\Psi}}
\item{covmat}{Posterior mean of covariance matrix}
\item{omega_record}{MCMC trace for diagonal element of \eqn{\Psi} (omega) with \link[posterior:draws_df]{posterior::draws_df} format.}
\item{eta_record}{MCMC trace for upper triangular element of \eqn{\Psi} (eta) with \link[posterior:draws_df]{posterior::draws_df} format.}
\item{param}{\link[posterior:draws_df]{posterior::draws_df} with every variable: alpha, lambda, tau, omega, and eta}
\item{df}{Numer of Coefficients: \verb{3m + 1} or \verb{3m}}
\item{p}{3 (The number of terms. It contains this element for usage in other functions.)}
\item{m}{Dimension of the data}
\item{obs}{Sample size used when training = \code{totobs} - \code{p}}
\item{totobs}{Total number of the observation}
\item{call}{Matched call}
\item{process}{Description of the model, e.g. \code{"VHAR_Horseshoe"}}
\item{type}{include constant term (\code{"const"}) or not (\code{"none"})}
\item{algo}{Usual Gibbs sampling (\code{"gibbs"}) or fast sampling (\code{"fast"})}
\item{spec}{Horseshoe specification defined by \code{\link[=set_horseshoe]{set_horseshoe()}}}
\item{iter}{Total iterations}
\item{burn}{Burn-in}
\item{thin}{Thinning}
\item{HARtrans}{VHAR linear transformation matrix}
\item{y0}{\eqn{Y_0}}
\item{design}{\eqn{X_0}}
\item{y}{Raw input}
}
}
\description{
This function fits VHAR with horseshoe prior.
}
\references{
Bai, R., & Ghosh, M. (2018). High-dimensional multivariate posterior consistency under global–local shrinkage priors. Journal of Multivariate Analysis, 167, 157–170. doi:\href{https://doi.org/10.1016/j.jmva.2018.04.010}{10.1016/j.jmva.2018.04.010}

Carvalho, C. M., Polson, N. G., & Scott, J. G. (2010). \emph{The horseshoe estimator for sparse signals}. Biometrika, 97(2), 465–480. doi:\href{https://doi.org/10.1093/biomet/asq017}{10.1093/biomet/asq017}

Makalic, E., & Schmidt, D. F. (2016). \emph{A Simple Sampler for the Horseshoe Estimator}. IEEE Signal Processing Letters, 23(1), 179–182. doi:\href{https://doi.org/10.1109/LSP.2015.2503725}{10.1109/lsp.2015.2503725}
}
