% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{choose_bayes}
\alias{choose_bayes}
\title{Finding the Set of Hyperparameters of Bayesian Model}
\usage{
choose_bayes(
  bayes_bound = bound_bvhar(),
  ...,
  eps = 1e-04,
  y,
  order = c(5, 22),
  include_mean = TRUE,
  parallel = list()
)
}
\arguments{
\item{bayes_bound}{Empirical Bayes optimization bound specification}

\item{...}{Additional arguments for \code{\link[stats:optim]{stats::optim()}}.}

\item{eps}{Hyperparameter \code{eps} is fixed. By default, \code{1e-04}.}

\item{y}{Time series data}

\item{order}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Order for BVAR or BVHAR. \code{p} of \code{\link[=bvar_minnesota]{bvar_minnesota()}} or \code{har} of \code{\link[=bvhar_minnesota]{bvhar_minnesota()}}. By default, \code{c(5, 22)} for \code{har}.}

\item{include_mean}{Add constant term (Default: \code{TRUE}) or not (\code{FALSE})}

\item{parallel}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} List the same argument of \code{\link[optimParallel:optimParallel]{optimParallel::optimParallel()}}. By default, this is empty, and the function does not execute parallel computation.}
}
\description{
This function chooses the set of hyperparameters of Bayesian model using \code{\link[stats:optim]{stats::optim()}} function.
}
\seealso{
Individual functions: \code{\link[=choose_bvar]{choose_bvar()}}
}
