% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{choose_bvar}
\alias{choose_bvar}
\title{Finding the Set of Hyperparameters of Bayesian Model}
\usage{
choose_bvar(
  sigma,
  lambda = 0.1,
  delta,
  lower,
  upper,
  ...,
  y,
  p,
  include_mean = TRUE
)
}
\arguments{
\item{sigma}{Initial vector for \code{sigma} (Default: \code{apply(y, 2, sd)})}

\item{lambda}{Initial value for \code{lambda} (Default: \code{.1})}

\item{delta}{Initial vector for \code{delta} (Default: \code{rep(0, ncol(y))})}

\item{lower}{Lower bound. By default, \code{-Inf}.}

\item{upper}{Upper bound. By default, \code{Inf}.}

\item{...}{Additional arguments for \code{\link[stats:optim]{stats::optim()}}.}

\item{y}{Time series data}

\item{p}{BVAR lag}

\item{include_mean}{Add constant term (Default: \code{TRUE}) or not (\code{FALSE})}
}
\description{
This function chooses the set of hyperparameters of Bayesian model using \code{\link[stats:optim]{stats::optim()}} function.
}
