% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{choose_sv}
\alias{choose_sv}
\title{Choose the Parameters Set of VAR-SV or VHAR-SV}
\usage{
choose_sv(y, ord, type = c("VAR", "VHAR"), include_mean = TRUE)
}
\arguments{
\item{y}{Time series data of which columns indicate the variables.}

\item{ord}{Order for VAR or VHAR.}

\item{type}{Model type (Default: \code{"VAR"} or \code{"VHAR"}).}

\item{include_mean}{Add constant term (Default: \code{TRUE}) or not (\code{FALSE}).}
}
\value{
\code{svinit} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} This function chooses parameters of Stochastic volatility using OLS.
}
\references{
Carriero, A., Chan, J., Clark, T. E., & Marcellino, M. (2022). \emph{Corrigendum to “Large Bayesian vector autoregressions with stochastic volatility and non-conjugate priors” [J. Econometrics 212 (1)(2019) 137–154]}. Journal of Econometrics, 227(2), 506-512.

Chan, J., Koop, G., Poirier, D., & Tobias, J. (2019). \emph{Bayesian Econometric Methods (2nd ed., Econometric Exercises)}. Cambridge: Cambridge University Press.
}
