% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{choose_var}
\alias{choose_var}
\title{Choose the Best VAR based on Information Criteria}
\usage{
choose_var(y, lag_max = 5, include_mean = TRUE, parallel = FALSE)
}
\arguments{
\item{y}{Time series data of which columns indicate the variables}

\item{lag_max}{Maximum Var lag to explore (default = 5)}

\item{include_mean}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Add constant term (Default: \code{TRUE}) or not (\code{FALSE})}

\item{parallel}{Parallel computation using \code{\link[foreach:foreach]{foreach::foreach()}}? By default, \code{FALSE}.}
}
\value{
Minimum order and information criteria values
}
\description{
This function computes AIC, FPE, BIC, and HQ up to p = \code{lag_max} of VAR model.
}
