% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-ghosh.R
\name{coef.bvarflat}
\alias{coef.bvarflat}
\title{Coefficients Method for \code{bvarflat} object}
\usage{
\method{coef}{bvarflat}(object, ...)
}
\arguments{
\item{object}{\code{bvarflat} object}

\item{...}{not used}
}
\description{
Matrix Normal mean of Minnesota BVAR
}
