% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-ghosh.R
\name{coef.bvarghosh}
\alias{coef.bvarghosh}
\title{Coefficients Method for \code{bvarghosh} object}
\usage{
\method{coef}{bvarghosh}(object, ...)
}
\arguments{
\item{object}{\code{bvarghosh} object}

\item{...}{not used}
}
\description{
Matrix Normal mean of Minnesota BVAR
}
