% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-minnesota.R
\name{coef.bvarmn}
\alias{coef.bvarmn}
\title{Coefficients Method for \code{bvarmn} object}
\usage{
\method{coef}{bvarmn}(object, ...)
}
\arguments{
\item{object}{\code{bvarmn} object}

\item{...}{not used}
}
\description{
Matrix Normal mean of Minnesota BVAR
}
