% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhar-minnesota.R
\name{coef.bvharmn}
\alias{coef.bvharmn}
\title{Coefficients Method for \code{bvharmn} object}
\usage{
\method{coef}{bvharmn}(object, ...)
}
\arguments{
\item{object}{\code{bvharmn} object}

\item{...}{not used}
}
\description{
Matrix Normal mean of Minnesota BVHAR
}
