% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member.R
\name{coef}
\alias{coef}
\alias{coef.varlse}
\alias{coef.vharlse}
\alias{coef.bvarmn}
\alias{coef.bvarflat}
\alias{coef.bvharmn}
\alias{coef.bvharsp}
\alias{coef.summary.bvharsp}
\title{Coefficient Matrix of Multivariate Time Series Models}
\usage{
\method{coef}{varlse}(object, ...)

\method{coef}{vharlse}(object, ...)

\method{coef}{bvarmn}(object, ...)

\method{coef}{bvarflat}(object, ...)

\method{coef}{bvharmn}(object, ...)

\method{coef}{bvharsp}(object, ...)

\method{coef}{summary.bvharsp}(object, ...)
}
\arguments{
\item{object}{Model object}

\item{...}{not used}
}
\value{
\link{matrix} object with appropriate dimension.
}
\description{
By defining \code{\link[stats:coef]{stats::coef()}} for each model, this function returns coefficient matrix estimates.
}
