% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vharlse.R
\name{coef.vharlse}
\alias{coef.vharlse}
\title{Coefficients Method for \code{vharlse} object}
\usage{
\method{coef}{vharlse}(object, ...)
}
\arguments{
\item{object}{varlse}

\item{...}{not used}
}
\description{
Coefficients Method for \code{vharlse} object
}
