% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlse.R
\name{coefficients.varlse}
\alias{coefficients.varlse}
\title{Coefficients Method for \code{varlse} object}
\usage{
\method{coefficients}{varlse}(object, ...)
}
\arguments{
\item{object}{\code{varlse} object}

\item{...}{not used}
}
\description{
Coefficients Method for \code{varlse} object
}
