% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_cov}
\alias{compute_cov}
\title{Covariance Estimate for Residual Covariance Matrix}
\usage{
compute_cov(z, num_design, dim_design)
}
\arguments{
\item{z}{Matrix, residual}

\item{num_design}{Integer, s = n - p}

\item{dim_design}{Ingeger, k = mp + 1}
}
\description{
Plausible estimator for residual covariance.
}
\details{
See Lütkepohl (2007).
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
