% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_covmse}
\alias{compute_covmse}
\title{Compute Forecast MSE Matrices}
\usage{
compute_covmse(object, step)
}
\arguments{
\item{step}{Integer, Step to forecast}

\item{`varlse`}{object}
}
\description{
Compute the forecast MSE matrices using VMA coefficients
}
\details{
See pp38 of Lütkepohl (2007).
Let \eqn{\Sigma} be the covariance matrix of VAR and let \eqn{W_j} be the VMA coefficients.
Recursively,
\deqn{\Sigma_y(1) = \Sigma}
\deqn{\Sigma_y(2) = \Sigma + W_1 \Sigma W_1^T}
\deqn{\Sigma_y(3) = \Sigma_y(2) + W_2 \Sigma W_2^T}
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
