% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_stablemat}
\alias{compute_stablemat}
\title{VAR(1) Representation of VAR(p)}
\usage{
compute_stablemat(object)
}
\arguments{
\item{object}{Model fit}
}
\description{
Compute the coefficient matrix of VAR(1) form
}
\details{
Each VAR(p) process can be represented by mp-dim VAR(1).

\deqn{Y_t = A Y_{t - 1} + C + U_t}

where

\deqn{
    A = \begin{bmatrix}
      A_1 & A_2 & \cdots A_{p - 1} & A_p \        \
      I_m & 0 & \cdots & 0 & 0 \                  \
      0 & I_m & \cdots & 0 & 0 \                  \
      \vdots & \vdots & \vdots & \vdots & \vdots \\
      0 & 0 & \cdots & I_m & 0
    \end{bmatrix}
}

\deqn{C = (c, 0, \ldots, 0)^T}

and

\deqn{U_t = (\epsilon_t, 0, \ldots, 0)^T}
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
