% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-sparse.R
\name{conf_fscore}
\alias{conf_fscore}
\alias{conf_fscore.ssvsmod}
\alias{conf_fscore.mvhsmod}
\title{Evaluate the Sparsity Estimation Based on F1 Score}
\usage{
conf_fscore(x, y, threshold, ...)

\method{conf_fscore}{ssvsmod}(x, y, threshold = 0.01, restrict = FALSE, ...)

\method{conf_fscore}{mvhsmod}(x, y, threshold = 0.01, ...)
}
\arguments{
\item{x}{Estimated model.}

\item{y}{True coefficient matrix.}

\item{threshold}{Threshold value indicating sparsity.}

\item{...}{not used}

\item{restrict}{Use restricted VAR. By default, \code{FALSE}.}
}
\description{
This function computes F1 score for sparse element of the true coefficients given threshold.
}
\details{
When using this function, the true coefficient matrix \eqn{\Phi} should be sparse.
If the element of the estimate \eqn{\hat\Phi} is smaller than some threshold,
it is treated to be zero.
Then the F1 score is computed by
\deqn{F_1 = \frac{2 precision \times recall}{precision + recall}}
}
