% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-sparse.R
\name{conf_prec}
\alias{conf_prec}
\alias{conf_prec.ssvsmod}
\alias{conf_prec.mvhsmod}
\title{Evaluate the Sparsity Estimation Based on Precision}
\usage{
conf_prec(x, y, threshold, ...)

\method{conf_prec}{ssvsmod}(x, y, threshold = 0.01, truth_thr = 0, restrict = FALSE, ...)

\method{conf_prec}{mvhsmod}(x, y, threshold = 0.01, truth_thr = 0, ...)
}
\arguments{
\item{x}{Estimated model.}

\item{y}{True coefficient matrix.}

\item{threshold}{Threshold value indicating sparsity.}

\item{...}{not used}

\item{truth_thr}{Threshold value when using non-sparse true coefficient matrix. By default, \code{0} for sparse matrix.}

\item{restrict}{Use restricted VAR. By default, \code{FALSE}.}
}
\description{
This function computes precision for sparse element of the true coefficients given threshold.
}
\details{
When using this function, the true coefficient matrix \eqn{\Phi} should be sparse.
If the element of the estimate \eqn{\hat\Phi} is smaller than some threshold,
it is treated to be zero.
Then the precision is computed by
\deqn{precision = \frac{TP}{TP + FP}}
where TP is true positive, and FP is false positive.
}
\references{
Bai, R., & Ghosh, M. (2018). High-dimensional multivariate posterior consistency under global–local shrinkage priors. Journal of Multivariate Analysis, 167, 157–170. doi:\href{https://doi.org/10.1016/j.jmva.2018.04.010}{10.1016/j.jmva.2018.04.010}
}
