% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-sparse.R
\name{conf_recall}
\alias{conf_recall}
\alias{conf_recall.summary.bvharsp}
\title{Evaluate the Sparsity Estimation Based on Recall}
\usage{
conf_recall(x, y, ...)

\method{conf_recall}{summary.bvharsp}(x, y, truth_thr = 0L, ...)
}
\arguments{
\item{x}{\code{summary.bvharsp} object.}

\item{y}{True inclusion variable.}

\item{...}{not used}

\item{truth_thr}{Threshold value when using non-sparse true coefficient matrix. By default, \code{0} for sparse matrix.}
}
\description{
This function computes recall for sparse element of the true coefficients given threshold.
}
\details{
Precision is computed by
\deqn{recall = \frac{TP}{TP + FN}}
where TP is true positive, and FN is false negative.
}
\references{
Bai, R., & Ghosh, M. (2018). High-dimensional multivariate posterior consistency under global–local shrinkage priors. Journal of Multivariate Analysis, 167, 157–170. doi:\href{https://doi.org/10.1016/j.jmva.2018.04.010}{10.1016/j.jmva.2018.04.010}
}
\seealso{
\code{\link[=confusion]{confusion()}}
}
