% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{est_energy}
\alias{est_energy}
\alias{est_energy_raw}
\alias{est_energy_rt}
\title{Estimated Energy Consumption}
\format{
\code{est_energy} is realized kernel variance dataset computed by parzen kernel.
Since \code{NI} and \code{PJM_Load} observations are too old, they are excluded.
The observations are subset of the raw data, from 2014-08-03 to 2018-08-01.
\describe{
\item{AEP}{\href{https://en.wikipedia.org/wiki/American_Electric_Power}{American Electric Power}}
\item{COMED}{\href{https://en.wikipedia.org/wiki/Commonwealth_Edison}{Commonwealth Edison}}
\item{DAYTON}{\href{https://en.wikipedia.org/wiki/DPL_Inc.}{The Dayton Power and Light Company}}
\item{DEOK}{\href{https://en.wikipedia.org/wiki/Duke_Energy}{Duke Energy Ohio/Kentucky}}
\item{DOM}{\href{https://en.wikipedia.org/wiki/Dominion_Energy}{Dominion Virginia Power}}
\item{DUQ}{\href{https://en.wikipedia.org/wiki/DQE}{Duquesne Light Co.}}
\item{EKPC}{\href{http://www.ekpc.coop/}{East Kentucky Power Cooperative}}
\item{FE}{\href{https://en.wikipedia.org/wiki/FirstEnergy}{FirstEnergy}}
\item{NI}{Northern Illinois Hub}
\item{PJME}{PJM East Region: 2001-2018}
\item{PJMW}{PJM West Region: 2001-2018}
\item{PJM_Load}{PJM Load Combined: 1998-2001}
}

\code{est_energy_raw} is a raw dataset that contains hourly date and corresponding energy consumption.
\describe{
\item{Datetime}{From 1998-04-01 10:00:00 to 2018-08-03 09:00:00}
}

\code{est_energy_rt} is a log-return dataset of \code{est_energy_raw}.
}
\source{
\url{https://www.kaggle.com/robikscube/time-series-forecasting-with-prophet/data}
}
\usage{
est_energy

est_energy_raw

est_energy_rt
}
\description{
US estimated energy consumption (in Megawatts) in each region.
}
\keyword{datasets}
