% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_bvar_mn}
\alias{estimate_bvar_mn}
\title{BVAR(p) Point Estimates based on Minnesota Prior}
\usage{
estimate_bvar_mn(x, y, x_dummy, y_dummy)
}
\arguments{
\item{x}{Matrix, X0}

\item{y}{Matrix, Y0}

\item{x_dummy}{Matrix, dummy X0}

\item{y_dummy}{Matrix, dummy Y0}
}
\description{
Point estimates for posterior distribution
}
\details{
Augment originally processed data and dummy observation.
OLS from this set gives the result.
}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}
}
