% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_ghosh_mn}
\alias{estimate_ghosh_mn}
\title{BVAR(p) Point Estimates based on Nonhierarchical Matrix Normal Prior}
\usage{
estimate_ghosh_mn(x, y, U)
}
\arguments{
\item{x}{Matrix, X0}

\item{y}{Matrix, Y0}

\item{U}{Positive definite matrix, covariance matrix corresponding to the column of the model parameter B}
}
\description{
Point estimates for Ghosh et al. (2018) nonhierarchical model for BVAR.
}
\details{
In Ghosh et al. (2018), there are many models for BVAR such as hierarchical or non-hierarchical.
Among these, this function chooses the most simple non-hierarchical matrix normal prior in Section 3.1.
}
\references{
Ghosh, S., Khare, K., & Michailidis, G. (2018). \emph{High-Dimensional Posterior Consistency in Bayesian Vector Autoregressive Models}. Journal of the American Statistical Association, 114(526). \url{https://doi:10.1080/01621459.2018.1437043}
}
