% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_har}
\alias{estimate_har}
\title{Compute Vector HAR Coefficient Matrices and Fitted Values}
\usage{
estimate_har(x, y)
}
\arguments{
\item{x}{X0 processed by \code{\link{build_design}}}

\item{y}{Y0 processed by \code{\link{build_y0}}}
}
\description{
Compute Vector HAR Coefficient Matrices and Fitted Values
}
\details{
Given Y0 and Y0, the function estimate least squares
Y0 = X1 Phi + Z
}
