% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_har_none}
\alias{estimate_har_none}
\title{Compute Vector HAR Coefficient Matrices and Fitted Values without Constant Term}
\usage{
estimate_har_none(x, y)
}
\arguments{
\item{x}{X0 processed by \code{\link{build_design}} (delete its last column)}

\item{y}{Y0 processed by \code{\link{build_y0}}}
}
\description{
This function fits VHAR given response and design matrices of multivariate time series, when the model has no constant term.
}
\details{
Given Y0 and Y0, the function estimate least squares
\deqn{Y_0 = X_1 \Phi + Z}
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}

Corsi, F. (2008). \emph{A Simple Approximate Long-Memory Model of Realized Volatility}. Journal of Financial Econometrics, 7(2), 174–196. \url{https://doi:10.1093/jjfinec/nbp001}
}
