% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_var}
\alias{estimate_var}
\title{Compute VAR(p) Coefficient Matrices and Fitted Values}
\usage{
estimate_var(x, y)
}
\arguments{
\item{x}{X0 processed by \code{\link[=build_design]{build_design()}}}

\item{y}{Y0 processed by \code{\link[=build_y0]{build_y0()}}}
}
\description{
This function fits VAR(p) given response and design matrices of multivariate time series.
}
\details{
Given Y0 and Y0, the function estimate least squares
Y0 = X0 A + Z
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
