% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etf_vix}
\alias{etf_vix}
\title{ETF volatility data}
\format{
A data frame of 1006 row and 9 columns
}
\usage{
etf_vix
}
\description{
Exchange Traded Funds (ETFs) volatility index from FRED.
Copyright, 2016, Chicago Board Options Exchange, Inc.
}
\keyword{datasets}
