% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etf_vix_raw}
\alias{etf_vix_raw}
\title{CBOE ETF Volatility Index Raw Dataset}
\format{
A data frmae of 1006 row and 10 columns,
including date column (\code{DATE})

From 2015-01-05 to 2018-12-28,
there exists 36 missing obervations.
}
\source{
Source: \url{https://www.cboe.com}

Release: \url{https://www.cboe.com/us/options/market_statistics/daily/}
}
\usage{
etf_vix_raw
}
\description{
Raw dataset for Chicago Board Options Exchage (CBOE) Exchange Traded Funds (ETFs) volatility index from FRED.
}
\details{
This dataset is included for convenience of usage of the CBOE ETF dataset when academic researching.
For the details of the data, see \link{etf_vix} documentation.
}
\keyword{datasets}
