% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-ghosh.R
\name{fitted.bvarflat}
\alias{fitted.bvarflat}
\title{Fitted Values Method for \code{bvarflat} object}
\usage{
\method{fitted}{bvarflat}(object, ...)
}
\arguments{
\item{object}{\code{bvarflat} object}

\item{...}{not used}
}
\description{
Fitted Values Method for \code{bvarflat} object
}
