% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-ghosh.R
\name{fitted.bvarghosh}
\alias{fitted.bvarghosh}
\title{Fitted Values Method for \code{bvarghosh} object}
\usage{
\method{fitted}{bvarghosh}(object, ...)
}
\arguments{
\item{object}{\code{bvarghosh} object}

\item{...}{not used}
}
\description{
Fitted Values Method for \code{bvarghosh} object
}
