% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-minnesota.R
\name{fitted.bvarmn}
\alias{fitted.bvarmn}
\title{Fitted Values Method for \code{bvarmn} object}
\usage{
\method{fitted}{bvarmn}(object, ...)
}
\arguments{
\item{object}{\code{bvarmn} object}

\item{...}{not used}
}
\description{
Fitted Values Method for \code{bvarmn} object
}
