% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhar-minnesota.R
\name{fitted.bvharmn}
\alias{fitted.bvharmn}
\title{Fitted Values Method for \code{bvharmn} object}
\usage{
\method{fitted}{bvharmn}(object, ...)
}
\arguments{
\item{object}{\code{bvharmn} object}

\item{...}{not used}
}
\description{
Fitted Values Method for \code{bvharmn} object
}
