% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member.R
\name{fitted}
\alias{fitted}
\alias{fitted.varlse}
\alias{fitted.vharlse}
\alias{fitted.bvarmn}
\alias{fitted.bvarflat}
\alias{fitted.bvharmn}
\title{Fitted Matrix from Multivariate Time Series Models}
\usage{
\method{fitted}{varlse}(object, ...)

\method{fitted}{vharlse}(object, ...)

\method{fitted}{bvarmn}(object, ...)

\method{fitted}{bvarflat}(object, ...)

\method{fitted}{bvharmn}(object, ...)
}
\arguments{
\item{object}{Model object}

\item{...}{not used}
}
\value{
\link{matrix} object.
}
\description{
By defining \code{\link[stats:fitted.values]{stats::fitted()}} for each model, this function returns fitted matrix.
}
