% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlse.R
\name{fitted.varlse}
\alias{fitted.varlse}
\title{Fitted Values Method for \code{varlse} object}
\usage{
\method{fitted}{varlse}(object, ...)
}
\arguments{
\item{object}{\code{varlse} object}

\item{...}{not used}
}
\description{
Fitted Values Method for \code{varlse} object
}
