% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vharlse.R
\name{fitted.vharlse}
\alias{fitted.vharlse}
\title{Fitted Values Method for \code{vharlse} object}
\usage{
\method{fitted}{vharlse}(object, ...)
}
\arguments{
\item{object}{varlse}

\item{...}{not used}
}
\description{
Fitted Values Method for \code{vharlse} object
}
