% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forecast_bvarmn_flat}
\alias{forecast_bvarmn_flat}
\title{Forecasting BVAR of Non-hierarchical Matrix Normal Prior}
\usage{
forecast_bvarmn_flat(object, step)
}
\arguments{
\item{object}{\code{bvarmn} object by \code{\link{bvar_minnesota}}}

\item{step}{Integer, Step to forecast}
}
\description{
Forecasting BVAR of Non-hierarchical Matrix Normal Prior
}
\details{
n-step ahead forecasting using VAR(p) recursively, based on pp35 of Lütkepohl (2007).
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}

Ghosh, S., Khare, K., & Michailidis, G. (2018). \emph{High-Dimensional Posterior Consistency in Bayesian Vector Autoregressive Models}. Journal of the American Statistical Association, 114(526). \url{https://doi:10.1080/01621459.2018.1437043}
}
