% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-forecast.R
\name{forecast_region}
\alias{forecast_region}
\title{Forecast Region for VAR(p)}
\usage{
forecast_region(object, h, level = 0.05, ...)
}
\arguments{
\item{object}{predvarlse object}

\item{level}{Confidence level}

\item{...}{not used}
}
\description{
Compute forecast region VAR(p) model
}
\details{
For h-step forecasting (e.g. n + 1, ... n + h),
joint forecast region of \eqn{100(1-\alpha)}\\% can be computed by
\deqn{\left\{ (y_{k, 1}, y_{k, h}) \mid y_{k, n}(i) - z_{(\alpha / 2h)} \sigma_n(i) \le y_{n, i} \le y_{k, n}(i) + z_{(\alpha / 2h)} \sigma_k(i), i = 1, \ldots, h \right\}}
See the pp41 of Lütkepohl (2007).

To compute covariance matrix, it needs VMA representation:
\deqn{Y_{t}(h) = c + \sum_{i = h}^{\infty} W_{i} \epsilon_{t + h - i} = c + \sum_{i = 0}^{\infty} W_{h + i} \epsilon_{t - i}}

Then

\deqn{\Sigma_y(h) = MSE \left[ y_t(h) \right] = \sum_{i = 0}^{h - 1} W_i \Sigma_{\epsilon} W_i^T = \Sigma_y(h - 1) + W_{h - 1} \Sigma_{\epsilon} W_{h - 1}^T}
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
