% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-forecast.R
\name{forecast_roll}
\alias{forecast_roll}
\alias{print.bvharcv}
\alias{knit_print.bvharcv}
\title{Out-of-sample Forecasting based on Rolling Window}
\usage{
forecast_roll(object, n_ahead, y_test)

\method{print}{bvharcv}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvharcv(x, ...)
}
\arguments{
\item{object}{Model object}

\item{n_ahead}{Step to forecast in rolling window scheme}

\item{y_test}{Test data to be compared. Use \code{\link[=divide_ts]{divide_ts()}} if you don't have separate evaluation dataset.}

\item{x}{\code{bvharcv} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\description{
This function forecast for out-of-sample.
}
\details{
Rolling windows forecasting fixes window size.
It moves the window ahead and forecast h-ahead in \code{y_test} set.
}
\references{
Hyndman, R. J., & Athanasopoulos, G. (2021). \emph{Forecasting: Principles and practice} (3rd ed.). OTEXTS. \url{https://otexts.com/fpp3/}
}
