% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forecast_var}
\alias{forecast_var}
\title{Forecasting Vector Autoregression}
\usage{
forecast_var(object, step)
}
\arguments{
\item{object}{\code{varlse} object by \code{\link{var_lm}}}

\item{step}{Integer, Step to forecast}
}
\description{
Forecasting Vector Autoregression
}
\details{
n-step ahead forecasting using VAR(p) recursively, based on pp35 of Lütkepohl (2007).
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
