% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forecast_vhar}
\alias{forecast_vhar}
\title{Forecasting Vector HAR}
\usage{
forecast_vhar(object, step)
}
\arguments{
\item{object}{\code{vharlse} object}

\item{step}{Integer, Step to forecast}
}
\description{
Forecasting Vector HAR
}
\details{
n-step ahead forecasting using VHAR recursively.
}
