% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{FPE}
\alias{FPE}
\title{Final Prediction Error Criterion}
\usage{
FPE(object, ...)
}
\arguments{
\item{object}{Model fit}

\item{...}{not used}
}
\value{
FPE value.
}
\description{
Generic function that computes FPE criterion.
}
