% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-varlse.R
\name{FPE.varlse}
\alias{FPE.varlse}
\title{FPE of VAR(p)}
\usage{
\method{FPE}{varlse}(object, ...)
}
\arguments{
\item{object}{\code{varlse} object}

\item{...}{not used}
}
\description{
FPE of VAR(p)
}
