% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-bayes.R
\name{gen_posterior.bvarmn}
\alias{gen_posterior.bvarmn}
\title{Posterior Distribution of Minnesota BVAR(p)}
\usage{
\method{gen_posterior}{bvarmn}(object, n = 100, ...)
}
\arguments{
\item{object}{\code{bvarmn} object}

\item{n}{number to generate (By default, 100)}

\item{...}{not used}
}
\description{
Generates Parameters of Minnesota BVAR \eqn{B, \Sigma_e}.
}
\details{
From Minnesota prior, set of coefficient matrices and residual covariance matrix have matrix Normal Inverse-Wishart distribution.

\deqn{(B, \Sigma) \sim MNIW(\hat{B}, \hat{U}, \hat{\Sigma}, \alpha_0 + n + 2)}
}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}
}
