% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-bayes.R
\name{gen_posterior}
\alias{gen_posterior}
\title{Generate Coefficient Matrix and Covariance Matrix from Bayesian Model}
\usage{
gen_posterior(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{not used}
}
\description{
Generate Coefficient Matrix and Covariance Matrix from Bayesian Model
}
