% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-models.R
\name{horseshoe_bvar_algo}
\alias{horseshoe_bvar_algo}
\title{Horseshoe Prior in BVAR}
\description{
This page describes Horseshoe prior and its Gibbs sampler
in a VAR model.
}
\references{
Carvalho, C. M., Polson, N. G., & Scott, J. G. (2010). The horseshoe estimator for sparse signals. Biometrika, 97(2), 465–480.

Makalic, E., & Schmidt, D. F. (2016). \emph{A Simple Sampler for the Horseshoe Estimator}. IEEE Signal Processing Letters, 23(1), 179–182.
}
\keyword{internal}
