% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam.R, R/print-bvharspec.R
\name{init_const}
\alias{init_const}
\alias{print.sclinit}
\alias{knit_print.sclinit}
\title{Initialize Parameters as Constant Values}
\usage{
init_const(scl = 1)

\method{print}{sclinit}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.sclinit(x, ...)
}
\arguments{
\item{scl}{Initialize by \code{scl * 1}.}

\item{x}{\code{sclinit}}

\item{digits}{digit option to print}

\item{...}{not used}
}
\description{
Initialize Parameters as Constant Values
}
