% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam-sparse.R
\name{init_horseshoe}
\alias{init_horseshoe}
\title{Initial Hyperparameters of Horseshoe Prior}
\usage{
init_horseshoe(init_local, init_global, init_priorvar)
}
\arguments{
\item{init_local}{Initial local shrinkage hyperparameters}

\item{init_global}{Initial global shrinkage hyperparameter}

\item{init_var}{Initial variance of the error term}
}
\description{
Set initial hyperparameters before starting Gibbs sampler for Horseshoe prior.
}
\references{
Carvalho, C. M., Polson, N. G., & Scott, J. G. (2010). The horseshoe estimator for sparse signals. Biometrika, 97(2), 465–480. doi:\href{https://doi.org/10.1093/biomet/asq017}{10.1093/biomet/asq017}

Makalic, E., & Schmidt, D. F. (2016). \emph{A Simple Sampler for the Horseshoe Estimator}. IEEE Signal Processing Letters, 23(1), 179–182. doi:\href{https://doi.org/10.1109/LSP.2015.2503725}{10.1109/lsp.2015.2503725}
}
