% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam-sparse.R
\name{init_ssvs}
\alias{init_ssvs}
\title{Initial parameters and Hyperparameters for SSVS Model}
\usage{
init_ssvs(init_coef, init_coef_dummy, init_chol, init_chol_dummy)
}
\arguments{
\item{init_coef}{Initial k x m coefficient matrix.}

\item{init_coef_dummy}{Initial k x m indicator matrix (1-0) corresponding to each component of coefficient.}

\item{init_chol}{Initial m x m variance matrix.}

\item{init_chol_dummy}{Initial m x m indicator matrix (1-0) corresponding to each component of variance matrix.}
}
\description{
Set initial parameters and hyperparameters of stochastic search variable selection for Bayesian VAR model.
}
\details{
Get the default SSVS setting for given VAR model.
}
\references{
George, E. I., Sun, D., & Ni, S. (2008). \emph{Bayesian stochastic search for VAR model restrictions. Journal of Econometrics}, 142(1), 553–580. doi:\href{https://doi.org/10.1016/j.jeconom.2007.08.017}{10.1016/j.jeconom.2007.08.017}

Koop, G., & Korobilis, D. (2009). \emph{Bayesian Multivariate Time Series Methods for Empirical Macroeconomics}. Foundations and Trends® in Econometrics, 3(4), 267–358. doi:\href{http://dx.doi.org/10.1561/0800000013}{10.1561/0800000013}
}
