% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam.R, R/print-bvharspec.R
\name{init_unif}
\alias{init_unif}
\alias{print.unifinit}
\alias{knit_print.unifinit}
\title{Initialize Parameters Randomly using Uniform Distribution}
\usage{
init_unif(min = -1, max = 1)

\method{print}{unifinit}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.unifinit(x, ...)
}
\arguments{
\item{min}{Lower limit of uniform distribution}

\item{max}{Upper limit of uniform distribution}

\item{x}{\code{unifinit}}

\item{digits}{digit option to print}

\item{...}{not used}
}
\description{
Initialize Parameters Randomly using Uniform Distribution
}
