% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ir_var}
\alias{ir_var}
\title{Orthogonal Impulse Response Functions of VAR}
\usage{
ir_var(object, lag_max)
}
\arguments{
\item{object}{\code{varlse} object}

\item{lag_max}{Maximum lag for VMA}
}
\description{
Compute orthogonal impulse responses of VAR
}
\details{
Based on variance decomposition (cholesky decomposition)
\deqn{\Sigma = P P^T}
where \eqn{P} is lower triangular matrix,
impulse response analysis if performed under MA representation
\deqn{y_t = \sum_{i = 0}^\infty \Theta_i v_{t - i}}
Here,
\deqn{\Theta_i = W_i P}
and \eqn{v_t = P^{-1} \epsilon_t} are orthogonal.
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
