% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ir_vhar}
\alias{ir_vhar}
\title{Orthogonal Impulse Response Functions of VHAR}
\usage{
ir_vhar(object, lag_max)
}
\arguments{
\item{object}{\code{vharlse} object}

\item{lag_max}{Maximum lag for VMA}
}
\description{
Compute orthogonal impulse responses of VHAR
}
\details{
Based on variance decomposition (cholesky decomposition)
\deqn{\Sigma = P P^T}
impulse response analysis if performed under MA representation.
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
