% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-ghosh.R
\name{is.bvarflat}
\alias{is.bvarflat}
\title{See if the Object \code{bvarflat}}
\usage{
is.bvarflat(x)
}
\arguments{
\item{x}{Object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
This function returns \code{TRUE} if the input is the output of \code{\link{bvar_flat}}.
}
