% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-ghosh.R
\name{is.bvarghosh}
\alias{is.bvarghosh}
\title{See if the Object \code{bvarghosh}}
\usage{
is.bvarghosh(x)
}
\arguments{
\item{x}{Object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
This function returns \code{TRUE} if the input is the output of \code{\link{bvar_ghosh}}.
}
