% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-minnesota.R
\name{is.bvarmn}
\alias{is.bvarmn}
\title{See if the Object \code{bvarmn}}
\usage{
is.bvarmn(x)
}
\arguments{
\item{x}{Object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
This function returns \code{TRUE} if the input is the output of \code{\link{bvar_minnesota}}.
}
