% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhar-minnesota.R
\name{is.bvharmn}
\alias{is.bvharmn}
\title{See if the Object \code{bvharmn}}
\usage{
is.bvharmn(x)
}
\arguments{
\item{x}{Object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
This function returns \code{TRUE} if the input is the output of \code{\link{bvhar_minnesota}}.
}
