% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-process.R
\name{is.stable}
\alias{is.stable}
\title{Stability of the process}
\usage{
is.stable(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{not used}
}
\value{
logical class
}
\description{
Stability of the process
}
