% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-varlse.R
\name{is.stable.varlse}
\alias{is.stable.varlse}
\title{Stability of VAR(p)}
\usage{
\method{is.stable}{varlse}(x, ...)
}
\arguments{
\item{x}{\code{varlse} object}

\item{...}{not used}
}
\description{
Stability of VAR(p)
}
