% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlse.R
\name{is.varlse}
\alias{is.varlse}
\title{See if the Object \code{varlse}}
\usage{
is.varlse(x)
}
\arguments{
\item{x}{Object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
This function returns \code{TRUE} if the input is the output of \code{\link{var_lm}}.
}
