% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vharlse.R
\name{is.vharlse}
\alias{is.vharlse}
\title{See if the Object \code{vharlse}}
\usage{
is.vharlse(x)
}
\arguments{
\item{x}{Object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
This function returns \code{TRUE} if the input is the output of \code{\link{vhar_lm}}.
}
