% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-forecast.R
\name{mape.predbvhar}
\alias{mape.predbvhar}
\title{Compute MAPE}
\usage{
\method{mape}{predbvhar}(x, y, ...)
}
\arguments{
\item{x}{\code{predbvhar} object}

\item{y}{test data to be compared. should be the same format with the train data and predict$forecast.}

\item{...}{not used}
}
\description{
Compute MAPE
}
