% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minnesota_prior}
\alias{minnesota_prior}
\title{Parameters of Normal Inverted Wishart Prior}
\usage{
minnesota_prior(x_dummy, y_dummy)
}
\arguments{
\item{x_dummy}{Matrix, dummy observation for X0}

\item{y_dummy}{Matrix, dummy observation for Y0}
}
\description{
Given dummy observations, compute parameters of Normal-IW prior for Minnesota.
}
\details{
Minnesota prior give prior to parameters \eqn{B} (VAR matrices) and \eqn{\Sigma_e} (residual covariance) the following distributions

\deqn{B \mid \Sigma_e, Y_0 \sim MN(B_0, \Omega_0, \Sigma_e)}
\deqn{\Sigma_e \mid Y_0 \sim IW(S_0, \alpha_0)}
(MN: \href{https://en.wikipedia.org/wiki/Matrix_normal_distribution}{matrix normal}, IW: \href{https://en.wikipedia.org/wiki/Inverse-Wishart_distribution}{inverse-wishart})

Bańbura et al. (2010) provides the formula how to find each matrix to match Minnesota moments.
}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}
}
