% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-forecast.R
\name{mrae}
\alias{mrae}
\alias{mrae.predbvhar}
\title{Evaluate the Model Based on MRAE (Mean Relative Absolute Error)}
\usage{
mrae(x, pred_bench, y, ...)

\method{mrae}{predbvhar}(x, pred_bench, y, ...)
}
\arguments{
\item{x}{\code{predbvhar} object to use}

\item{pred_bench}{\code{predbvhar} from benchmark model}

\item{y}{test data to be compared. should be the same format with the train data and \code{predict$forecast}.}

\item{...}{not used}
}
\description{
This function computes MRAE given prediction result versus evaluation set.
}
\details{
MRAE implements benchmark model as scaling method.
Relative error is defined by
\deqn{r_t = \frac{e_t}{e_t^{\ast}}}
where \eqn{e_t^\ast} is the error from the benchmark method.
Then

\deqn{MRAE = mean(\lvert r_t \rvert)}
}
\references{
Hyndman, R. J., & Koehler, A. B. (2006). \emph{Another look at measures of forecast accuracy}. International Journal of Forecasting, 22(4), 679–688. doi:\href{https://doi.org/10.1016/j.ijforecast.2006.03.001}{10.1016/j.ijforecast.2006.03.001}
}
