% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-forecast.R
\name{mse}
\alias{mse}
\title{Evaluate the Model Based on MSE}
\usage{
mse(x, y, ...)
}
\arguments{
\item{x}{\code{predbvhar} object}

\item{y}{test data to be compared}

\item{...}{not used}
}
\description{
Evaluate the Model Based on MSE
}
