% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oxfordman}
\alias{oxfordman}
\title{Oxford-Man Institute Realized Library}
\format{
A data frame of 53646 row and 20 columns:
\describe{
\item{Date}{Date}
\item{Symbol}{Name of the Assets}
\itemize{
\item{.AEX - AEX index}
\item{.AORD - All Ordinaries}
\item{.BFX -	Bell 20 Index}
\item{.BSESN -	S&P BSE Sensex}
\item{.BVLG - PSI All-Share Index}
\item{.BVSP - BVSP BOVESPA Index}
\item{.DJI - Dow Jones Industrial Average}
\item{.FCHI - CAC 40}
\item{.FTMIB - FTSE MIB}
\item{.FTSE - FTSE 100}
\item{.GDAXI - DAX}
\item{.GSPTSE - S&P/TSX Composite index}
\item{.HSI -	HANG SENG Index}
\item{.IBEX - IBEX 35 Index}
\item{.IXIC - Nasdaq 100}
\item{.KS11 - Korea Composite Stock Price Index (KOSPI)}
\item{.KSE - Karachi SE 100 Index}
\item{.MXX - IPC Mexico}
\item{.N225 - Nikkei 225}
\item{.NSEI - NIFTY 50}
\item{.OMXC20 - OMX Copenhagen 20 Index}
\item{.OMXHPI - OMX Helsinki All Share Index}
\item{.OMXSPI - OMX Stockholm All Share Index}
\item{.OSEAX - Oslo Exchange All-share Index}
\item{.RUT - Russel 2000}
\item{.SMSI - Madrid General Index}
\item{.SPX - S&P 500 Index}
\item{.SSEC - Shanghai Composite Index}
\item{.SSMI - Swiss Stock Market Index}
\item{.STI - Straits Times Index}
\item{.STOXX50E - EURO STOXX 50}
}
\item{nobs}{Number of observations}
\item{by_ss}{Bipower Variation (5-min Sub-sampled)}
\item{rsv}{Realized Semi-variance (5-min)}
\item{rk_parzen}{Realized Kernel Variance (Non-Flat Parzen)}
\item{rv10}{Realized Variance (10-min)}
\item{rv5_ss}{Realized Variance (5-min Sub-sampled)}
\item{rv5}{Realized Variance (5-min)}
\item{rv10_ss}{Realized Variance (10-min Sub-sampled)}
\item{rk_twoscale}{Realized Kernel Variance (Two-Scale/Bartlett)}
\item{close_price}{Closing (Last) Price}
\item{rsv_ss}{Realized Semi-variance (5-min Sub-sampled)}
\item{rk_th2}{Realized Kernel Variance (Tukey-Hanning(2))}
\item{open_time}{Opening Time}
\item{medrv}{Median Realized Variance (5-min)}
\item{open_price}{Opening (First) Price}
\item{bv}{Bipower Variation (5-min)}
\item{open_to_close}{Open to Close Return}
\item{close_time}{Closing Time}
}
}
\usage{
oxfordman
}
\description{
Famous volatility measure dataset collected daily by \href{https://realized.oxford-man.ox.ac.uk}{Oxford-man Institute of Quantitative Finance}.
}
\references{
\url{https://realized.oxford-man.ox.ac.uk/data}

Available estimators: \url{https://realized.oxford-man.ox.ac.uk/documentation/estimators}

Asset lists: \url{https://realized.oxford-man.ox.ac.uk/data/assets}
}
\keyword{datasets}
