% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{predict.bvarflat}
\alias{predict.bvarflat}
\title{Predict Method for \code{bvarghosh} object}
\usage{
\method{predict}{bvarflat}(object, n.ahead, n_iter = 100L, level = 0.05, ...)
}
\arguments{
\item{object}{\code{bvarghosh} object}

\item{n.ahead}{step to forecast}

\item{n_iter}{Number to sample residual matrix from inverse-wishart distribution. By default, 100.}

\item{level}{Specify alpha of confidence interval level 100(1 - alpha) percentage. By default, .05.}

\item{...}{not used}
}
\value{
\code{predbvhar} \link{class} with the following components:
\describe{
\item{process}{object class: bvarghosh}
\item{forecast}{forecast matrix}
\item{y}{bvarflat$y}
}
}
\description{
Point forecasting for Nonhierarchical Ghosh BVAR(p)
}
\details{
n-step ahead forecasting using BVAR recursively
}
