% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{predict.bvarmn}
\alias{predict.bvarmn}
\title{Predict Method for \code{bvarmn} object}
\usage{
\method{predict}{bvarmn}(object, n.ahead, ...)
}
\arguments{
\item{object}{\code{bvarmn} object}

\item{n.ahead}{step to forecast}

\item{...}{not used}
}
\value{
matrix
}
\description{
Point forecasting for Minnesota BVAR
}
\details{
n-step ahead forecasting using BVAR recursively
}
