% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{print.predbvhar}
\alias{print.predbvhar}
\alias{knit_print.predbvhar}
\title{Print Method for \code{predbvhar} object}
\usage{
\method{print}{predbvhar}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.predbvhar(x, ...)
}
\arguments{
\item{x}{\code{predbvhar} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\description{
Print Method for \code{predbvhar} object
}
