% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-ghosh.R
\name{residuals.bvarflat}
\alias{residuals.bvarflat}
\title{Residuals Method for \code{bvarflat} object}
\usage{
\method{residuals}{bvarflat}(object, ...)
}
\arguments{
\item{object}{\code{bvarghosh} object}

\item{...}{not used}
}
\description{
Residuals Method for \code{bvarflat} object
}
