% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-ghosh.R
\name{residuals.bvarghosh}
\alias{residuals.bvarghosh}
\title{Residuals Method for \code{bvarghosh} object}
\usage{
\method{residuals}{bvarghosh}(object, ...)
}
\arguments{
\item{object}{\code{bvarghosh} object}

\item{...}{not used}
}
\description{
Residuals Method for \code{bvarghosh} object
}
