% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-minnesota.R
\name{residuals.bvarmn}
\alias{residuals.bvarmn}
\title{Residuals Method for \code{bvarmn} object}
\usage{
\method{residuals}{bvarmn}(object, ...)
}
\arguments{
\item{object}{\code{bvarmn} object}

\item{...}{not used}
}
\description{
Residuals Method for \code{bvarmn} object
}
