% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhar-minnesota.R
\name{residuals.bvharmn}
\alias{residuals.bvharmn}
\title{Residuals Method for \code{bvharmn} object}
\usage{
\method{residuals}{bvharmn}(object, ...)
}
\arguments{
\item{object}{\code{bvharmn} object}

\item{...}{not used}
}
\description{
Residuals Method for \code{bvharmn} object
}
