% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member.R
\name{residuals}
\alias{residuals}
\alias{residuals.varlse}
\alias{residuals.vharlse}
\alias{residuals.bvarmn}
\alias{residuals.bvarflat}
\alias{residuals.bvharmn}
\title{Residual Matrix from Multivariate Time Series Models}
\usage{
\method{residuals}{varlse}(object, ...)

\method{residuals}{vharlse}(object, ...)

\method{residuals}{bvarmn}(object, ...)

\method{residuals}{bvarflat}(object, ...)

\method{residuals}{bvharmn}(object, ...)
}
\arguments{
\item{object}{Model object}

\item{...}{not used}
}
\value{
\link{matrix} object.
}
\description{
By defining \code{\link[stats:residuals]{stats::residuals()}} for each model, this function returns residual.
}
