% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlse.R
\name{residuals.varlse}
\alias{residuals.varlse}
\title{Residuals Method for \code{varlse} object}
\usage{
\method{residuals}{varlse}(object, ...)
}
\arguments{
\item{object}{\code{varlse} object}

\item{...}{not used}
}
\description{
Residuals Method for \code{varlse} object
}
