% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vharlse.R
\name{residuals.vharlse}
\alias{residuals.vharlse}
\title{Residuals Method for \code{vharlse} object}
\usage{
\method{residuals}{vharlse}(object, ...)
}
\arguments{
\item{object}{varlse}

\item{...}{not used}
}
\description{
Residuals Method for \code{vharlse} object
}
