% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scale_har}
\alias{scale_har}
\title{Build a Linear Transformation Matrix for Vector HAR}
\usage{
scale_har(m)
}
\arguments{
\item{m}{integer, dimension}
}
\description{
Build a Linear Transformation Matrix for Vector HAR
}
\details{
VHAR is linearly restricted VAR(22) in Y0 = X0 B + Z.
}
