% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperprior.R
\name{set_bvar}
\alias{set_bvar}
\title{Hyperpriors for Minnesota BVAR}
\usage{
set_bvar(sigma, lambda = 0.1, delta, eps = 1e-04)
}
\arguments{
\item{sigma}{standard error for each variable (Default: sd)}

\item{lambda}{tightness of the prior around a random walk or white noise (Default: .1)}

\item{delta}{Persistence (Litterman sets 1 = random walk prior, Default: White noise prior = 0)}

\item{eps}{very small number}
}
\description{
Sets hyperprior for \code{\link[=bvar_minnesota]{bvar_minnesota()}}.
}
