% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperprior.R
\name{set_bvhar_har}
\alias{set_bvhar_har}
\title{Hyperpriors for HAR-type Minnesota BHVAR}
\usage{
set_bvhar_har(sigma, lambda = 0.1, eps = 1e-04, daily, weekly, monthly)
}
\arguments{
\item{sigma}{Standard error vector for each variable}

\item{lambda}{Tightness of the prior around a random walk or white noise}

\item{eps}{very small number}

\item{delta}{Prior belief about white noise (Litterman sets 1: default)}
}
\description{
Sets hyperprior for \code{\link[=bvhar_minnesota]{bvhar_minnesota()}} with \code{mn_type = "VHAR"}.
}
