% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperprior.R
\name{set_bvhar_mn}
\alias{set_bvhar_mn}
\title{Hyperpriors for VAR-type Minnesota BHVAR}
\usage{
set_bvhar_mn(sigma, lambda = 0.1, delta, eps = 1e-04)
}
\arguments{
\item{sigma}{Standard error vector for each variable}

\item{lambda}{Tightness of the prior around a random walk or white noise}

\item{eps}{very small number}

\item{daily}{Same as delta in VHAR type}

\item{weekly}{Fill the second part in the first block}

\item{monthly}{Fill the third part in the first block}
}
\description{
Sets hyperprior for \code{\link[=bvhar_minnesota]{bvhar_minnesota()}} with \code{mn_type = "VAR"}.
}
