% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam.R, R/member.R
\name{set_gdp}
\alias{set_gdp}
\alias{is.gdpspec}
\title{Generalized Double Pareto Shrinkage Hyperparameters for Coefficients and Contemporaneous Coefficients}
\usage{
set_gdp(shape_grid = 100L, rate_grid = 100L)

is.gdpspec(x)
}
\arguments{
\item{shape_grid}{Griddy gibbs grid size for Gamma shape hyperparameter}

\item{rate_grid}{Griddy gibbs grid size for Gamma rate hyperparameter}

\item{x}{Any object}
}
\value{
\code{gdpspec} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Set GDP hyperparameters for VAR or VHAR coefficient and contemporaneous coefficient.
}
\references{
Armagan, A., Dunson, D. B., & Lee, J. (2013). \emph{GENERALIZED DOUBLE PARETO SHRINKAGE}. Statistica Sinica, 23(1), 119–143.

Korobilis, D., & Shimizu, K. (2022). \emph{Bayesian Approaches to Shrinkage and Sparse Estimation}. Foundations and Trends® in Econometrics, 11(4), 230-354.
}
