% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam.R, R/print-bvharspec.R
\name{set_horseshoe}
\alias{set_horseshoe}
\alias{print.horseshoeinit}
\alias{knit_print.horseshoeinit}
\title{Initial Hyperparameters of Horseshoe Prior}
\usage{
set_horseshoe(init_local, init_global, init_priorvar)

\method{print}{horseshoeinit}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.horseshoeinit(x, ...)
}
\arguments{
\item{init_local}{Initial local shrinkage hyperparameters}

\item{init_global}{Initial global shrinkage hyperparameter}

\item{init_priorvar}{Initial variance of the error term}

\item{x}{\code{horseshoeinit} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\description{
Set initial hyperparameters before starting Gibbs sampler for Horseshoe prior.
}
\details{
Set horseshoe prior initialization for VAR family.
\itemize{
\item \code{init_local}: (Initial) local shrinkage for each row of coefficients matrix.
\item \code{init_global}: (Initial) global shrinkage.
\item \code{init_priorvar}: Initial covariance matrix.
}

In this package, horseshoe prior model is estimated by Gibbs sampling,
initial means initial values for that gibbs sampler.
}
\references{
Carvalho, C. M., Polson, N. G., & Scott, J. G. (2010). The horseshoe estimator for sparse signals. Biometrika, 97(2), 465–480. doi:\href{https://doi.org/10.1093/biomet/asq017}{10.1093/biomet/asq017}

Makalic, E., & Schmidt, D. F. (2016). \emph{A Simple Sampler for the Horseshoe Estimator}. IEEE Signal Processing Letters, 23(1), 179–182. doi:\href{https://doi.org/10.1109/LSP.2015.2503725}{10.1109/lsp.2015.2503725}
}
