% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam.R, R/print-bvharspec.R, R/member.R
\name{set_intercept}
\alias{set_intercept}
\alias{print.interceptspec}
\alias{is.interceptspec}
\alias{knit_print.interceptspec}
\title{Prior for Constant Term}
\usage{
set_intercept(mean = 0, sd = 0.1)

\method{print}{interceptspec}(x, digits = max(3L, getOption("digits") - 3L), ...)

is.interceptspec(x)

\method{knit_print}{interceptspec}(x, ...)
}
\arguments{
\item{mean}{Normal mean of constant term}

\item{sd}{Normal standard deviance for constant term}

\item{x}{Any object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\description{
Set Normal prior hyperparameters for constant term
}
