% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam.R, R/print-bvharspec.R, R/member.R
\name{set_ng}
\alias{set_ng}
\alias{print.ngspec}
\alias{is.ngspec}
\title{Normal-Gamma Hyperparameter for Coefficients and Contemporaneous Coefficients}
\usage{
set_ng(
  local_shape = 1,
  group_shape = 1,
  group_scale = 1,
  global_shape = 1,
  global_scale = 1,
  contem_shape = 1,
  contem_global_shape = 1,
  contem_global_scale = 1
)

\method{print}{ngspec}(x, digits = max(3L, getOption("digits") - 3L), ...)

is.ngspec(x)
}
\arguments{
\item{local_shape}{Gamma prior shape for coefficient local shrinkage}

\item{group_shape}{Inverse gamma prior shape for coefficient group shrinkage}

\item{group_scale}{Inverse gamma prior scale for coefficient group shrinkage}

\item{global_shape}{Inverse gamma prior shape for coefficient global shrinkage}

\item{global_scale}{Inverse gamma prior scale for coefficient global shrinkage}

\item{contem_shape}{Gamma prior shape for contemporaneous coefficient local shrinkage}

\item{contem_global_shape}{Inverse gamma prior shape for contemporaneous coefficient global shrinkage}

\item{contem_global_scale}{Inverse gamma prior scale for contemporaneous coefficient global shrinkage}

\item{x}{\code{ngspec}}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{ngspec} object
}
\description{
Set NG hyperparameters for VAR or VHAR coefficient and contemporaneous coefficient.
}
\references{
George, E. I., & McCulloch, R. E. (1993). \emph{Variable Selection via Gibbs Sampling}. Journal of the American Statistical Association, 88(423), 881–889.

George, E. I., Sun, D., & Ni, S. (2008). \emph{Bayesian stochastic search for VAR model restrictions}. Journal of Econometrics, 142(1), 553–580.

Koop, G., & Korobilis, D. (2009). \emph{Bayesian Multivariate Time Series Methods for Empirical Macroeconomics}. Foundations and Trends® in Econometrics, 3(4), 267–358.
}
