% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam-sparse.R
\name{set_spikeslab_chol}
\alias{set_spikeslab_chol}
\title{Spike and Slab Hyperparameter for VAR Covariance}
\usage{
set_spikeslab_chol(
  shape = NULL,
  rate = NULL,
  spike_sd = NULL,
  slab_sd = NULL,
  slab_weight = NULL
)
}
\arguments{
\item{shape}{Gamma shape parameters for precision matrix}

\item{rate}{Gamma rate parameters for precision matrix}

\item{spike_sd}{Standard deviance for Spike normal distribution, in the precision prior.}

\item{slab_sd}{Standard deviance for Slab normal distribution, in the precision prior.}

\item{slab_weight}{Bernoulli parameter for sparsity proportion, in the precision prior.}
}
\description{
Set Hyperparameters for cholesky factor of VAR precision matrix.
}
\details{
For precision matrix \eqn{\Sigma_e^{-1}}, SSVS applies Cholesky decomposition.
\deqn{\Sigma_e^{-1} = \Psi \Psi^T}
where \eqn{\Psi = \{\psi_{ij}\}} is upper triangular.

Diagonal components follow the gamma distribution.
\deqn{\psi_{jj}^2 \sim Gamma(shape = a_j, rate = b_j)}
For each row of off-diagonal (upper-triangular) components, we apply spike-slab prior again.
\deqn{\psi_{ij} \mid w_{ij} \sim (1 - w_{ij}) N(0, \kappa_{0,ij}^2) + w_{ij} N(0, \kappa_{1,ij}^2)}
\deqn{w_{ij} \sim Bernoulli(q_{ij})}
\itemize{
\item \code{shape}: \eqn{a_j}
\item \code{rate}: \eqn{b_j}
\item \code{spike_sd}: \eqn{\kappa_{0,ij}}
\item \code{slab_sd}: \eqn{\kappa_{1,ij}}
\item \code{slab_weight}: \eqn{q_{ij}}
\item \eqn{j = 1, \ldots, mk}: vectorized format corresponding to coefficient matrix
\item \eqn{i = 1, \ldots, j - 1} and \eqn{j = 2, \ldots, m}: \eqn{\eta = (\psi_{12}, \psi_{13}, \psi_{23}, \psi_{14}, \ldots, \psi_{34}, \ldots, \psi_{1m}, \ldots, \psi_{m - 1, m})^T}
}
}
\references{
George, E. I., Sun, D., & Ni, S. (2008). \emph{Bayesian stochastic search for VAR model restrictions. Journal of Econometrics}, 142(1), 553–580. doi:\href{https://doi.org/10.1016/j.jeconom.2007.08.017}{10.1016/j.jeconom.2007.08.017}

Koop, G., & Korobilis, D. (2009). \emph{Bayesian Multivariate Time Series Methods for Empirical Macroeconomics}. Foundations and Trends® in Econometrics, 3(4), 267–358. doi:\href{http://dx.doi.org/10.1561/0800000013}{10.1561/0800000013}
}
