% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam-sparse.R
\name{set_spikeslab_coef}
\alias{set_spikeslab_coef}
\title{Spike and Slab Hyperparameter for VAR Coefficient}
\usage{
set_spikeslab_coef(spike_sd = NULL, slab_sd = NULL, slab_weight = NULL)
}
\arguments{
\item{spike_sd}{Standard deviance for Spike normal distribution (See Details).}

\item{slab_sd}{Standard deviance for Slab normal distribution (See Details).}

\item{slab_weight}{Bernoulli parameter for sparsity proportion (See Details).}
}
\description{
Set Hyperparameters for VAR coefficient matrix.
}
\details{
Let \eqn{\alpha} be the vectorized coefficient, \eqn{\alpha = vec(A)}.
Spike-slab prior is given using two normal distributions.
\deqn{\alpha_j \mid \gamma_j \sim (1 - \gamma_j) N(0, \tau_{0j}^2) + \gamma_j N(0, \tau_{1j}^2)}
As spike-slab prior itself suggests, set \eqn{\tau_{0j}} small (point mass at zero: spike distribution)
and set \eqn{\tau_{1j}} large (symmetric by zero: slab distribution).

\eqn{\gamma_j} is the proportion of the nonzero coefficients and it follows
\deqn{\gamma_j \sim Bernoulli(p_j)}
\itemize{
\item \code{spike_sd}: \eqn{\tau_{0j}}
\item \code{slab_sd}: \eqn{\tau_{1j}}
\item \code{slab_weight}: \eqn{p_j}
\item \eqn{j = 1, \ldots, mk}: vectorized format corresponding to coefficient matrix
}
}
\references{
George, E. I., Sun, D., & Ni, S. (2008). \emph{Bayesian stochastic search for VAR model restrictions. Journal of Econometrics}, 142(1), 553–580. doi:\href{https://doi.org/10.1016/j.jeconom.2007.08.017}{10.1016/j.jeconom.2007.08.017}

Koop, G., & Korobilis, D. (2009). \emph{Bayesian Multivariate Time Series Methods for Empirical Macroeconomics}. Foundations and Trends® in Econometrics, 3(4), 267–358. doi:\href{http://dx.doi.org/10.1561/0800000013}{10.1561/0800000013}
}
