% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam-sparse.R, R/print-bvharspec.R
\name{set_spikeslab_cov}
\alias{set_spikeslab_cov}
\alias{print.bvharss_sig}
\alias{knit_print.bvharss_sig}
\title{Spike and Slab Hyperparameter for VAR Covariance}
\usage{
set_spikeslab_cov(
  shape = NULL,
  rate = NULL,
  spike_sd = NULL,
  slab_sd = NULL,
  prop_sparse = NULL
)

\method{print}{bvharss_sig}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvharss_sig(x, ...)
}
\arguments{
\item{shape}{Gamma shape parameters for precision matrix}

\item{rate}{Gamma rate parameters for precision matrix}

\item{spike_sd}{Standard deviance for Spike normal distribution, in the precision prior.}

\item{slab_sd}{Standard deviance for Slab normal distribution, in the precision prior.}

\item{prop_sparse}{Bernoulli parameter for sparsity proportion, in the precision prior.}

\item{x}{\code{bvharss_sig} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\description{
Set Hyperparameters for VAR covariance matrix (precision matrix).
}
\details{
For precision matrix \eqn{\Sigma_e^{-1}}, SSVS applies Cholesky decomposition.
\deqn{\Sigma_e^{-1} = \Psi \Psi^T}
where \eqn{\Psi = \{\psi_{ij}\}} is upper triangular.

Diagonal components follow the gamma distribution.
\deqn{\psi_{jj}^2 \sim Gamma(shape = a_j, rate = b_j)}
For each row of off-diagonal (upper-triangular) components, we apply spike-slab prior again.
\deqn{\psi_{ij} \mid w_{ij} \sim (1 - w_{ij}) N(0, \kappa_{0,ij}^2) + w_{ij} N(0, \kappa_{1,ij}^2)}
\deqn{w_{ij} \sim Bernoulli(q_{ij})}
\itemize{
\item \code{shape}: \eqn{a_j}
\item \code{rate}: \eqn{b_j}
\item \code{spike_sd}: \eqn{\kappa_{0,ij}}
\item \code{slab_sd}: \eqn{\kappa_{1,ij}}
\item \code{prop_sparse}: \eqn{q_{ij}}
\item \eqn{j = 1, \ldots, mk}: vectorized format corresponding to coefficient matrix
\item \eqn{i = 1, \ldots, j - 1} and \eqn{j = 2, \ldots, m}: \eqn{\eta = (\psi_{12}, \psi_{13}, \psi_{23}, \psi_{14}, \ldots, \psi_{34}, \ldots, \psi_{1m}, \ldots, \psi_{m - 1, m})^T}
}
}
\references{
Jochmann, M., Koop, G., & Strachan, R. W. (2010). \emph{Bayesian forecasting using stochastic search variable selection in a VAR subject to breaks}. International Journal of Forecasting, 26(2), 326–347. doi:\href{https://www.sciencedirect.com/science/article/abs/pii/S0169207009001782?via\%3Dihub}{10.1016/j.ijforecast.2009.11.002}

George, E. I., Sun, D., & Ni, S. (2008). \emph{Bayesian stochastic search for VAR model restrictions}. Journal of Econometrics, 142(1), 553–580. doi:\href{https://www.sciencedirect.com/science/article/abs/pii/S0304407607001753?via\%3Dihub}{10.1016/j.jeconom.2007.08.017}
}
