% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-param.R
\name{sim_hs_var}
\alias{sim_hs_var}
\alias{sim_hs_vhar}
\title{Generate Matrix-variate Horseshoe Parameters}
\usage{
sim_hs_var(bayes_spec, p, dim_data = NULL, include_mean = TRUE)

sim_hs_vhar(bayes_spec, har = c(5, 22), dim_data = NULL, include_mean = TRUE)
}
\arguments{
\item{bayes_spec}{Horseshoe specification by \code{\link[=set_horseshoe]{set_horseshoe()}}.}

\item{p}{VAR lag}

\item{dim_data}{Specify the dimension of the data if hyperparameters of \code{bayes_spec} have constant values.}

\item{include_mean}{Add constant term (Default: \code{TRUE}) or not (\code{FALSE})}

\item{har}{Numeric vector for weekly and monthly order. By default, \code{c(5, 22)}.}
}
\description{
This function generates parameters of VAR with SSVS prior.
}
\references{
Bai, R., & Ghosh, M. (2018). High-dimensional multivariate posterior consistency under global–local shrinkage priors. Journal of Multivariate Analysis, 167, 157–170. doi:\href{https://doi.org/10.1016/j.jmva.2018.04.010}{10.1016/j.jmva.2018.04.010}

Carvalho, C. M., Polson, N. G., & Scott, J. G. (2010). \emph{The horseshoe estimator for sparse signals}. Biometrika, 97(2), 465–480. doi:\href{https://doi.org/10.1093/biomet/asq017}{10.1093/biomet/asq017}

Makalic, E., & Schmidt, D. F. (2016). \emph{A Simple Sampler for the Horseshoe Estimator}. IEEE Signal Processing Letters, 23(1), 179–182. doi:\href{https://doi.org/10.1109/LSP.2015.2503725}{10.1109/lsp.2015.2503725}
}
