% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_matgaussian}
\alias{sim_matgaussian}
\title{Generate Matrix Normal Random Matrix}
\usage{
sim_matgaussian(mat_mean, mat_scale_u, mat_scale_v)
}
\arguments{
\item{mat_mean}{Mean matrix}

\item{mat_scale_u}{First scale matrix}

\item{mat_scale_v}{Second scale matrix}
}
\description{
This function samples one matrix gaussian matrix.
}
