% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-param.R
\name{sim_mncoef}
\alias{sim_mncoef}
\title{Generate Minnesota BVAR Parameters}
\usage{
sim_mncoef(p, sigma, lambda, delta, eps = 1e-04)
}
\arguments{
\item{p}{VAR lag}

\item{sigma}{standard error for each variable}

\item{lambda}{tightness of the prior around a random walk or white noise}

\item{delta}{Persistence (Litterman sets 1 = random walk prior, Default: White noise prior = 0)}

\item{eps}{very small number}
}
\description{
This function generates parameters of BVAR with Minnesota prior.
}
\details{
Implementing dummy observation constructions,
Bańbura et al. (2010) sets Normal-IW prior.
\deqn{B \mid \Sigma_e \sim MN(B_0, \Omega_0, \Sigma_e)}
\deqn{\Sigma_e \sim IW(S_0, \alpha_0)}
}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}
}
\seealso{
\code{\link{build_ydummy}} and \code{\link{build_xdummy}} for defining dummy observations
}
