% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_stable_var}
\alias{sim_stable_var}
\title{Generate Stable VAR(p) Process}
\usage{
sim_stable_var(num_sim, var_coef, var_lag, sig_error)
}
\arguments{
\item{num_sim}{Number to generated process}

\item{var_coef}{VAR coefficient. The format should be the same as the output of \code{\link{var_lm}}}

\item{var_lag}{Lag of VAR}

\item{sig_error}{Variance matrix of the error term. Try \code{diag(dim)}.}
}
\description{
This function generates stable VAR(p)
}
\details{
Recall the relation between stable VAR(p) and VMA.
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
\seealso{
\code{\link{VARtoVMA}} computes VMA representation.
}
