% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_stable_vhar}
\alias{sim_stable_vhar}
\title{Generate Stable VHAR Process}
\usage{
sim_stable_vhar(num_sim, vhar_coef, sig_error)
}
\arguments{
\item{num_sim}{Number to generated process}

\item{vhar_coef}{VHAR coefficient. The format should be the same as the output of \code{\link{vhar_lm}}}

\item{sig_error}{Variance matrix of the error term. Try \code{diag(dim)}.}
}
\description{
This function generates stable VHAR
}
\details{
Recall the relation between stable VAR(p) and VMA.
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
\seealso{
\code{\link{VHARtoVMA}} computes VMA representation.
}
