% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-r.R
\name{split_coef}
\alias{split_coef}
\alias{split_coef.bvharmod}
\alias{split_coef.bvharirf}
\title{Splitting Coefficient Matrix into List}
\usage{
split_coef(object, ...)

\method{split_coef}{bvharmod}(object, ...)

\method{split_coef}{bvharirf}(object, ...)
}
\arguments{
\item{object}{\code{bvharmod} object}

\item{...}{not used}
}
\description{
Split \code{coefficients} into matrix list.
}
\details{
Each result of \code{\link[=var_lm]{var_lm()}}, \code{\link[=vhar_lm]{vhar_lm()}}, \code{\link[=bvar_minnesota]{bvar_minnesota()}}, \code{\link[=bvar_flat]{bvar_flat()}}, and \code{\link[=bvhar_minnesota]{bvhar_minnesota()}} is a subclass of \code{bvharmod}.
For example,
\code{c("varlse", "bvharmod")}.
}
