% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-models.R
\name{ssvs_algo}
\alias{ssvs_algo}
\title{Stochastic Search Variable Selection in VAR and VHAR}
\description{
This page describes a stochastic search variable selection (SSVS) MCMC algorithm
in VAR and VHAR models.
}
\keyword{internal}
