% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-models.R
\name{ssvs_bvar_algo}
\alias{ssvs_bvar_algo}
\title{Stochastic Search Variable Selection in VAR}
\description{
This page describes a stochastic search variable selection (SSVS) MCMC algorithm
in a VAR model.
}
\section{SSVS Prior}{

}

\section{Gibbs Sampling}{
Data: \eqn{X_0}, \eqn{Y_0}

Input:
\itemize{
\item VAR order p
\item MCMC iteration number
\item Weight of each slab: Bernoulli distribution parameters
\itemize{
\item \eqn{p_j}: of coefficients
\item \eqn{q_{ij}}: of cholesky factor
}
\item Gamma distribution parameters for cholesky factor diagonal elements \eqn{\psi_{jj}}
\itemize{
\item \eqn{a_j}: shape
\item \eqn{b_j}: rate
}
\item Correlation matrix of coefficient vector: \eqn{R = I_{k^2p}}
\item Correlation matrix to restrict cholesky factor (of \eqn{\eta_j}): \eqn{R_j = I_{j - 1}}
\item Tuning parameters for spike-and-slab sd semi-automatic approach
\itemize{
\item \eqn{c_0}: small value (0.1)
\item \eqn{c_1}: large value (10)
}
\item Constant to reduce prior influence on constant term: \eqn{c}
}

Gibbs sampling:
\enumerate{
\item Initialize \eqn{\Psi}, \eqn{\omega}, \eqn{\alpha}, \eqn{\gamma}
\item Iterate
\enumerate{
\item Diagonal elements of cholesky factor: \eqn{\psi^{(t)} \mid \alpha^{(t - 1)}, \gamma^{(t - 1)}, \omega^{(t - 1)}, Y_0}
\item Off-diagonal elements of cholesky factor: \eqn{\eta^{(t)} \mid \psi^{(t)} \alpha^{(t - 1)}, \gamma^{(t - 1)}, \omega^{(t - 1)}, Y_0}
\item Dummy vector for cholesky factor: \eqn{\omega^{(t)} \mid \eta^{(t)}, \psi^{(t)} \alpha^{(t - 1)}, \gamma^{(t - 1)}, \omega^{(t - 1)}, Y_0}
\item Coefficient vector: \eqn{\alpha^{(t)} \mid \gamma^{(t - 1)}, \Sigma^{(t)}, \omega^{(t)}, Y_0}
\item Dummy vector for coefficient vector: \eqn{\gamma^{(t)} \mid \alpha^{(t)}, \psi^{(t)}, \eta^{(t)}, \omega^{(t)}, Y_0}
}
}
}

\references{
George, E. I., Sun, D., & Ni, S. (2008). \emph{Bayesian stochastic search for VAR model restrictions. Journal of Econometrics}, 142(1), 553–580. doi:\href{https://doi.org/10.1016/j.jeconom.2007.08.017}{10.1016/j.jeconom.2007.08.017}

Koop, G., & Korobilis, D. (2009). \emph{Bayesian Multivariate Time Series Methods for Empirical Macroeconomics}. Foundations and Trends® in Econometrics, 3(4), 267–358. doi:\href{http://dx.doi.org/10.1561/0800000013}{10.1561/0800000013}
}
\keyword{internal}
