% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-varlse.R
\name{stableroot.varlse}
\alias{stableroot.varlse}
\title{Characteristic polynomial roots for VAR(p)}
\usage{
\method{stableroot}{varlse}(x, ...)
}
\arguments{
\item{x}{\code{varlse} object}

\item{...}{not used}
}
\description{
Characteristic polynomial roots for VAR(p)
}
