% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-bvarmn.R, R/summary-bayes.R
\name{print.summary.bvarmn}
\alias{print.summary.bvarmn}
\alias{knit_print.summary.bvarmn}
\alias{summary.bvarmn}
\title{Summary of \code{\link{bvar_minnesota}}}
\usage{
\method{print}{summary.bvarmn}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.summary.bvarmn(x, ...)

\method{summary}{bvarmn}(object, n_iter = 100, ...)
}
\arguments{
\item{x}{\code{summary.bvarmn} object}

\item{digits}{digit option to print}

\item{...}{not used}

\item{object}{\code{bvarmn} object}

\item{n_iter}{Number to sample Matrix Normal Inverse-Wishart distribution}
}
\value{
\code{gen_posterior} for \code{bvarmn} object returns \code{minnesota} \link{class}.
\describe{
\item{coefficients}{iter x k x m array: each column of the array indicate the draw for each lag corresponding to that variable}
\item{covmat}{iter x m x m array: each column of teh array indicate the draw for each varable corresponding to that variable}
}
}
\description{
Summary of \code{\link{bvar_minnesota}}
}
\details{
From Minnesota prior, set of coefficient matrices and residual covariance matrix have matrix Normal Inverse-Wishart distribution.

\deqn{(B, \Sigma) \sim MNIW(\hat{B}, \hat{U}, \hat{\Sigma}, \alpha_0 + n + 2)}
}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}
}
