% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-bayes.R, R/print-bvharmn.R
\name{summary.bvharmn}
\alias{summary.bvharmn}
\alias{print.summary.bvharmn}
\title{Summarizing Bayesian VHAR Model}
\usage{
\method{summary}{bvharmn}(object, n_iter = 100L, ...)

\method{print}{summary.bvharmn}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{\code{bvhar} object}

\item{n_iter}{Number to sample Matrix Normal Inverse-Wishart distribution}

\item{...}{not used}

\item{x}{\code{summary.bvharmn} object}

\item{digits}{digit option to print}
}
\description{
\code{summary} method for \code{bvharmn} class.
}
\details{
From Minnesota prior, set of coefficient matrices and residual covariance matrix have matrix Normal Inverse-Wishart distribution.

\deqn{(\Phi, \Sigma_e) \sim MNIW(\hat\Phi, \hat{V}_H^{-1}, \hat\Sigma_e, d_0 + s)}
where \eqn{\hat{V}_H = X_{+}^T X_{+}} is the posterior precision of MN.
}
