% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover.R, R/print-spillover.R
\name{summary.bvharspillover}
\alias{summary.bvharspillover}
\alias{print.summary.bvharspillover}
\alias{knit_print.summary.bvharspillover}
\title{Summarizing Spillover}
\usage{
\method{summary}{bvharspillover}(object, ...)

\method{print}{summary.bvharspillover}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.summary.bvharspillover(x, ...)
}
\arguments{
\item{object}{\code{bvharspillover} object}

\item{...}{not used}

\item{x}{\code{summary.bvharspillover} object}

\item{digits}{digit option to print}
}
\description{
This function computes directional spillovers.
}
