% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-bvharsp.R, R/summary-sparse.R
\name{print.summary.hsmod}
\alias{print.summary.hsmod}
\alias{knit_print.summary.hsmod}
\alias{summary.hsmod}
\title{Summarizing VAR and VHAR with Matrix-variate Horseshoe Prior Model}
\usage{
\method{print}{summary.hsmod}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.summary.hsmod(x, ...)

\method{summary}{hsmod}(object, level = 0.05, correction = FALSE, ...)
}
\arguments{
\item{x}{\code{summary.hsmod} object}

\item{digits}{digit option to print}

\item{...}{not used}

\item{object}{\code{hsmod} object}

\item{level}{Specify alpha of credible interval level 100(1 - alpha) percentage. By default, \code{.05}.}

\item{correction}{Multiple testing correction. By default, \code{FALSE}.}
}
\description{
Conduct variable selection.
}
\details{
MCMC can construct \eqn{100 (1 - \alpha)} credible interval.
This interval can help variable selection.
}
\references{
Bai, R., & Ghosh, M. (2018). High-dimensional multivariate posterior consistency under global–local shrinkage priors. Journal of Multivariate Analysis, 167, 157–170. doi:\href{https://doi.org/10.1016/j.jmva.2018.04.010}{10.1016/j.jmva.2018.04.010}
}
