% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-bayes.R, R/print-summarymniw.R
\name{summary.normaliw}
\alias{summary.normaliw}
\alias{print.summary.normaliw}
\alias{knit_print.summary.normaliw}
\title{Summarizing Bayesian Multivariate Time Series Model}
\usage{
\method{summary}{normaliw}(
  object,
  num_iter = 10000L,
  num_burn = floor(num_iter/2),
  thinning = 1L,
  ...
)

\method{print}{summary.normaliw}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.summary.normaliw(x, ...)
}
\arguments{
\item{object}{\code{normaliw} object}

\item{num_iter}{Number to sample MNIW distribution}

\item{num_burn}{Number of burn-in}

\item{thinning}{Thinning every thinning-th iteration}

\item{...}{not used}

\item{x}{\code{summary.normaliw} object}

\item{digits}{digit option to print}
}
\value{
\code{summary} for \code{bvarmn} object returns \code{summary.bvarmn} \link{class}.
\describe{
\item{coefficients}{iter x k x m array: each column of the array indicate the draw for each lag corresponding to that variable}
\item{covmat}{iter x m x m array: each column of teh array indicate the draw for each varable corresponding to that variable}
}
}
\description{
\code{summary} method for \code{normaliw} class.
}
\details{
From Minnesota prior, set of coefficient matrices and residual covariance matrix have matrix Normal Inverse-Wishart distribution.

BVAR:

\deqn{(A, \Sigma_e) \sim MNIW(\hat{A}, \hat{V}^{-1}, \hat\Sigma_e, \alpha_0 + s)}
where \eqn{\hat{V} = X_\ast^T X_\ast} is the posterior precision of MN.

BVHAR:

\deqn{(\Phi, \Sigma_e) \sim MNIW(\hat\Phi, \hat{V}_H^{-1}, \hat\Sigma_e, d_0 + s)}
where \eqn{\hat{V}_H = X_{+}^T X_{+}} is the posterior precision of MN.
}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. doi:\href{https://doi.org/10.2307/1391384}{10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). doi:\href{https://doi:10.1002/jae.1137}{10.1002/jae.1137}
}
