% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-models.R
\name{ts_forecasting_cv}
\alias{ts_forecasting_cv}
\title{Time Series Cross-Validation}
\description{
This page describes the out-of-sample forecasting method in time series scheme.
While the most simple way to compute test error is splitting training-test set,
but popular approach in time series literature is out-of-sample forecasting.
}
\section{Rolling Window Forecasting}{
Rolling window forecasting fixes its window size.
The window is used as training set.
This window will be moved to the end as possible as it can be.

The step should set the same step ahead forecasting at every iteration, saying one-step or h-step.
After fitting the model in the window, researcher should forecast next one-step or h-step ahead.
The longest forecast horizon is \code{(final_period - window_size) - h + 1}.

After this window, move the window one step ahead and do the same process.
Get forecasted values until possible (longest forecast horizon).
}

\section{Expanding Windows}{
Expanding window forecasting fixes its starting period.
Rolling window method moves the window with constant size,
but expanding window method just literally expands window the window fixing the starting period.

The other procedure is the same.
}

\references{
Hyndman, R. J., & Athanasopoulos, G. (2021). \emph{Forecasting: Principles and practice} (3rd ed.). OTEXTS. \url{https://otexts.com/fpp3/}
}
\keyword{internal}
