% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover.R, R/print-spillover.R
\name{varshares}
\alias{varshares}
\alias{print.bvharspillover}
\alias{knit_print.bvharspillover}
\alias{varshares.bvharmod}
\title{h-step ahead Normalized Spillover}
\usage{
varshares(
  object,
  n_ahead,
  num_iter = 10000L,
  num_burn = floor(num_iter/2),
  thinning = 1L,
  ...
)

\method{print}{bvharspillover}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvharspillover(x, ...)

\method{varshares}{bvharmod}(
  object,
  n_ahead,
  num_iter = 10000L,
  num_burn = floor(num_iter/2),
  thinning = 1L,
  ...
)
}
\arguments{
\item{object}{Model object}

\item{n_ahead}{step to forecast}

\item{num_iter}{Number to sample MNIW distribution}

\item{num_burn}{Number of burn-in}

\item{thinning}{Thinning every thinning-th iteration}

\item{...}{not used}

\item{x}{\code{bvharspillover} object}

\item{digits}{digit option to print}
}
\description{
This function gives connectedness table with h-step ahead normalized spillover index (a.k.a. variance shares).
}
