% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VARtoVMA}
\alias{VARtoVMA}
\title{Convert VAR to VMA(infinite)}
\usage{
VARtoVMA(var_coef, p)
}
\arguments{
\item{var_coef}{Matrix (mp + 1 x m), VAR matrix augmented}

\item{p}{Integer, VAR lag}
}
\description{
Convert VAR process to infinite vector MA process
}
\details{
Let VAR(p) be stable.
\deqn{Y_t = c + \sum_{j = 0} W_j Z_{t - j}}
For VAR coefficient \eqn{B_1, B_2, \ldots, B_p},
\deqn{I = (W_0 + W_1 L + W_2 L^2 + \cdots + ) (I - B_1 L - B_2 L^2 - \cdots - B_p L^p)}
Recursively,
\deqn{W_0 = I}
\deqn{W_1 = W_0 B_1}
\deqn{W_2 = W_1 B_1 + W_0 B_2}
\deqn{W_j = \sum_{j = 1}^k W_{k - j} B_j}
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
