% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-vharlse.R, R/vharlse.R
\name{print.vharlse}
\alias{print.vharlse}
\alias{vhar_lm}
\title{Fit Vector HAR}
\usage{
\method{print}{vharlse}(x, digits = max(3L, getOption("digits") - 3L), ...)

vhar_lm(y)
}
\arguments{
\item{y}{matrix, Time series data of which columns indicate the variables}
}
\value{
vharlse object
}
\description{
\code{vhar_lm} fits VHAR using OLS method
}
\details{
For VHAR model
\deqn{Y_{t} = Phi^{(d)} Y_{t - 1} + \Phi^{(w)} Y_{t - 1}^{(w)} + \Phi^{(m)} Y_{t - 1}^{(m)} + \epsilon_t}
the function gives basic values.
}
