% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VHARtoVMA}
\alias{VHARtoVMA}
\title{Convert VHAR to VMA(infinite)}
\usage{
VHARtoVMA(object, lag_max)
}
\arguments{
\item{object}{\code{vharlse} object}

\item{lag_max}{Maximum lag for VMA}
}
\value{
VMA coefficient of k(lag-max + 1) x k dimension
}
\description{
Convert VHAR process to infinite vector MA process
}
\details{
Let VAR(p) be stable
and let VAR(p) be
\eqn{Y_0 = X_0 B + Z}

VHAR is VAR(22) with
\deqn{Y_0 = X_1 B + Z = ((X_0 \tilde{T}^T)) \Phi + Z}

Observe that
\deqn{B = \tilde{T}^T \Phi}
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing.
}
