% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spillover.R
\name{autoplot.bvhardynsp}
\alias{autoplot.bvhardynsp}
\title{Dynamic Spillover Indices Plot}
\usage{
\method{autoplot}{bvhardynsp}(
  object,
  type = c("tot", "to", "from", "net"),
  hcol = "grey",
  hsize = 1.5,
  row_facet = NULL,
  col_facet = NULL,
  ...
)
}
\arguments{
\item{object}{\code{bvhardynsp} object}

\item{type}{Index to draw}

\item{hcol}{color of horizontal line = 0 (By default, grey)}

\item{hsize}{size of horizontal line = 0 (By default, 1.5)}

\item{row_facet}{\code{nrow} of \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}

\item{col_facet}{\code{ncol} of \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}

\item{...}{Additional}
}
\description{
Draws dynamic directional spillover plot.
}
