% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-bayes.R
\name{autoplot.minnesota}
\alias{autoplot.minnesota}
\title{Density Plot for \code{minnesota} Object}
\usage{
\method{autoplot}{minnesota}(
  object,
  type = c("mean", "residual"),
  var_name = NULL,
  NROW = NULL,
  NCOL = NULL,
  ...
)
}
\arguments{
\item{object}{\code{minnesota} object}

\item{type}{Draw whether Mean or residuals}

\item{var_name}{variable name (for mean)}

\item{NROW}{Numer of facet row}

\item{NCOL}{Numer of facet col}

\item{...}{not used}
}
\description{
Density Plot for \code{minnesota} Object
}
