% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-forecast.R
\name{autoplot.predbvhar}
\alias{autoplot.predbvhar}
\title{Plot Forecast Result}
\usage{
\method{autoplot}{predbvhar}(
  object,
  type = c("grid", "wrap"),
  ci_fill = "grey70",
  x_cut = 1,
  NROW = NULL,
  NCOL = NULL,
  ...
)
}
\arguments{
\item{object}{\code{predbvhar} object usually generated by \code{predict()} function in this package}

\item{type}{Divide variables using \code{\link[ggplot2]{facet_grid}} ("grid": default) or \code{\link[ggplot2]{facet_wrap}} ("wrap")}

\item{ci_fill}{color of CI}

\item{x_cut}{plot x axes from \code{x_cut} for visibility}

\item{NROW}{\code{nrow} of \code{\link[ggplot2]{facet_wrap}}}

\item{NCOL}{\code{ncol} of \code{\link[ggplot2]{facet_wrap}}}

\item{...}{additional option for \code{\link[ggplot2]{geom_path}}}
}
\description{
Plots the forecasting result with forecast regions.
}
