% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-forecast.R
\name{autoplot.predbvhar}
\alias{autoplot.predbvhar}
\title{Plot Forecast Result}
\usage{
\method{autoplot}{predbvhar}(object, type = c("facet", "shape"), ...)
}
\arguments{
\item{object}{\code{predbvhar} object usually generated by \code{predict()} function in this package}

\item{type}{Divide variables using \code{\link[ggplot2]{facet_grid}} ("facet": default) or just line type ("color")}

\item{...}{additional option for \code{\link[ggplot2]{geom_path}}}
}
\description{
Plots the forecasting result with forecast regions.
}
\details{
For h-step forecasting (e.g. n + 1, ... n + h),
joint forecast region of \eqn{100(1-\alpha)}\\% can be computed by
\deqn{\left\{ (y_{k, 1}, y_{k, h}) \mid y_{k, n}(i) - z_{(\alpha / 2h)} \sigma_n(i) \le y_{n, i} \le y_{k, n}(i) + z_{(\alpha / 2h)} \sigma_k(i), i = 1, \ldots, h \right\}}
See the pp41 of Lütkepohl (2007).
}
