% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{autoplot.predbvhar}
\alias{autoplot.predbvhar}
\title{Plot Forecast Result}
\usage{
\method{autoplot}{predbvhar}(object, type = c("facet", "shape"), ...)
}
\arguments{
\item{object}{\code{predbvhar} object usually generated by \code{predict()} function in this package}

\item{type}{Divide variables using \code{\link[ggplot2]{facet_grid}} ("facet": default) or just line type ("color")}

\item{...}{additional option for \code{\link[ggplot2]{geom_path}}}
}
\description{
Plots the forecasting result
}
