% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.predbvhar}
\alias{autoplot.predbvhar}
\alias{autolayer.predbvhar}
\title{Plot Forecast Result}
\usage{
\method{autoplot}{predbvhar}(
  object,
  type = c("grid", "wrap"),
  ci_alpha = 0.7,
  alpha_scale = 0.3,
  x_cut = 1,
  line_type = c(`FALSE` = "dotted", `TRUE` = "solid"),
  NROW = NULL,
  NCOL = NULL,
  ...
)

\method{autolayer}{predbvhar}(
  object,
  ci_fill = "grey70",
  ci_alpha = 0.5,
  alpha_scale = 0.3,
  line_type = c(`FALSE` = "dotted", `TRUE` = "solid"),
  ...
)
}
\arguments{
\item{object}{\code{predbvhar} object usually generated by \code{predict()} function in this package}

\item{type}{Divide variables using \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid}} ("grid": default) or \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}} ("wrap")}

\item{ci_alpha}{Transparency of CI}

\item{alpha_scale}{Scale of transparency parameter (\code{alpha}) between the two layers. \code{alpha} of CI ribbon = \code{alpha_scale} * \code{alpha} of path (By default, .5)}

\item{x_cut}{plot x axes from \code{x_cut} for visibility}

\item{line_type}{linetype regarding to forecasting. See \code{\link[ggplot2:scale_linetype_manual]{ggplot2::scale_linetype_manual}}}

\item{NROW}{\code{nrow} of \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}}

\item{NCOL}{\code{ncol} of \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}}

\item{...}{additional option for \code{\link[ggplot2:geom_path]{ggplot2::geom_path}}}

\item{ci_fill}{color of CI}
}
\description{
Plots the forecasting result with forecast regions.
}
