% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.predbvhar}
\alias{autoplot.predbvhar}
\alias{autolayer.predbvhar}
\title{Plot Forecast Result}
\usage{
\method{autoplot}{predbvhar}(
  object,
  type = c("grid", "wrap"),
  ci_alpha = 0.7,
  alpha_scale = 0.3,
  x_cut = 1,
  viridis = FALSE,
  viridis_option = "D",
  NROW = NULL,
  NCOL = NULL,
  ...
)

\method{autolayer}{predbvhar}(object, ci_fill = "grey70", ci_alpha = 0.5, alpha_scale = 0.3, ...)
}
\arguments{
\item{object}{A \code{predbvhar} object}

\item{type}{Divide variables using \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}} ("grid": default) or \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} ("wrap")}

\item{ci_alpha}{Transparency of CI}

\item{alpha_scale}{Scale of transparency parameter (\code{alpha}) between the two layers. \code{alpha} of CI ribbon = \code{alpha_scale} * \code{alpha} of path (By default, .5)}

\item{x_cut}{plot x axes from \code{x_cut} for visibility}

\item{viridis}{If \code{TRUE}, scale CI and forecast line using \code{\link[ggplot2:scale_viridis]{ggplot2::scale_fill_viridis_d()}} and \link[ggplot2:scale_viridis]{ggplot2::scale_colour_viridis_d}, respectively.}

\item{viridis_option}{Option for viridis string. See \code{option} of \link[ggplot2:scale_viridis]{ggplot2::scale_colour_viridis_d}. Choose one of \code{c("A", "B", "C", "D", "E")}. By default, \code{D}.}

\item{NROW}{\code{nrow} of \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}

\item{NCOL}{\code{ncol} of \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}

\item{...}{additional option for \code{\link[ggplot2:geom_path]{ggplot2::geom_path()}}}

\item{ci_fill}{color of CI}
}
\value{
A ggplot object

A ggplot layer
}
\description{
Plots the forecasting result with forecast regions.
}
