% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.summary.mvhsmod}
\alias{autoplot.summary.mvhsmod}
\title{Plot the Credible Interval of Matrix-variate Horseshoe Coefficients}
\usage{
\method{autoplot}{summary.mvhsmod}(object, colour = NULL, linetype = NULL, fill = NULL, ...)
}
\arguments{
\item{object}{\code{summary.mvhsmod} object}

\item{colour}{Color of the credible interval line (Try \code{"black"}).}

\item{linetype}{Line type of the credible interval line (Try \code{1}: solid).}

\item{fill}{Color to fill between credible interval lines (Try none: \code{NA}).}

\item{...}{Other arguments passed on the \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}}.}
}
\description{
Draw credible interval for horseshoe prior coefficients.
}
