% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.summary.normaliw}
\alias{autoplot.summary.normaliw}
\title{Density Plot for Minnesota Prior VAR Model}
\usage{
\method{autoplot}{summary.normaliw}(
  object,
  type = c("trace", "dens", "area"),
  pars = character(),
  regex_pars = character(),
  ...
)
}
\arguments{
\item{object}{\code{summary.normaliw} object}

\item{type}{The type of the plot. Trace plot (\code{"trace"}), kernel density plot (\code{"dens"}), and interval estimates plot (\code{"area"}).}

\item{pars}{Parameter names to draw.}

\item{regex_pars}{Regular expression parameter names to draw.}

\item{...}{Other options for each \code{\link[bayesplot:MCMC-traces]{bayesplot::mcmc_trace()}}, \code{\link[bayesplot:MCMC-distributions]{bayesplot::mcmc_dens()}}, and \code{\link[bayesplot:MCMC-intervals]{bayesplot::mcmc_areas()}}.}
}
\value{
A ggplot object
}
\description{
This function draws density plot for coefficient matrices of Minnesota prior VAR model.
}
