% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R, R/print-tuning.R, R/member.R
\name{bound_bvhar}
\alias{bound_bvhar}
\alias{print.boundbvharemp}
\alias{is.boundbvharemp}
\alias{knit_print.boundbvharemp}
\title{Setting Empirical Bayes Optimization Bounds}
\usage{
bound_bvhar(
  init_spec = set_bvhar(),
  lower_spec = set_bvhar(),
  upper_spec = set_bvhar()
)

\method{print}{boundbvharemp}(x, digits = max(3L, getOption("digits") - 3L), ...)

is.boundbvharemp(x)

\method{knit_print}{boundbvharemp}(x, ...)
}
\arguments{
\item{init_spec}{Initial Bayes model specification}

\item{lower_spec}{Lower bound Bayes model specification}

\item{upper_spec}{Upper bound Bayes model specification}

\item{x}{Any object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{boundbvharemp} \link{class}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} This function sets lower and upper bounds for \code{\link[=set_bvar]{set_bvar()}}, \code{\link[=set_bvhar]{set_bvhar()}}, or \code{\link[=set_weight_bvhar]{set_weight_bvhar()}}.
}
