% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_design}
\alias{build_design}
\title{Build X0 matrix in VAR(p)}
\usage{
build_design(x, p)
}
\arguments{
\item{x}{Matrix, time series data}

\item{p}{VAR lag}
}
\description{
Build X0 matrix in VAR(p)
}
\details{
X0 is
\deqn{X0 = [Y_p, \ldots, Y_1, 1]}
i.e. (n - p) x (mp + 1) matrix
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
