% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-minnesota.R, R/print-bvarmn.R
\name{bvar_minnesota}
\alias{bvar_minnesota}
\alias{print.bvarmn}
\alias{knit_print.bvarmn}
\title{Fit Bayesian VAR(p) of Minnesota Prior}
\usage{
bvar_minnesota(y, p, bayes_spec = set_bvar(), include_mean = TRUE)

\method{print}{bvarmn}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvarmn(x, ...)
}
\arguments{
\item{y}{matrix, Time series data of which columns indicate the variables}

\item{p}{VAR lag}

\item{bayes_spec}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} A BVAR model specification by \code{\link[=set_bvar]{set_bvar()}}.}

\item{include_mean}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Add constant term (Default: \code{TRUE}) or not (\code{FALSE})}

\item{x}{\code{bvarmn} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{bvar_minnesota} returns an object \code{bvarmn} \link{class}.
It is a list with the following components:

\describe{
\item{design}{\eqn{X_0}}
\item{y0}{\eqn{Y_0}}
\item{y}{Raw input}
\item{p}{Lag of VAR}
\item{m}{Dimension of the data}
\item{obs}{Sample size used when training = \code{totobs} - \code{p}}
\item{totobs}{Total number of the observation}
\item{process}{Process: BVAR_Minnesota}
\item{spec}{Model specification (\code{bvharspec})}
\item{type}{include constant term (\code{const}) or not (\code{none})}
\item{call}{Matched call}
\item{mn_mean}{Location of posterior matrix normal distribution}
\item{fitted.values}{Fitted values}
\item{residuals}{Residuals}
\item{mn_prec}{Precision matrix of posterior matrix normal distribution}
\item{iw_scale}{Scale matrix of posterior inverse-wishart distribution}
\item{a0}{\eqn{\alpha_0}: nrow(Dummy observation) - k}
}
}
\description{
This function fits BVAR(p) with Minnesota prior.
}
\details{
Minnesota prior give prior to parameters \eqn{B} (VAR matrices) and \eqn{\Sigma_e} (residual covariance).

\deqn{B \mid \Sigma_e \sim MN(B_0, \Omega_0, \Sigma_e)}
\deqn{\Sigma_e \sim IW(S_0, \alpha_0)}
(MN: \href{https://en.wikipedia.org/wiki/Matrix_normal_distribution}{matrix normal}, IW: \href{https://en.wikipedia.org/wiki/Inverse-Wishart_distribution}{inverse-wishart})

\eqn{\delta_i} are related to the belief to random walk.
\itemize{
\item If \eqn{\delta_i = 1} for all i, random walk prior
\item If \eqn{\delta_i = 0} for all i, white noise prior
}

\eqn{\lambda} controls the overall tightness of the prior around these two prior beliefs.
\itemize{
\item If \eqn{\lambda = 0}, the posterior is equivalent to prior and the data do not influence the estimates.
\item If \eqn{\lambda = \infty}, the posterior mean becomes OLS estimates (VAR).
}

\eqn{\sigma_i^2 / \sigma_j^2} in Minnesota moments explain the data scales.
}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}
}
