% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-minnesota.R, R/print-bvarmn.R
\name{bvar_minnesota}
\alias{bvar_minnesota}
\alias{print.bvarmn}
\alias{knit_print.bvarmn}
\title{Fit Bayesian VAR(p) of Minnesota Prior}
\usage{
bvar_minnesota(y, p, sigma, lambda, delta, eps = 1e-04)

\method{print}{bvarmn}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvarmn(x, ...)
}
\arguments{
\item{y}{matrix, Time series data of which columns indicate the variables}

\item{p}{VAR lag}

\item{sigma}{standard error for each variable}

\item{lambda}{tightness of the prior around a random walk or white noise}

\item{delta}{Persistence (Litterman sets 1 = random walk prior, Default: White noise prior = 0)}

\item{eps}{very small number}

\item{x}{\code{bvarmn} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{bvar_minnesota} returns an object \code{bvarmn} \link{class}.

It is a list with the following components:

\describe{
\item{design}{\eqn{X_0}}
\item{y0}{\eqn{Y_0}}
\item{y}{Raw input}
\item{p}{Lag of VAR}
\item{m}{Dimension of the data}
\item{obs}{Sample size used when training = \code{totobs} - \code{p}}
\item{totobs}{Total number of the observation}
\item{process}{Process: Minnesota}
\item{call}{Matched call}
\item{mn_mean}{Location of posterior matrix normal distribution}
\item{fitted.values}{Fitted values}
\item{residuals}{Residuals}
\item{mn_prec}{Precision matrix of posterior matrix normal distribution}
\item{iw_scale}{Scale matrix of posterior inverse-wishart distribution}
\item{a0}{\eqn{\alpha_0}: nrow(Dummy observation) - k}
}
}
\description{
This function fits BVAR(p) with Minnesota prior.
}
\details{
Minnesota prior give prior to parameters \eqn{B} (VAR matrices) and \eqn{\Sigma_e} (residual covariance).

\deqn{B \mid \Sigma_e \sim MN(B_0, \Omega_0, \Sigma_e)}
\deqn{\Sigma_e \sim IW(S_0, \alpha_0)}
(MN: \href{https://en.wikipedia.org/wiki/Matrix_normal_distribution}{matrix normal}, IW: \href{https://en.wikipedia.org/wiki/Inverse-Wishart_distribution}{inverse-wishart})
}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}
}
