% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-minnesota.R, R/print-bvarmn.R
\name{bvar_minnesota}
\alias{bvar_minnesota}
\alias{print.bvarmn}
\alias{knit_print.bvarmn}
\title{Fitting Bayesian VAR(p) of Minnesota Prior}
\usage{
bvar_minnesota(y, p, bayes_spec = set_bvar(), include_mean = TRUE)

\method{print}{bvarmn}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvarmn(x, ...)
}
\arguments{
\item{y}{Time series data of which columns indicate the variables}

\item{p}{VAR lag}

\item{bayes_spec}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} A BVAR model specification by \code{\link[=set_bvar]{set_bvar()}}.}

\item{include_mean}{Add constant term (Default: \code{TRUE}) or not (\code{FALSE})}

\item{x}{\code{bvarmn} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{bvar_minnesota} returns an object \code{bvarmn} \link{class}.
It is a list with the following components:

\describe{
\item{coefficients}{Posterior Mean matrix of Matrix Normal distribution}
\item{fitted.values}{Fitted values}
\item{residuals}{Residuals}
\item{mn_prec}{Posterior precision matrix of Matrix Normal distribution}
\item{iw_scale}{Posterior scale matrix of posterior inverse-wishart distribution}
\item{iw_shape}{Posterior shape of inverse-wishart distribution (\eqn{alpha_0} - obs + 2). \eqn{\alpha_0}: nrow(Dummy observation) - k}
\item{df}{Numer of Coefficients: mp + 1 or mp}
\item{p}{Lag of VAR}
\item{m}{Dimension of the time series}
\item{obs}{Sample size used when training = \code{totobs} - \code{p}}
\item{totobs}{Total number of the observation}
\item{call}{Matched call}
\item{process}{Process string in the \code{bayes_spec}: \code{"BVAR_Minnesota"}}
\item{spec}{Model specification (\code{bvharspec})}
\item{type}{include constant term (\code{"const"}) or not (\code{"none"})}
\item{prior_mean}{Prior mean matrix of Matrix Normal distribution: \eqn{A_0}}
\item{prior_precision}{Prior precision matrix of Matrix Normal distribution: \eqn{\Omega_0}}
\item{prior_scale}{Prior scale matrix of inverse-wishart distribution: \eqn{S_0}}
\item{prior_shape}{Prior shape of inverse-wishart distribution: \eqn{\alpha_0}}
\item{y0}{\eqn{Y_0}}
\item{design}{\eqn{X_0}}
\item{y}{Raw input}
}
}
\description{
This function fits BVAR(p) with Minnesota prior.
}
\details{
Minnesota prior give prior to parameters \eqn{B} (VAR matrices) and \eqn{\Sigma_e} (residual covariance).

\deqn{A \mid \Sigma_e \sim MN(A_0, \Omega_0, \Sigma_e)}
\deqn{\Sigma_e \sim IW(S_0, \alpha_0)}
(MN: \href{https://en.wikipedia.org/wiki/Matrix_normal_distribution}{matrix normal}, IW: \href{https://en.wikipedia.org/wiki/Inverse-Wishart_distribution}{inverse-wishart})

\eqn{\delta_i} are related to the belief to random walk.
\itemize{
\item If \eqn{\delta_i = 1} for all i, random walk prior
\item If \eqn{\delta_i = 0} for all i, white noise prior
}

\eqn{\lambda} controls the overall tightness of the prior around these two prior beliefs.
\itemize{
\item If \eqn{\lambda = 0}, the posterior is equivalent to prior and the data do not influence the estimates.
\item If \eqn{\lambda = \infty}, the posterior mean becomes OLS estimates (VAR).
}

\eqn{\sigma_i^2 / \sigma_j^2} in Minnesota moments explain the data scales.
}
\examples{
# Perform the function using etf_vix dataset
\dontrun{
  fit <- bvar_minnesota(y = etf_vix, p = 5)
  class(fit)
  str(fit)
}

# Extract coef, fitted values, and residuals
\dontrun{
  coef(fit)
  residuals(fit)
  fitted(fit)
}

}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}

KADIYALA, K.R. and KARLSSON, S. (1997), \emph{NUMERICAL METHODS FOR ESTIMATION AND INFERENCE IN BAYESIAN VAR-MODELS}. J. Appl. Econ., 12: 99-132. \url{https://doi.org/10.1002/(SICI)1099-1255(199703)12:2<99::AID-JAE429>3.0.CO;2-A}

Sims, C. A., & Zha, T. (1998). \emph{Bayesian Methods for Dynamic Multivariate Models}. International Economic Review, 39(4), 949–968. \url{https://doi.org/10.2307/2527347}

Domenico Giannone, Michele Lenza, Giorgio E. Primiceri; \emph{Prior Selection for Vector Autoregressions}. The Review of Economics and Statistics 2015; 97 (2): 436–451. doi: \url{https://doi.org/10.1162/REST_a_00483}
}
\seealso{
\itemize{
\item \code{\link[=set_bvar]{set_bvar()}} to specify the hyperparameters of Minnesota prior.
\item \code{\link[=build_ydummy]{build_ydummy()}} and \code{\link[=build_xdummy]{build_xdummy()}} to construct dummy observations.
\item \code{\link[=estimate_bvar_mn]{estimate_bvar_mn()}} to compute BVAR prior and posterior.
}
}
