% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-ssvs.R
\name{bvar_ssvs}
\alias{bvar_ssvs}
\alias{print.bvarsp}
\alias{knit_print.bvarsp}
\title{Fitting Bayesian VAR(p) of SSVS Prior}
\usage{
bvar_ssvs(
  y,
  p,
  num_iter = 100,
  num_burn = 0,
  thinning = 1,
  bayes_spec = set_ssvs(),
  init_spec = init_ssvs(),
  include_mean = TRUE
)

\method{print}{bvarsp}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvarsp(x, ...)
}
\arguments{
\item{y}{Time series data of which columns indicate the variables}

\item{p}{VAR lag}

\item{num_iter}{MCMC iteration number}

\item{num_burn}{Number of burn-in}

\item{thinning}{Thinning every thinning-th iteration}

\item{bayes_spec}{A BVAR model specification by \code{\link[=set_ssvs]{set_ssvs()}}.}

\item{init_spec}{SSVS initialization specification by \code{\link[=init_ssvs]{init_ssvs()}}.}

\item{include_mean}{Add constant term (Default: \code{TRUE}) or not (\code{FALSE})}

\item{x}{\code{bvarsp} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{bvar_ssvs} returns an object named \code{bvarsp} \link{class}.
It is a list with the following components:

\describe{
\item{coefficients}{Coefficient Matrix}
\item{p}{Lag of VAR}
\item{m}{Dimension of the data}
\item{obs}{Sample size used when training = \code{totobs} - \code{p}}
\item{totobs}{Total number of the observation}
\item{call}{Matched call}
\item{type}{include constant term (\code{"const"}) or not (\code{"none"})}
\item{y0}{\eqn{Y_0}}
\item{design}{\eqn{X_0}}
\item{y}{Raw input}
}
}
\description{
This function fits BVAR(p) with stochastic search variable selection (SSVS) prior.
}
\details{
SSVS prior gives prior to parameters \eqn{\alpha = vec(A)} (VAR coefficient) and \eqn{\Sigma_e^{-1} = \Psi \Psi^T} (residual covariance).

\deqn{\alpha_j \mid \gamma_j \sim (1 - \gamma_j) N(0, \kappa_{0j}^2) + \gamma_j N(0, \kappa_{1j}^2)}
\deqn{\gamma_j \sim Bernoulli(q_j)}

and for upper triangular matrix \eqn{\Psi},

\deqn{\psi_{jj}^2 \sim Gamma(shape = a_j, rate = b_j)}
\deqn{\psi_{ij} \mid w_{ij} \sim (1 - w_{ij}) N(0, \kappa_{0,ij}^2) + w_{ij} N(0, \kappa_{1,ij}^2)}
\deqn{w_{ij} \sim Bernoulli(q_{ij})}

Gibbs sampler is used for the estimation.
See \link{ssvs_bvar_algo} how it works.
}
\references{
George, E. I., Sun, D., & Ni, S. (2008). \emph{Bayesian stochastic search for VAR model restrictions. Journal of Econometrics}, 142(1), 553–580. doi:\href{https://doi.org/10.1016/j.jeconom.2007.08.017}{10.1016/j.jeconom.2007.08.017}

Koop, G., & Korobilis, D. (2009). \emph{Bayesian Multivariate Time Series Methods for Empirical Macroeconomics}. Foundations and Trends® in Econometrics, 3(4), 267–358. doi:\href{http://dx.doi.org/10.1561/0800000013}{10.1561/0800000013}
}
\seealso{
\itemize{
\item Vectorization formulation \link{var_vec_formulation}
\item Gibbs sampler algorithm \link{ssvs_bvar_algo}
}
}
