% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhar-package.R
\docType{package}
\name{bvhar-package}
\alias{bvhar}
\alias{bvhar-package}
\title{bvhar: Bayesian Vector HAR Model}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A package aiming at researching Bayesian Vector heterogeneous autoregressive process.
    You can model VAR, VHAR, Bayesian VAR (BVAR), and Bayesian VHAR (BVHAR).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ygeunkim/bvhar}
  \item Report bugs at \url{https://github.com/ygeunkim/bvhar/issues}
}

}
\author{
\strong{Maintainer}: Young Geun Kim \email{dudrms33@g.skku.edu} [copyright holder]

Other contributors:
\itemize{
  \item Changryong Baek [contributor]
}

}
\keyword{internal}
