% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhar-package.R
\docType{package}
\name{bvhar-package}
\alias{bvhar}
\alias{bvhar-package}
\title{bvhar: Bayesian Vector Autoregressive Heterogeneous Modeling}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A package aiming at researching Bayesian Vector heterogeneous autoregressive process. `bvhar` can model VAR, VHAR, Bayesian VAR (BVAR), and Bayesian VHAR (BVHAR).
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The bvhar package provides function to analyze and forecast multivariate time series data via vector autoregressive modelling.
Here, vector autoregressive modelling includes:
\itemize{
\item Vector autoregressive (VAR) model: \code{\link[=var_lm]{var_lm()}}
\item Vector heterogeneous autoregressive (VHAR) model: \code{\link[=vhar_lm]{vhar_lm()}}
\item Bayesian VAR (BVAR) model: \code{\link[=bvar_minnesota]{bvar_minnesota()}}, \code{\link[=bvar_flat]{bvar_flat()}}
\item Bayesian VHAR (BVHAR) model: \code{\link[=bvhar_minnesota]{bvhar_minnesota()}}
}

Each function returns S3 class \code{varlse}, \code{vharlse}, \code{bvarmn}, \code{bvarflat}, and \code{bvharmn}.
As dealing with other statistical functions (such as \code{\link[stats:lm]{stats::lm()}} and \code{\link[stats:glm]{stats::glm()}}),
users can analyze time series with S3 methods:
\itemize{
\item Extract: \code{\link[=coef.varlse]{coef.varlse()}}, \code{\link[=residuals.varlse]{residuals.varlse()}}, \code{\link[=fitted.varlse]{fitted.varlse()}}
\item log-likelihood: \code{\link[=logLik.varlse]{logLik.varlse()}}
\item Forecasting: \code{\link[=predict.varlse]{predict.varlse()}}
}
}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}

Ghosh, S., Khare, K., & Michailidis, G. (2018). \emph{High-Dimensional Posterior Consistency in Bayesian Vector Autoregressive Models}. Journal of the American Statistical Association, 114(526). \url{https://doi:10.1080/01621459.2018.1437043}

Baek, C. and Park, M. (2021). \emph{Sparse vector heterogeneous autoregressive modeling for realized volatility}. J. Korean Stat. Soc. 50, 495–510. \url{https://doi.org/10.1007/s42952-020-00090-5}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ygeunkim/bvhar}
  \item Report bugs at \url{https://github.com/ygeunkim/bvhar/issues}
}

}
\author{
\strong{Maintainer}: Young Geun Kim \email{dudrms33@g.skku.edu} [copyright holder]

Other contributors:
\itemize{
  \item Changryong Baek [contributor]
}

}
\keyword{internal}
