% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhar-minnesota.R, R/print-bvharmn.R
\name{bvhar_minnesota}
\alias{bvhar_minnesota}
\alias{print.bvharmn}
\alias{knit_print.bvharmn}
\title{Fit Bayesian VHAR of Minnesota Prior}
\usage{
bvhar_minnesota(y, bayes_spec = set_bvhar(), include_mean = TRUE)

\method{print}{bvharmn}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvharmn(x, ...)
}
\arguments{
\item{y}{Time series data of which columns indicate the variables}

\item{bayes_spec}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} A BVHAR model specification by \code{\link[=set_bvhar]{set_bvhar()}} (default) or \code{\link[=set_weight_bvhar]{set_weight_bvhar()}}.}

\item{include_mean}{Add constant term (Default: \code{TRUE}) or not (\code{FALSE})}

\item{x}{\code{bvarmn} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{bvhar_minnesota} returns an object \code{bvharmn} \link{class}.
It is a list with the following components:

\describe{
\item{design}{\eqn{X_0}}
\item{y0}{\eqn{Y_0}}
\item{y}{Raw input}
\item{m}{Dimension of the data}
\item{obs}{Sample size used when training = \code{totobs} - \code{p}}
\item{totobs}{Total number of the observation}
\item{process}{Process: BVHAR_MN_VAR or BVHAR_MN_VHAR}
\item{spec}{Model specification (\code{bvharspec})}
\item{type}{include constant term (\code{"const"}) or not (\code{"none"})}
\item{call}{Matched call}
\item{mn_mean}{Location of posterior matrix normal distribution}
\item{fitted.values}{Fitted values}
\item{residuals}{Residuals}
\item{mn_prec}{Precision matrix of posterior matrix normal distribution}
\item{iw_scale}{Scale matrix of posterior inverse-wishart distribution}
\item{a0}{\eqn{\alpha_0}: nrow(Dummy observation) - k}
}
}
\description{
This function fits BVHAR with Minnesota prior.
}
\details{
Apply Minnesota prior to Vector HAR: \eqn{\Phi} (VHAR matrices) and \eqn{\Sigma_e} (residual covariance).

\deqn{\Phi \mid \Sigma_e \sim MN(P_0, \Psi_0, \Sigma_e)}
\deqn{\Sigma_e \sim IW(U_0, d_0)}
(MN: \href{https://en.wikipedia.org/wiki/Matrix_normal_distribution}{matrix normal}, IW: \href{https://en.wikipedia.org/wiki/Inverse-Wishart_distribution}{inverse-wishart})

Two types of Minnesota priors builds different dummy variables for Y0.
\code{mn_type = "VAR"} constructs dummy Y0 with \code{p = 3} of \code{\link[=build_ydummy]{build_ydummy()}}.
The only difference from BVAR is dimension.

On the other hand, dummy Y0 for \code{mn_type = "VHAR"} has its own function \code{\link[=build_ydummy_bvhar]{build_ydummy_bvhar()}}.
It fills the zero matrix in the first block in Bańbura et al. (2010).
}
\examples{
# Perform the function using etf_vix dataset
\dontrun{
  fit <- bvhar_minnesota(y = etf_vix)
  class(fit)
  str(fit)
}

# Extract coef, fitted values, and residuals
\dontrun{
  coef(fit)
  residuals(fit)
  fitted(fit)
}

}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}

Baek, C. and Park, M. (2021). \emph{Sparse vector heterogeneous autoregressive modeling for realized volatility}. J. Korean Stat. Soc. 50, 495–510. \url{https://doi.org/10.1007/s42952-020-00090-5}

Corsi, F. (2008). \emph{A Simple Approximate Long-Memory Model of Realized Volatility}. Journal of Financial Econometrics, 7(2), 174–196. \url{https://doi:10.1093/jjfinec/nbp001}
}
\seealso{
\itemize{
\item \code{\link[=set_bvhar]{set_bvhar()}} to specify the hyperparameters of VAR-type Minnesota prior.
\item \code{\link[=set_weight_bvhar]{set_weight_bvhar()}} to specify the hyperparameters of HAR-type Minnesota prior.
\item \code{\link[=build_ydummy]{build_ydummy()}} and \code{\link[=build_xdummy]{build_xdummy()}}, and \code{\link[=build_ydummy_bvhar]{build_ydummy_bvhar()}} to add dummy observations.
\item \code{\link[=estimate_bvar_mn]{estimate_bvar_mn()}} to compute Minnesota prior and posterior.
}
}
