% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhar-minnesota.R, R/print-bvharmn.R
\name{bvhar_minnesota}
\alias{bvhar_minnesota}
\alias{print.bvharmn}
\alias{knit_print.bvharmn}
\title{Fit Bayesian BVHAR of Minnesota Prior}
\usage{
bvhar_minnesota(y, sigma, lambda, delta, eps = 1e-04)

\method{print}{bvharmn}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvharmn(x, ...)
}
\arguments{
\item{y}{matrix, Time series data of which columns indicate the variables}

\item{sigma}{standard error for each variable}

\item{lambda}{tightness of the prior around a random walk or white noise}

\item{delta}{prior belief about white noise (Litterman sets 1: default)}

\item{eps}{very small number}

\item{x}{\code{bvarmn} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{bvarmn} object with
\item{\code{design}}{\eqn{X_0}}
\item{\code{y0}}{\eqn{Y_0}}
\item{\code{y}}{raw input}
\item{\code{m}}{Dimension of the data}
\item{\code{obs}}{Sample size used when training = \code{totobs} - \code{p}}
\item{\code{totobs}}{Total number of the observation}
\item{\code{process}}{Process: VAR}
\item{\code{call}}{Matched call}
\item{\code{mn_mean}}{Location of posterior matrix normal distribution}
\item{\code{fitted.values}}{Fitted values}
\item{\code{mn_scale}}{First scale matrix of posterior matrix normal distribution}
\item{\code{iw_mean}}{Scale matrix of posterior inverse-wishart distribution}
}
\description{
This function fits BVAR(p) with Minnesota prior.
}
\details{
Apply Minnesota prior to Vector HAR: \eqn{\Phi} (VHAR matrices) and \eqn{\Sigma_e} (residual covariance).

\deqn{\Phi \mid \Sigma_e, Y_0 \sim MN(B_0, \Omega_0, \Sigma_e)}
\deqn{\Sigma_e \mid Y_0 \sim IW(S_0, \alpha_0)}
(MN: \href{https://en.wikipedia.org/wiki/Matrix_normal_distribution}{matrix normal}, IW: \href{https://en.wikipedia.org/wiki/Inverse-Wishart_distribution}{inverse-wishart})
}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}

Corsi, F. (2008). \emph{A Simple Approximate Long-Memory Model of Realized Volatility}. Journal of Financial Econometrics, 7(2), 174–196. \url{https://doi:10.1093/jjfinec/nbp001}
}
