% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R, R/print-tuning.R
\name{choose_bvar}
\alias{choose_bvar}
\alias{choose_bvhar}
\alias{print.bvharemp}
\alias{knit_print.bvharemp}
\title{Finding the Set of Hyperparameters of Bayesian Model}
\usage{
choose_bvar(
  bayes_spec = set_bvar(),
  lower = 0.01,
  upper = Inf,
  ...,
  eps = 1e-04,
  y,
  p,
  include_mean = TRUE
)

choose_bvhar(
  bayes_spec = set_bvhar(),
  lower = 0.01,
  upper = Inf,
  ...,
  eps = 1e-04,
  y,
  include_mean = TRUE
)

\method{print}{bvharemp}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvharemp(x, ...)
}
\arguments{
\item{bayes_spec}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} A Bayes model specification.}

\item{lower}{Lower bound. By default, \code{-Inf}.}

\item{upper}{Upper bound. By default, \code{Inf}.}

\item{...}{not used}

\item{eps}{Hyperparameter \code{eps} is fixed. By default, \code{1e-04}.}

\item{y}{Time series data}

\item{p}{BVAR lag}

\item{include_mean}{Add constant term (Default: \code{TRUE}) or not (\code{FALSE})}

\item{x}{\code{bvharemp} object}

\item{digits}{digit option to print}
}
\description{
This function chooses the set of hyperparameters of Bayesian model using \code{\link[stats:optim]{stats::optim()}} function.
}
\details{
Empirical Bayes method maximizes marginal likelihood and selects the set of hyperparameters.
}
