% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_cov}
\alias{compute_cov}
\title{Covariance Estimate for Residual Covariance Matrix}
\usage{
compute_cov(z, num_design, dim_design)
}
\arguments{
\item{z}{Matrix, residual}

\item{num_design}{Integer, Number of sample used (s = n - p)}

\item{dim_design}{Ingeger, Number of parameter for each dimension (k = mp + 1)}
}
\description{
Compute ubiased estimator for residual covariance.
}
\details{
See pp75 Lütkepohl (2007).
\itemize{
\item s = n - p: sample used (\code{num_design})
\item k = mp + 1 (m: dimension, p: VAR lag): number of parameter for each dimension (\code{dim_design})
}

Then an unbiased estimator for \eqn{\Sigma_e} is

\deqn{\hat{\Sigma}_e = \frac{1}{s - k} (Y_0 - \hat{A} X_0)^T (Y_0 - \hat{A} X_0)}
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing.
}
