% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{compute_dic}
\alias{compute_dic}
\title{Deviance Information Criterion}
\usage{
compute_dic(object, ...)
}
\arguments{
\item{object}{Model fit}

\item{...}{not used}
}
\value{
DIC value.
}
\description{
Generic function that computes DIC.
}
