% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-sparse.R
\name{conf_fscore}
\alias{conf_fscore}
\alias{conf_fscore.summary.bvharsp}
\title{Evaluate the Sparsity Estimation Based on F1 Score}
\usage{
conf_fscore(x, y, ...)

\method{conf_fscore}{summary.bvharsp}(x, y, truth_thr = 0, ...)
}
\arguments{
\item{x}{\code{summary.bvharsp} object.}

\item{y}{True inclusion variable.}

\item{...}{not used}

\item{truth_thr}{Threshold value when using non-sparse true coefficient matrix. By default, \code{0} for sparse matrix.}
}
\description{
This function computes F1 score for sparse element of the true coefficients given threshold.
}
\details{
The F1 score is computed by
\deqn{F_1 = \frac{2 precision \times recall}{precision + recall}}
}
\seealso{
\code{\link[=confusion]{confusion()}}
}
