% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-sparse.R
\name{conf_recall}
\alias{conf_recall}
\alias{conf_recall.ssvsmod}
\alias{conf_recall.mvhsmod}
\title{Evaluate the Sparsity Estimation Based on Recall}
\usage{
conf_recall(x, y, threshold, ...)

\method{conf_recall}{ssvsmod}(x, y, threshold = 0.01, restrict = FALSE, ...)

\method{conf_recall}{mvhsmod}(x, y, threshold = 0.01, ...)
}
\arguments{
\item{x}{Estimated model.}

\item{y}{True coefficient matrix.}

\item{threshold}{Threshold value indicating sparsity.}

\item{...}{not used}

\item{restrict}{Use restricted VAR. By default, \code{FALSE}.}
}
\description{
This function computes recall for sparse element of the true coefficients given threshold.
}
\details{
When using this function, the true coefficient matrix \eqn{\Phi} should be sparse.
If the element of the estimate \eqn{\hat\Phi} is smaller than some threshold,
it is treated to be zero.
Then the precision is computed by
\deqn{recall = \frac{TP}{TP + FN}}
where TP is true positive, and FN is false negative.
}
\references{
Bai, R., & Ghosh, M. (2018). High-dimensional multivariate posterior consistency under global–local shrinkage priors. Journal of Multivariate Analysis, 167, 157–170. doi:\href{https://doi.org/10.1016/j.jmva.2018.04.010}{10.1016/j.jmva.2018.04.010}
}
