% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-forecast.R
\name{divide_ts}
\alias{divide_ts}
\title{Split a Time Series Dataset into Train-Test Set}
\usage{
divide_ts(y, n_ahead)
}
\arguments{
\item{y}{Time series data of which columns indicate the variables}

\item{n_ahead}{step to evaluate}
}
\description{
Split a given time series dataset into train and test set for evaluation.
}
