% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_har}
\alias{estimate_har}
\title{Compute Vector HAR Coefficient Matrices and Fitted Values}
\usage{
estimate_har(x, y)
}
\arguments{
\item{x}{X0 processed by \code{\link{build_design}}}

\item{y}{Y0 processed by \code{\link{build_y0}}}
}
\description{
Compute Vector HAR Coefficient Matrices and Fitted Values
}
\details{
Given Y0 and Y0, the function estimate least squares
Y0 = X1 Phi + Z
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}

Corsi, F. (2008). \emph{A Simple Approximate Long-Memory Model of Realized Volatility}. Journal of Financial Econometrics, 7(2), 174–196. \url{https://doi:10.1093/jjfinec/nbp001}
}
