% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forecast_bvarmn}
\alias{forecast_bvarmn}
\title{Forecasting BVAR of Minnesota Prior}
\usage{
forecast_bvarmn(object, step)
}
\arguments{
\item{object}{\code{bvarmn} object by \code{\link{bvar_minnesota}}}

\item{step}{Integer, Step to forecast}
}
\description{
Forecasting BVAR of Minnesota Prior
}
\details{
n-step ahead forecasting using VAR(p) recursively, based on pp35 of Lütkepohl (2007).
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}

Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}
}
