% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forecast_bvarmn_flat}
\alias{forecast_bvarmn_flat}
\title{Forecasting BVAR(p) of Flat Prior}
\usage{
forecast_bvarmn_flat(object, step)
}
\arguments{
\item{object}{\code{bvarflat} object}

\item{step}{Integer, Step to forecast}
}
\description{
Forecasting BVAR(p) of Flat Prior
}
\details{
n-step ahead point forecasting using BVAR(p) recursively,
using point estimate for coefficient matrix.
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}

Ghosh, S., Khare, K., & Michailidis, G. (2018). \emph{High-Dimensional Posterior Consistency in Bayesian Vector Autoregressive Models}. Journal of the American Statistical Association, 114(526). \url{https://doi:10.1080/01621459.2018.1437043}
}
