% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forecast_bvharmn}
\alias{forecast_bvharmn}
\title{Forecasting Bayesian VHAR}
\usage{
forecast_bvharmn(object, step)
}
\arguments{
\item{object}{\code{bvharmn} object by \code{\link{vhar_lm}}}

\item{step}{Integer, Step to forecast}
}
\description{
Forecasting Bayesian VHAR
}
\details{
n-step ahead forecasting using VHAR recursively.
}
