% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-forecast.R
\name{forecast_expand}
\alias{forecast_expand}
\alias{forecast_expand.bvharmod}
\alias{forecast_expand.svmod}
\title{Out-of-sample Forecasting based on Expanding Window}
\usage{
forecast_expand(object, n_ahead, y_test, num_thread = 1, ...)

\method{forecast_expand}{bvharmod}(object, n_ahead, y_test, num_thread = 1, ...)

\method{forecast_expand}{svmod}(
  object,
  n_ahead,
  y_test,
  num_thread = 1,
  innovation = TRUE,
  use_fit = TRUE,
  ...
)
}
\arguments{
\item{object}{Model object}

\item{n_ahead}{Step to forecast in rolling window scheme}

\item{y_test}{Test data to be compared. Use \code{\link[=divide_ts]{divide_ts()}} if you don't have separate evaluation dataset.}

\item{num_thread}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Number of threads}

\item{...}{Additional arguments.}

\item{innovation}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Include heteroskedastic covariance of innovation when forecasting. By default, \code{TRUE}.}

\item{use_fit}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Use \code{object} result for the first window. By default, \code{TRUE}.}
}
\value{
\code{predbvhar_expand} \link{class}
}
\description{
This function conducts expanding window forecasting.
}
\details{
Expanding windows forecasting fixes the starting period.
It moves the window ahead and forecast h-ahead in \code{y_test} set.
}
\references{
Hyndman, R. J., & Athanasopoulos, G. (2021). \emph{Forecasting: Principles and practice} (3rd ed.). OTEXTS. \url{https://otexts.com/fpp3/}
}
\seealso{
See \link{ts_forecasting_cv} for out-of-sample forecasting methods.
}
