% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-forecast.R, R/print-forecast.R,
%   R/member.R
\name{forecast_roll}
\alias{forecast_roll}
\alias{print.bvharcv}
\alias{is.bvharcv}
\alias{knit_print.bvharcv}
\alias{forecast_roll.olsmod}
\alias{forecast_roll.normaliw}
\alias{forecast_roll.svmod}
\title{Out-of-sample Forecasting based on Rolling Window}
\usage{
forecast_roll(object, n_ahead, y_test, num_thread = 1, ...)

\method{print}{bvharcv}(x, digits = max(3L, getOption("digits") - 3L), ...)

is.bvharcv(x)

\method{knit_print}{bvharcv}(x, ...)

\method{forecast_roll}{olsmod}(object, n_ahead, y_test, num_thread = 1, ...)

\method{forecast_roll}{normaliw}(object, n_ahead, y_test, num_thread = 1, use_fit = TRUE, ...)

\method{forecast_roll}{svmod}(
  object,
  n_ahead,
  y_test,
  num_thread = 1,
  use_sv = TRUE,
  sparse = FALSE,
  lpl = FALSE,
  use_fit = TRUE,
  ...
)
}
\arguments{
\item{object}{Model object}

\item{n_ahead}{Step to forecast in rolling window scheme}

\item{y_test}{Test data to be compared. Use \code{\link[=divide_ts]{divide_ts()}} if you don't have separate evaluation dataset.}

\item{num_thread}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Number of threads}

\item{...}{not used}

\item{x}{\code{bvharcv} object}

\item{digits}{digit option to print}

\item{use_fit}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Use \code{object} result for the first window. By default, \code{TRUE}.}

\item{use_sv}{Use SV term}

\item{sparse}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Apply restriction. By default, \code{FALSE}.}

\item{lpl}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Compute log-predictive likelihood (LPL). By default, \code{FALSE}.}
}
\value{
\code{predbvhar_roll} \link{class}
}
\description{
This function conducts rolling window forecasting.
}
\details{
Rolling windows forecasting fixes window size.
It moves the window ahead and forecast h-ahead in \code{y_test} set.
}
\references{
Hyndman, R. J., & Athanasopoulos, G. (2021). \emph{Forecasting: Principles and practice} (3rd ed.). OTEXTS.
}
