% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-varlse.R
\name{FPE}
\alias{FPE}
\alias{FPE.varlse}
\title{FPE}
\usage{
FPE(object, ...)

\method{FPE}{varlse}(object, ...)
}
\arguments{
\item{object}{model fit}

\item{...}{not used}
}
\description{
FPE
}
\section{Methods (by class)}{
\itemize{
\item \code{varlse}: of VAR(p)
}}

