% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{FPE.varlse}
\alias{FPE.varlse}
\alias{FPE.vharlse}
\title{Final Prediction Error Criterion of Multivariate Time Series Model}
\usage{
\method{FPE}{varlse}(object, ...)

\method{FPE}{vharlse}(object, ...)
}
\arguments{
\item{object}{Model fit}

\item{...}{not used}
}
\value{
FPE value.
}
\description{
Compute FPE of VAR(p), VHAR, BVAR(p), and BVHAR
}
\details{
Let \eqn{\tilde{\Sigma}_e} be the MLE
and let \eqn{\hat{\Sigma}_e} be the unbiased estimator (\code{covmat}) for \eqn{\Sigma_e}.
Note that

\deqn{\tilde{\Sigma}_e = \frac{s - k}{n} \hat{\Sigma}_e}

Then

\deqn{FPE(p) = (\frac{s + k}{s - k})^m \det \tilde{\Sigma}_e}
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing.
}
