% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{geom_eval}
\alias{geom_eval}
\title{Adding Test Data Layer}
\usage{
geom_eval(data, colour = "red", ...)
}
\arguments{
\item{data}{Test data to draw, which has the same format with the train data.}

\item{colour}{Color of the line (By default, \code{"red"}).}

\item{...}{Other arguments passed on the \code{\link[ggplot2:geom_path]{ggplot2::geom_path()}}.}
}
\description{
This function adds a layer of test dataset.
}
