% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{geom_predbvhar}
\alias{geom_predbvhar}
\title{Forecasting Lines and Region}
\usage{
geom_predbvhar(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ci_param = list(),
  line_param = list(),
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes}} or \code{\link[ggplot2:aes_]{ggplot2::aes_}}.
If specified and \code{inherit.aes = TRUE} (the default), it is combined with the default mapping at the top level of the plot.
You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer.
This should be generated by interal function, \code{gather_predbvhar}.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{ci_param}{Parameter lists for \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon}}. See details.}

\item{line_param}{Parameter lists for \code{\link[ggplot2:geom_path]{ggplot2::geom_path}}}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them.
This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification,
e.g. \code{\link[ggplot2:borders]{ggplot2::borders}}.}
}
\description{
This function draws h-step ahead forecasts their confidence (credible) region.
}
\details{
\code{gather_predbvhar} produces a \code{tibble} with columns named

\itemize{
\item forecast - If the row is observed one or forecasted one
\item id - Index of the row
\item variable - Name of the variable
\item value_forecast - Point forecasts
\item value_lower_joint - Lower CI
\item value_upper_joint - Upper CI
\item model - Fitting Model
}

In case of \code{ci_param}, since it is controlling CI, try to specify \code{alpha} and \code{colour}.
}
