% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{HQ.varlse}
\alias{HQ.varlse}
\alias{HQ.vharlse}
\title{Hannan-Quinn Criterion of Multivariate Time Series Model}
\usage{
\method{HQ}{varlse}(object, ...)

\method{HQ}{vharlse}(object, ...)
}
\arguments{
\item{object}{Model fit}

\item{...}{not used}
}
\description{
Compute HQ of VAR(p), VHAR, BVAR(p), and BVHAR
}
\details{
Let \eqn{\tilde{\Sigma}_e} be the MLE
and let \eqn{\hat{\Sigma}_e} be the unbiased estimator (from \code{\link[=compute_cov]{compute_cov()}} and the member named \code{covmat}) for \eqn{\Sigma_e}.
Note that

\deqn{\tilde{\Sigma}_e = \frac{s - k}{n} \hat{\Sigma}_e}

Then

\deqn{HQ(p) = \log \det \Sigma_e + \frac{2 \log \log s}{s}(\text{number of freely estimated parameters})}

where the number of freely estimated parameters is \eqn{pm^2}.
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
