% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R, R/print-bvharspec.R
\name{init_sv}
\alias{init_sv}
\alias{print.svinit}
\title{Initialize Stochastic Volatility Parameters}
\usage{
init_sv(lvol = 0, lvol_init = 0.1, lvol_sig = 0.1, type = c("user", "auto"))

\method{print}{svinit}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{lvol}{Time-varying log-volatility.
The length is same with data dimension times sample size (\eqn{n = T - p}).}

\item{lvol_init}{Initial state of log-volatility.
The length is same with data dimension.}

\item{lvol_sig}{Variance of log-volatility.
The length is same with data dimension.}

\item{type}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Type to choose initial values. One of \code{"user"} (User-given) and \code{"auto"} (OLS for coefficients and 1 for dummy).}

\item{x}{\code{svinit}}

\item{digits}{digit option to print}

\item{...}{not used}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} This function initializes stochastic volatility parameters in Gibbs sampler.
}
