% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member.R
\name{is.varlse}
\alias{is.varlse}
\alias{is.vharlse}
\alias{is.bvarmn}
\alias{is.bvarflat}
\alias{is.bvharmn}
\alias{is.predbvhar}
\alias{is.bvharcv}
\alias{is.bvharspec}
\alias{is.bvharss_coef}
\alias{is.bvharss_sig}
\title{See if the Object a class in this package}
\usage{
is.varlse(x)

is.vharlse(x)

is.bvarmn(x)

is.bvarflat(x)

is.bvharmn(x)

is.predbvhar(x)

is.bvharcv(x)

is.bvharspec(x)

is.bvharss_coef(x)

is.bvharss_sig(x)
}
\arguments{
\item{x}{Object}
}
\value{
logical class
}
\description{
This function returns \code{TRUE} if the input is the \link{class} defined by this package.
}
