% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member.R
\name{is.varlse}
\alias{is.varlse}
\alias{is.vharlse}
\alias{is.bvarmn}
\alias{is.bvarflat}
\alias{is.bvharmn}
\alias{is.predbvhar}
\alias{is.bvharcv}
\alias{is.bvharspec}
\alias{is.bvharpriorspec}
\alias{is.bvharemp}
\alias{is.boundbvharemp}
\alias{is.ssvsinput}
\alias{is.ssvsinit}
\alias{is.horseshoespec}
\alias{is.svspec}
\title{See if the Object a class in this package}
\usage{
is.varlse(x)

is.vharlse(x)

is.bvarmn(x)

is.bvarflat(x)

is.bvharmn(x)

is.predbvhar(x)

is.bvharcv(x)

is.bvharspec(x)

is.bvharpriorspec(x)

is.bvharemp(x)

is.boundbvharemp(x)

is.ssvsinput(x)

is.ssvsinit(x)

is.bvharpriorspec(x)

is.horseshoespec(x)

is.svspec(x)
}
\arguments{
\item{x}{Object}
}
\value{
logical class
}
\description{
This function returns \code{TRUE} if the input is the \link{class} defined by this package.
}
