% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{logLik.varlse}
\alias{logLik.varlse}
\alias{logLik.vharlse}
\title{Extract Log-Likelihood of Multivariate Time Series Model}
\usage{
\method{logLik}{varlse}(object, ...)

\method{logLik}{vharlse}(object, ...)
}
\arguments{
\item{object}{Model fit}

\item{...}{not used}
}
\description{
Compute log-likelihood function value of VAR(p), VHAR, BVAR(p), and BVHAR
}
\details{
Consider Y0 matrix from \code{\link[=build_y0]{build_y0()}}.
Likelihood of VAR(p) has

\deqn{Y_0 \sim MN(X_0 B, I_s, \Sigma_e)}

where X0 from \code{\link[=build_design]{build_design()}}.


}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
