% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-forecast.R
\name{mae}
\alias{mae}
\alias{mae.predbvhar}
\alias{mae.bvharcv}
\title{Evaluate the Model Based on MAE (Mean Absolute Error)}
\usage{
mae(x, y, ...)

\method{mae}{predbvhar}(x, y, ...)

\method{mae}{bvharcv}(x, y, ...)
}
\arguments{
\item{x}{Forecasting object}

\item{y}{Test data to be compared. should be the same format with the train data.}

\item{...}{not used}
}
\description{
This function computes MAE given prediction result versus evaluation set.
}
\details{
Let \eqn{e_t = y_t - \hat{y}_t}.
MAE is defined by

\deqn{MSE = mean(\lvert e_t \rvert)}

Some researchers prefer MAE to MSE because it is less sensitive to outliers.
}
\references{
Hyndman, R. J., & Koehler, A. B. (2006). \emph{Another look at measures of forecast accuracy}. International Journal of Forecasting, 22(4), 679–688.
}
