% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-forecast.R
\name{mape}
\alias{mape}
\title{Evaluate the Model Based on MAPE (Mean Absolute Percentage Error)}
\usage{
mape(x, y, ...)
}
\arguments{
\item{x}{\code{predbvhar} object}

\item{y}{test data to be compared}

\item{...}{not used}
}
\description{
Evaluate the Model Based on MAPE (Mean Absolute Percentage Error)
}
