% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_loss}
\alias{plot_loss}
\title{Compare Lists of Models}
\usage{
plot_loss(mod_list, y, type = c("mse", "mape"), ...)
}
\arguments{
\item{mod_list}{lists of forecast results (\code{predbvhar} objects)}

\item{y}{test data to be compared. should be the same format with the train data and predict$forecast.}

\item{type}{loss function to be used (\code{"mse"}: MSE, \code{mape}: MAPE)}

\item{...}{additional options for \code{\link[ggplot2:geom_line]{ggplot2::geom_line}}}
}
\description{
Draw plot of test error for given models
}
