% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{predict.bvarflat}
\alias{predict.bvarflat}
\title{Predict Method for \code{bvarghosh} object}
\usage{
\method{predict}{bvarflat}(object, n.ahead, n_iter = 100L, level = 0.05, ...)
}
\arguments{
\item{object}{\code{bvarghosh} object}

\item{n.ahead}{step to forecast}

\item{n_iter}{Number to sample residual matrix from inverse-wishart distribution. By default, 100.}

\item{level}{Specify alpha of confidence interval level 100(1 - alpha) percentage. By default, .05.}

\item{...}{not used}
}
\value{
\code{predbvhar} \link{class} with the following components:
\describe{
\item{process}{object class: bvarghosh}
\item{forecast}{forecast matrix}
\item{se}{standard error matrix}
\item{lower}{lower confidence interval}
\item{upper}{upper confidence interval}
\item{lower_joint}{lower CI adjusted (Bonferroni)}
\item{upper_joint}{upper CI adjusted (Bonferroni)}
\item{y}{bvarflat$y}
}
}
\description{
Point forecasting for Nonhierarchical Ghosh BVAR(p)
}
\details{
n-step ahead forecasting using BVAR recursively
}
\references{
Ghosh, S., Khare, K., & Michailidis, G. (2018). \emph{High-Dimensional Posterior Consistency in Bayesian Vector Autoregressive Models}. Journal of the American Statistical Association, 114(526). \url{https://doi:10.1080/01621459.2018.1437043}

Gelman, A., Carlin, J. B., Stern, H. S., & Rubin, D. B. (2013). \emph{Bayesian data analysis}. Chapman and Hall/CRC. \url{http://www.stat.columbia.edu/~gelman/book/}
}
