% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{predict.bvarghosh}
\alias{predict.bvarghosh}
\title{Predict Method for \code{bvarghosh} object}
\usage{
\method{predict}{bvarghosh}(object, n.ahead, ...)
}
\arguments{
\item{object}{\code{bvarghosh} object}

\item{n.ahead}{step to forecast}

\item{...}{not used}
}
\value{
matrix
}
\description{
Point forecasting for Nonhierarchical Ghosh BVAR(p)
}
\details{
n-step ahead forecasting using BVAR recursively
}
