% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{predict.bvharmn}
\alias{predict.bvharmn}
\title{Predict Method for \code{bvharmn} object}
\usage{
\method{predict}{bvharmn}(object, n.ahead, ...)
}
\arguments{
\item{object}{\code{vharlse} object}

\item{n.ahead}{step to forecast}

\item{...}{not used}
}
\value{
\code{predbvhar} \link{class} with the following components:
\describe{
\item{process}{object class: bvharmn}
\item{forecast}{forecast matrix}
\item{y}{bvharmn$y}
}
}
\description{
Forecasting BVHAR
}
\details{
n-step ahead forecasting using VHAR recursively
}
