% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{predict.varlse}
\alias{predict.varlse}
\title{Predict Method for \code{varlse} object}
\usage{
\method{predict}{varlse}(object, n.ahead, ...)
}
\arguments{
\item{object}{\code{varlse} object}

\item{n.ahead}{step to forecast}

\item{...}{not used}
}
\value{
matrix
}
\description{
Forecasting VAR(p)
}
\details{
n-step ahead forecasting using VAR(p) recursively.
See pp35 of Lütkepohl (2007).
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
