% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{predict.varlse}
\alias{predict.varlse}
\title{Predict Method for \code{varlse} object}
\usage{
\method{predict}{varlse}(object, n.ahead, ...)
}
\arguments{
\item{object}{\code{varlse} object}

\item{n.ahead}{step to forecast}

\item{...}{not used}
}
\value{
matrix
}
\description{
Forecasting VAR(p)
}
\details{
n-step ahead forecasting using VAR(p) recursively
}
