% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{predict.vharlse}
\alias{predict.vharlse}
\title{Predict Method for \code{vharlse} object}
\usage{
\method{predict}{vharlse}(object, n.ahead, ...)
}
\arguments{
\item{object}{\code{vharlse} object}

\item{n.ahead}{step to forecast}

\item{...}{not used}
}
\value{
\code{predbvhar} \link{class} with the following components:
\describe{
\item{process}{object class: vharlse}
\item{forecast}{forecast matrix}
\item{y}{vharlse$y}
}
}
\description{
Forecasting VHAR
}
\details{
n-step ahead forecasting using VHAR recursively
}
