% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scale_har}
\alias{scale_har}
\title{Building a Linear Transformation Matrix for Vector HAR}
\usage{
scale_har(m)
}
\arguments{
\item{m}{integer, dimension}
}
\description{
This function produces a linear transformation matrix for VHAR for given dimension.
}
\details{
VHAR is linearly restricted VAR(22) in \eqn{Y_0 = X_0 A + Z}.

\deqn{Y_0 = X_1 \Phi + Z = (X_0 T_{HAR}^T) \Phi + Z}

This function computes above \eqn{T_{HAR}}.
}
