% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam.R
\name{set_bvar}
\alias{set_bvar}
\alias{set_bvar_flat}
\alias{set_bvhar}
\alias{set_weight_bvhar}
\alias{print.bvharspec}
\alias{knit_print.bvharspec}
\title{Hyperparameters for Bayesian Models}
\usage{
set_bvar(sigma, lambda = 0.1, delta, eps = 1e-04)

set_bvar_flat(U)

set_bvhar(sigma, lambda = 0.1, delta, eps = 1e-04)

set_weight_bvhar(sigma, lambda = 0.1, eps = 1e-04, daily, weekly, monthly)

\method{print}{bvharspec}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvharspec(x, ...)
}
\arguments{
\item{sigma}{Standard error vector for each variable (Default: sd)}

\item{lambda}{Tightness of the prior around a random walk or white noise (Default: .1)}

\item{delta}{Persistence (Default: Litterman sets 1 = random walk prior, White noise prior = 0)}

\item{eps}{Very small number (Default: 1e-04)}

\item{U}{Positive definite matrix. By default, identity matrix of dimension ncol(X0)}

\item{daily}{Same as delta in VHAR type (Default: 1 as Litterman)}

\item{weekly}{Fill the second part in the first block (Default: 1)}

\item{monthly}{Fill the third part in the first block (Default: 1)}

\item{x}{\code{bvharspec} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
Every function returns \code{bvharspec} \link{class}.
It is the list of which the components are the same as the arguments provided.
}
\description{
Set hyperparameters of Bayesian VAR and VHAR models.
}
\details{
\itemize{
\item Missing arguments will be set to be default values in each model function mentioned above.
\item \code{set_bvar} sets hyperparameters for \code{\link[=bvar_minnesota]{bvar_minnesota()}}.
}

\itemize{
\item \code{set_bvar_flat} sets hyperparameters for \code{\link[=bvar_flat]{bvar_flat()}}.
}

\itemize{
\item \code{set_bvhar} sets hyperparameters for \code{\link[=bvhar_minnesota]{bvhar_minnesota()}} with VAR-type Minnesota prior.
}

\itemize{
\item \code{set_weight_bvhar} sets hyperparameters for \code{\link[=bvhar_minnesota]{bvhar_minnesota()}} with VHAR-type Minnesota prior.
}
}
\examples{
# Minnesota BVAR specification
# m = 3
\dontrun{
  bvar_spec <- set_bvar(
    sigma = c(.03, .02, .01), # Sigma = diag(.03^2, .02^2, .01^2)
    lambda = .2, # lambda = .2
    delta = rep(.1, 3), # delta1 = .1, delta2 = .1, delta3 = .1
    eps = 1e-04 # eps = 1e-04
  )
  class(bvar_spec)
  str(bvar_spec)
}
# Flat BVAR specification
# m = 3
# p = 5 with constant term
# U = 500 * I(mp + 1)
\dontrun{
  bvar_spec <- set_bvar_flat(U = 500 * diag(16))
  class(bvar_spec)
  str(bvar_spec)
}
# VAR-type Minnesota BVHAR specification
# m = 3
\dontrun{
  bvhar_spec <- set_bvhar(
    sigma = c(.03, .02, .01), # Sigma = diag(.03^2, .02^2, .01^2)
    lambda = .2, # lambda = .2
    delta = rep(.1, 3), # delta1 = .1, delta2 = .1, delta3 = .1
    eps = 1e-04 # eps = 1e-04
  )
  class(bvhar_spec)
  str(bvhar_spec)
}
# HAR-type Minnesota BVHAR specification
# m = 3
\dontrun{
  bvhar_spec <- set_weight_bvhar(
    sigma = c(.03, .02, .01), # Sigma = diag(.03^2, .02^2, .01^2)
    lambda = .2, eps = 1e-04, # lambda = .2
    eps = 1e-04, # eps = 1e-04
    daily = rep(.2, 3), # daily1 = .2, daily2 = .2, daily3 = .2
    weekly = rep(.1, 3), # weekly1 = .1, weekly2 = .1, weekly3 = .1
    monthly = rep(.05, 3) # monthly1 = .05, monthly2 = .05, monthly3 = .05
  )
  class(bvhar_spec)
  str(bvhar_spec)
}
}
