% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam.R
\name{set_factor}
\alias{set_factor}
\title{Factor Specification}
\usage{
set_factor(size_factor = 0, factor_lag = 0)
}
\arguments{
\item{size_factor}{Factor dimension. If \code{0}, factor is not used.}

\item{factor_lag}{Lag for the factor autoregressions.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Set factor size and prior for factor.
}
\references{
Korobilis, D. (2022). A new algorithm for structural restrictions in Bayesian vector autoregressions. European Economic Review, 148, 104241.
}
