% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam.R, R/print-bvharspec.R
\name{set_horseshoe}
\alias{set_horseshoe}
\alias{print.horseshoespec}
\alias{knit_print.horseshoespec}
\title{Horseshoe Prior Specification}
\usage{
set_horseshoe(process = c("VAR", "VHAR"))

\method{print}{horseshoespec}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.horseshoespec(x, ...)
}
\arguments{
\item{process}{Time series process. \code{"VAR"} (default) or \code{"VHAR"}.}

\item{x}{\code{ssvsspec}}

\item{digits}{digit option to print}

\item{...}{not used}
}
\description{
Set Horseshoe prior.
}
