% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar-hierarchical.R
\name{set_lambda}
\alias{set_lambda}
\title{Hyperpriors for Bayesian Models}
\usage{
set_lambda(mode, sd)
}
\arguments{
\item{mode}{Mode of Gamma distribution}

\item{sd}{Standard deviation of Gamma distribution}
}
\description{
Set hyperpriors of Bayesian VAR and VHAR models.
}
\details{
Parameters of Gamma distribution is computed using \href{https://en.wikipedia.org/wiki/Quadratic_formula}{quadratic formula}.
}
\references{
Giannone, D., Lenza, M., & Primiceri, G. E. (2015). \emph{Prior Selection for Vector Autoregressions}. Review of Economics and Statistics, 97(2). doi:\href{https://doi.org/10.1162/REST_a_00483}{10.1162/REST_a_00483}
}
