% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam-sparse.R, R/print-bvharspec.R
\name{set_spikeslab_coef}
\alias{set_spikeslab_coef}
\alias{print.bvharss_coef}
\alias{knit_print.bvharss_coef}
\title{Spike and Slab Hyperparameter for VAR Coefficient}
\usage{
set_spikeslab_coef(spike_sd = NULL, slab_sd = NULL, prop_sparse = NULL)

\method{print}{bvharss_coef}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvharss_coef(x, ...)
}
\arguments{
\item{spike_sd}{Standard deviance for Spike normal distribution (See Details).}

\item{slab_sd}{Standard deviance for Slab normal distribution (See Details).}

\item{prop_sparse}{Bernoulli parameter for sparsity proportion (See Details).}

\item{x}{\code{bvharss_coef} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\description{
Set Hyperparameters for VAR coefficient matrix.
}
\details{
Let \eqn{\alpha} be the vectorized coefficient,
\deqn{\alpha = vec(A)}
Spike-slab prior is given using two normal distributions.
\deqn{\alpha_j \mid \gamma_j \sim (1 - \gamma_j) N(0, \kappa_{0j}^2) + \gamma_j N(0, \kappa_{1j}^2)}
As spike-slab prior itself suggests, set \eqn{\kappa_{0j}} small (point mass at zero: spike distribution)
and set \eqn{\kappa_{1j}} large (symmetric by zero: slab distribution).

\eqn{\gamma_j} is the proportion of the nonzero coefficients and it follows
\deqn{\gamma_j \sim Bernoulli(q_j)}
\itemize{
\item \code{spike_sd}: \eqn{\kappa_{0j}}
\item \code{slab_sd}: \eqn{\kappa_{1j}}
\item \code{prop_sparse}: \eqn{q_j}
\item \eqn{j = 1, \ldots, mk}: vectorized format corresponding to coefficient matrix
}
}
\references{
Jochmann, M., Koop, G., & Strachan, R. W. (2010). \emph{Bayesian forecasting using stochastic search variable selection in a VAR subject to breaks}. International Journal of Forecasting, 26(2), 326–347. doi:\href{https://www.sciencedirect.com/science/article/abs/pii/S0169207009001782?via\%3Dihub}{10.1016/j.ijforecast.2009.11.002}

George, E. I., Sun, D., & Ni, S. (2008). \emph{Bayesian stochastic search for VAR model restrictions}. Journal of Econometrics, 142(1), 553–580. doi:\href{https://www.sciencedirect.com/science/article/abs/pii/S0304407607001753?via\%3Dihub}{10.1016/j.jeconom.2007.08.017}
}
