% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam-sparse.R
\name{set_ssvs}
\alias{set_ssvs}
\title{Hyperparameters for SSVS Model}
\usage{
set_ssvs(coef_ss = set_spikeslab_coef(), sig_ss = set_spikeslab_cov())
}
\arguments{
\item{coef_ss}{Spike and slab specification for coefficient matrix.}

\item{sig_ss}{Spike and slab specification for covariance matrix.}
}
\description{
Set hyperparameters of stochastic search variable selection for Bayesian VAR model.
}
\details{
Setting for MCMC.
}
\references{
Jochmann, M., Koop, G., & Strachan, R. W. (2010). \emph{Bayesian forecasting using stochastic search variable selection in a VAR subject to breaks}. International Journal of Forecasting, 26(2), 326–347. doi:\href{https://www.sciencedirect.com/science/article/abs/pii/S0169207009001782?via\%3Dihub}{10.1016/j.ijforecast.2009.11.002}

George, E. I., Sun, D., & Ni, S. (2008). \emph{Bayesian stochastic search for VAR model restrictions}. Journal of Econometrics, 142(1), 553–580. doi:\href{https://www.sciencedirect.com/science/article/abs/pii/S0304407607001753?via\%3Dihub}{10.1016/j.jeconom.2007.08.017}
}
