% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_mgaussian}
\alias{sim_mgaussian}
\title{Generate Multivariate Normal Random Vector with Zero Mean}
\usage{
sim_mgaussian(num_sim, sig)
}
\arguments{
\item{num_sim}{Number to generate process}

\item{sig}{Variance matrix}
}
\description{
This function samples n x muti-dimensional normal random matrix with zero mean vector.
}
