% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-param.R
\name{sim_mnvhar_coef}
\alias{sim_mnvhar_coef}
\title{Generate Minnesota BVAR Parameters}
\usage{
sim_mnvhar_coef(
  type = c("VAR", "VHAR"),
  sigma,
  lambda,
  delta,
  daily,
  weekly,
  monthly,
  eps = 1e-04
)
}
\arguments{
\item{type}{Prior mean type to apply. ("VAR" = Original Minnesota. "VHAR" = fills zero in the first block)}

\item{sigma}{Standard error vector for each variable}

\item{lambda}{Tightness of the prior around a random walk or white noise}

\item{delta}{Prior belief about white noise (Litterman sets 1: default)}

\item{daily}{Same as delta in VHAR type}

\item{weekly}{Fill the second part in the first block}

\item{monthly}{Fill the third part in the first block}

\item{eps}{Very small number}
}
\description{
This function generates parameters of BVAR with Minnesota prior.
}
\details{
Normal-IW family for vector HAR model:
\deqn{\Phi \mid \Sigma_e \sim MN(P_0, \Psi_0, \Sigma_e)}
\deqn{\Sigma_e \sim IW(U_0, d_0)}
}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}

Corsi, F. (2008). \emph{A Simple Approximate Long-Memory Model of Realized Volatility}. Journal of Financial Econometrics, 7(2), 174–196. \url{https://doi:10.1093/jjfinec/nbp001}
}
\seealso{
\code{\link{build_ydummy}}, \code{\link{build_xdummy}}, and \code{\link{build_ydummy_bvhar}} for defining dummy observation.
}
