% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_var}
\alias{sim_var}
\title{Generate Multivariate Time Series Process Following VAR(p)}
\usage{
sim_var(num_sim, num_burn, var_coef, var_lag, sig_error, init)
}
\arguments{
\item{num_sim}{Number to generated process}

\item{num_burn}{Number of burn-in}

\item{var_coef}{VAR coefficient. The format should be the same as the output of \code{\link[=var_lm]{var_lm()}}}

\item{var_lag}{Lag of VAR}

\item{sig_error}{Variance matrix of the error term. Try \code{diag(dim)}.}

\item{init}{Initial y1, ..., yp matrix to simulate VAR model. Try \code{matrix(0L, nrow = var_lag, ncol = dim)}.}
}
\description{
This function generates multivariate time series dataset that follows VAR(p).
}
\details{
\enumerate{
\item Generate \eqn{\epsilon_1, \epsilon_n \sim N(0, \Sigma)}
\item For i = 1, ... n,
\deqn{y_{p + i} = (y_{p + i - 1}^T, \ldots, y_i^T, 1)^T B + \epsilon_i}
\item Then the output is \eqn{(y_{p + 1}, \ldots, y_{n + p})^T}
}

Initial values might be set to be zero vector or \eqn{(I_m - A_1 - \cdots - A_p)^{-1} c}.
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
