% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-varlse.R
\name{stableroot}
\alias{stableroot}
\title{Roots of characteristic polynomial}
\usage{
stableroot(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{not used}
}
\description{
Roots of characteristic polynomial
}
