% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-process.R
\name{stableroot}
\alias{stableroot}
\alias{stableroot.varlse}
\alias{stableroot.vharlse}
\alias{stableroot.bvarmn}
\alias{stableroot.bvarflat}
\alias{stableroot.bvharmn}
\title{Roots of characteristic polynomial}
\usage{
stableroot(x, ...)

\method{stableroot}{varlse}(x, ...)

\method{stableroot}{vharlse}(x, ...)

\method{stableroot}{bvarmn}(x, ...)

\method{stableroot}{bvarflat}(x, ...)

\method{stableroot}{bvharmn}(x, ...)
}
\arguments{
\item{x}{Model fit}

\item{...}{not used}
}
\value{
Numeric vector.
}
\description{
Compute the character polynomial of coefficient matrix.
}
\details{
To know whether the process is stable or not, make characteristic polynomial.

\deqn{\det(I_m - A z) = 0}

where \eqn{A} is VAR(1) coefficient matrix representation.
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing.
}
