% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-process.R
\name{stableroot.varlse}
\alias{stableroot.varlse}
\alias{stableroot.vharlse}
\alias{stableroot.bvarmn}
\alias{stableroot.bvarflat}
\alias{stableroot.bvharmn}
\title{Characteristic polynomial roots for VAR Coefficient Matrix}
\usage{
\method{stableroot}{varlse}(x, ...)

\method{stableroot}{vharlse}(x, ...)

\method{stableroot}{bvarmn}(x, ...)

\method{stableroot}{bvarflat}(x, ...)

\method{stableroot}{bvharmn}(x, ...)
}
\arguments{
\item{x}{Model fit}

\item{...}{not used}
}
\description{
Compute the character polynomial of VAR(p) coefficient matrix.
}
\details{
To know whether the process is stable or not, make characteristic polynomial.

\deqn{\det(I_m - A z) = 0}

where \eqn{A} is VAR(1) matrix defined in \code{\link[=compute_stablemat]{compute_stablemat()}}.
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
