% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-bayes.R, R/print-bvarmn.R,
%   R/print-bvharmn.R
\name{summary.bvarmn}
\alias{summary.bvarmn}
\alias{print.summary.bvarmn}
\alias{knit_print.summary.bvarmn}
\alias{knit_print.summary.bvharmn}
\title{Summarizing Bayesian VAR of Minnesota Prior Model}
\usage{
\method{summary}{bvarmn}(object, n_iter = 100L, ...)

\method{print}{summary.bvarmn}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.summary.bvarmn(x, ...)

knit_print.summary.bvharmn(x, ...)
}
\arguments{
\item{object}{\code{bvarmn} object}

\item{n_iter}{Number to sample Matrix Normal Inverse-Wishart distribution}

\item{...}{not used}

\item{x}{\code{summary.bvharmn} object}

\item{digits}{digit option to print}
}
\value{
\code{summary} for \code{bvarmn} object returns \code{summary.bvarmn} \link{class}.
\describe{
\item{coefficients}{iter x k x m array: each column of the array indicate the draw for each lag corresponding to that variable}
\item{covmat}{iter x m x m array: each column of teh array indicate the draw for each varable corresponding to that variable}
}
}
\description{
\code{summary} method for \code{bvarmn} class.
}
\details{
From Minnesota prior, set of coefficient matrices and residual covariance matrix have matrix Normal Inverse-Wishart distribution.

\deqn{(A, \Sigma_e) \sim MNIW(\hat{A}, \hat{V}^{-1}, \hat\Sigma_e, \alpha_0 + s)}
where \eqn{\hat{V} = X_\ast^T X_\ast} is the posterior precision of MN.
}
\references{
Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}
}
