% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-sparse.R
\name{summary.mvhsmod}
\alias{summary.mvhsmod}
\title{Summarizing VAR and VHAR with Matrix-variate Horseshoe Prior Model}
\usage{
\method{summary}{mvhsmod}(object, level = 0.05, ...)
}
\arguments{
\item{object}{\code{mvhsmod} object}

\item{level}{Specify alpha of credible interval level 100(1 - alpha) percentage. By default, \code{.05}.}

\item{...}{not used}
}
\description{
Conduct variable selection.
}
\details{
MCMC can construct \eqn{100 (1 - \alpha)} credible interval.
This interval can help variable selection.
}
\references{
Bai, R., & Ghosh, M. (2018). High-dimensional multivariate posterior consistency under global–local shrinkage priors. Journal of Multivariate Analysis, 167, 157–170. doi:\href{https://doi.org/10.1016/j.jmva.2018.04.010}{10.1016/j.jmva.2018.04.010}
}
