% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-sparse.R
\name{summary.ssvsmod}
\alias{summary.ssvsmod}
\title{Summarizing VAR and VHAR with SSVS Prior Model}
\usage{
\method{summary}{ssvsmod}(
  object,
  coef_threshold = 0.5,
  chol_threshold = 0.5,
  restrict = FALSE,
  ...
)
}
\arguments{
\item{object}{\code{ssvsmod} object}

\item{coef_threshold}{Threshold for variable selection. By default, \code{0.5}.}

\item{chol_threshold}{Threshold for variable selection in cholesky factor. By default, \code{0.5}.}

\item{restrict}{Use restricted VAR. By default, \code{FALSE}.}

\item{...}{not used}
}
\description{
Conduct variable selection.
}
\references{
George, E. I., & McCulloch, R. E. (1993). \emph{Variable Selection via Gibbs Sampling}. Journal of the American Statistical Association, 88(423), 881–889. doi:\href{https://www.tandfonline.com/doi/abs/10.1080/01621459.1993.10476353}{10.1080/01621459.1993.10476353}

George, E. I., Sun, D., & Ni, S. (2008). \emph{Bayesian stochastic search for VAR model restrictions}. Journal of Econometrics, 142(1), 553–580. doi:\href{https://doi.org/10.1016/j.jeconom.2007.08.017}{10.1016/j.jeconom.2007.08.017}

Koop, G., & Korobilis, D. (2009). \emph{Bayesian Multivariate Time Series Methods for Empirical Macroeconomics}. Foundations and Trends® in Econometrics, 3(4), 267–358. doi:\href{http://dx.doi.org/10.1561/0800000013}{10.1561/0800000013}

O’Hara, R. B., & Sillanpää, M. J. (2009). \emph{A review of Bayesian variable selection methods: what, how and which}. Bayesian Analysis, 4(1), 85–117. doi:\href{https://doi.org/10.1214/09-BA403}{10.1214/09-ba403}
}
