% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-varlse.R, R/summary-varlse.R
\name{print.summary.varlse}
\alias{print.summary.varlse}
\alias{summary.varlse}
\title{Summary of varlse}
\usage{
\method{print}{summary.varlse}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{summary}{varlse}(object, ...)
}
\arguments{
\item{...}{not used}

\item{object}{\code{varlse} object}
}
\description{
Summary of varlse
}
