% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlse.R, R/print-varlse.R
\name{var_lm}
\alias{var_lm}
\alias{print.varlse}
\alias{knit_print.varlse}
\title{Fit Vector Autoregressive Model of Order p Model}
\usage{
var_lm(y, p, include_mean = TRUE)

\method{print}{varlse}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.varlse(x, ...)
}
\arguments{
\item{y}{Time series data of which columns indicate the variables}

\item{p}{integer, lags of VAR}

\item{include_mean}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Add constant term (Default: \code{TRUE}) or not (\code{FALSE})}

\item{x}{\code{varlse} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{var_lm} returns an object named \code{varlse} \link{class}.
It is a list with the following components:

\describe{
\item{design}{\eqn{X_0}}
\item{y0}{\eqn{Y_0}}
\item{y}{Raw input}
\item{p}{Lag of VAR}
\item{m}{Dimension of the data}
\item{df}{Numer of Coefficients: mp + 1}
\item{obs}{Sample size used when training = \code{totobs} - \code{p}}
\item{totobs}{Total number of the observation}
\item{process}{Process: VAR}
\item{type}{include constant term (\code{const}) or not (\code{none})}
\item{call}{Matched call}
\item{coefficients}{Coefficient Matrix}
\item{fitted.values}{Fitted response values}
\item{residuals}{Residuals}
}
}
\description{
This function fits VAR(p) using OLS method.
}
\details{
This package specifies VAR(p) model as
\deqn{Y_{t} = c + B_1 Y_{t - 1} + \cdots + B_p Y_{t - p} + \epsilon_t}

If \code{include_type = TRUE}, there is \eqn{c} term.
Otherwise (\code{include_type = FALSE}), there is no \eqn{c} term.
The function estimates every coefficient matrix \eqn{c, B_1, \ldots, B_p}.
\itemize{
\item \code{\link[=build_y0]{build_y0()}} gives response matrix, \eqn{Y_0}.
\item \code{\link[=build_design]{build_design()}} gives design matrix, \eqn{X_0}.
\item Coefficient matrix is the form of \eqn{B = [B_1, B_2, \ldots, B_p, c]^T}.
}

Then perform least squares to the following

\deqn{Y_0 = X_0 B + error}

which gives

\deqn{\hat{B} = (X_0^T X_0)^{-1} X_0^T Y_0}
}
\examples{
# Perform the function using etf_vix dataset
\dontrun{
  fit <- var_lm(y = etf_vix, p = 5)
  class(fit)
  str(fit)
}

# Extract coef, fitted values, and residuals
\dontrun{
  coef(fit)
  residuals(fit)
  fitted(fit)
}

}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}
}
\seealso{
\itemize{
\item \code{\link[=build_y0]{build_y0()}} and \code{\link[=build_design]{build_design()}} to define Y0 and X0 matrix.
\item \code{\link[=estimate_var]{estimate_var()}} to compute coefficient VAR matrix.
\item Other package \code{\link[vars:VAR]{vars::VAR()}} is famous in VAR modeling.
}
}
