% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlse.R, R/print-varlse.R
\name{var_lm}
\alias{var_lm}
\alias{print.varlse}
\alias{knit_print.varlse}
\title{Fit Vector Autoregressive Model of Order p Model}
\usage{
var_lm(y, p)

\method{print}{varlse}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.varlse(x, ...)
}
\arguments{
\item{y}{Time series data of which columns indicate the variables}

\item{p}{integer, lags of VAR}

\item{x}{\code{varlse} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{var_lm} returns an object named \code{varlse} \link{class}.

It is a list with the following components:

\describe{
\item{design}{\eqn{X_0}}
\item{y0}{\eqn{Y_0}}
\item{y}{Raw input}
\item{p}{Lag of VAR}
\item{m}{Dimension of the data}
\item{df}{Numer of Coefficients: mp + 1}
\item{obs}{Sample size used when training = \code{totobs} - \code{p}}
\item{totobs}{Total number of the observation}
\item{process}{Process: VAR}
\item{call}{Matched call}
\item{coefficients}{Coefficient Matrix}
\item{fitted.values}{Fitted response values}
\item{residuals}{Residuals}
}
}
\description{
This function fits VAR(p) using OLS method
}
\details{
For VAR(p) model
\deqn{Y_{t} = c + B_1 Y_{t - 1} + \cdots + B_p Y_{t - p} + \epsilon_t}
the function gives basic values.
}
\examples{
# Perform the function using etf_vix dataset
\dontrun{
  fit <- var_lm(y = etf_vix, p = 5)
  class(fit)
  str(fit)
}

# Extract coef, fitted values, and residuals
\dontrun{
  coef(fit)
  residuals(fit)
  fitted(fit)
}

}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}

Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}
}
\seealso{
\code{\link{build_y0}} and \code{\link{build_design}} for defining Y0 and X0 matrix,

and \code{\link{estimate_var}} for computing coefficient VAR matrix.

Other package \code{\link[vars]{VAR}} is famous in VAR modeling.
}
\author{
Young Geun Kim \email{dudrms33@g.skku.edu}
}
