% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-varlse.R, R/varlse.R
\name{print.varlse}
\alias{print.varlse}
\alias{var_lm}
\title{Fit VAR(p)}
\usage{
\method{print}{varlse}(x, digits = max(3L, getOption("digits") - 3L), ...)

var_lm(y, p)
}
\arguments{
\item{y}{matrix, Time series data of which columns indicate the variables}

\item{p}{integer, lags of VAR}
}
\value{
\code{varlse} object
}
\description{
\code{var_lm} fits VAR(p) using OLS method
}
\details{
For VAR(p) model
\deqn{Y_{t} = B_1 Y_{t - 1} + \cdots + B_p Y_{t - p} + \epsilon_t}
the function gives basic values.
}
