% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlse.R, R/print-varlse.R
\name{var_lm}
\alias{var_lm}
\alias{print.varlse}
\alias{knit_print.varlse}
\title{Fit VAR(p)}
\usage{
var_lm(y, p)

\method{print}{varlse}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.varlse(x, ...)
}
\arguments{
\item{y}{matrix, Time series data of which columns indicate the variables}

\item{p}{integer, lags of VAR}

\item{x}{\code{varlse} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{varlse} \link{class} with
\item{\code{design}}{\eqn{X_0}}
\item{\code{y0}}{\eqn{Y_0}}
\item{\code{y}}{raw input}
\item{\code{p}}{lag of VAR: p}
\item{\code{m}}{Dimension of the data}
\item{\code{obs}}{Sample size used when training = \code{totobs} - \code{p}}
\item{\code{totobs}}{Total number of the observation}
\item{\code{process}}{Process: VAR}
\item{\code{call}}{Matched call}
\item{\code{coefficients}}{Coefficient Matrix}
\item{\code{fitted.values}}{Fitted response values}
\item{\code{residuals}}{Residuals}
}
\description{
The function fits VAR(p) using OLS method
}
\details{
For VAR(p) model
\deqn{Y_{t} = c + B_1 Y_{t - 1} + \cdots + B_p Y_{t - p} + \epsilon_t}
the function gives basic values.
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}

Litterman, R. B. (1986). \emph{Forecasting with Bayesian Vector Autoregressions: Five Years of Experience}. Journal of Business & Economic Statistics, 4(1), 25. \url{https://doi:10.2307/1391384}

Bańbura, M., Giannone, D., & Reichlin, L. (2010). \emph{Large Bayesian vector auto regressions}. Journal of Applied Econometrics, 25(1). \url{https://doi:10.1002/jae.1137}
}
\seealso{
\code{\link{build_y0}} and \code{\link{build_design}} for defining Y0 and X0 matrix,

and \code{\link{estimate_var}} for computing coefficient VAR matrix.

Other package \code{\link[vars]{VAR}} is famous in VAR modeling.
}
