% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-models.R
\name{var_vec_formulation}
\alias{var_vec_formulation}
\title{Vectorized OLS Formulation}
\description{
This page specifies the OLS formulation, which is vectorized of \link{var_design_formulation}.
Notation and format here will be used in this entire package document.
}
\section{Vector Autoregressive Model}{
Recall k-dim VAR model \eqn{Y_0 = X_0 A + Z_0}.
Applying \eqn{vec} operation, we have

\deqn{vec(Y_0) = (I_k \otimes X_0) vec(A) + vec(Z_0)}

Estimating \eqn{\alpha = vec(A)} is equivalent to estimating usual OLS.
}

\section{Vector Heterogeneous Autoregressive Model}{
Recall k-dim VHAR model \deqn{Y_0 = X_1 \Phi + Z_0 = (X_0 C_{HAR}^\intercal) \Phi + Z_0}.
Then
\deqn{vec(Y_0) = (I_k \otimes X_0 C_{HAR}^\intercal) vec(\Phi) + vec(Z_0) = (I_k \otimes X_1) vec(\Phi) + vec(Z_0)}
}

\section{Restrictions}{
\eqn{\phi = vec(\Phi)} is linear restriction of \eqn{\alpha = vec(A)}.
Then
\deqn{\alpha = vec(C_{HAR}^\intercal \Phi) = (I_k \otimes C_{HAR}^\intercal) \phi}
}

\section{Restrictions for Sparsity}{
Value of 0, 1 corresponding to \eqn{A_{i,j}} gives sparsity of the element.
Let \eqn{\psi = (\psi_1, \ldots, \psi_{k^2 p})^\intercal} where \eqn{\psi_j = 0,1},
and let \eqn{\Psi = diag(\psi) \in \mathbb{R}^{k^2 p \times k^2 p}}.

Then the new VAR coefficient vector \eqn{\theta = \Psi \alpha \in \mathbb{R}^{k^2 p}} becomes
\eqn{\alpha_j = 0} if \eqn{\psi_j = 0} and \eqn{\alpha_j \neq 0} if \eqn{\psi = 1}.
}

\references{
Jochmann, M., Koop, G., & Strachan, R. W. (2010). \emph{Bayesian forecasting using stochastic search variable selection in a VAR subject to breaks}. International Journal of Forecasting, 26(2), 326–347. doi:\href{https://doi.org/10.1016/j.ijforecast.2009.11.002}{10.1016/j.ijforecast.2009.11.002}

Korobilis, D. (2013). \emph{VAR FORECASTING USING BAYESIAN VARIABLE SELECTION}. Journal of Applied Econometrics, 28(2). doi:\href{https://doi.org/10.1002/jae.1271}{10.1002/jae.1271}

Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. doi:\href{https://doi.org/10.1007/978-3-540-27752-1}{10.1007/978-3-540-27752-1}
}
\keyword{internal}
