% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vharlse.R, R/print-vharlse.R
\name{vhar_lm}
\alias{vhar_lm}
\alias{print.vharlse}
\alias{knit_print.vharlse}
\title{Fit Vector HAR}
\usage{
vhar_lm(y)

\method{print}{vharlse}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.vharlse(x, ...)
}
\arguments{
\item{y}{matrix, Time series data of which columns indicate the variables}

\item{x}{\code{vharlse} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{vharlse} \link{class} with
\item{\code{design}}{\eqn{X_0}}
\item{\code{y0}}{\eqn{Y_0}}
\item{\code{y}}{raw input}
\item{\code{m}}{Dimension of the data}
\item{\code{obs}}{Sample size used when training = \code{totobs} - \code{p}}
\item{\code{totobs}}{Total number of the observation}
\item{\code{process}}{Process: VHAR}
\item{\code{call}}{Matched call}
\item{\code{coefficients}}{Coefficient Matrix}
\item{\code{fitted.values}}{Fitted response values}
\item{\code{residuals}}{Residuals}
}
\description{
This function fits VHAR using OLS method
}
\details{
For VHAR model
\deqn{Y_{t} = Phi^{(d)} Y_{t - 1} + \Phi^{(w)} Y_{t - 1}^{(w)} + \Phi^{(m)} Y_{t - 1}^{(m)} + \epsilon_t}
the function gives basic values.
}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}

Corsi, F. (2008). \emph{A Simple Approximate Long-Memory Model of Realized Volatility}. Journal of Financial Econometrics, 7(2), 174–196. \url{https://doi:10.1093/jjfinec/nbp001}
}
