% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vharlse.R, R/print-vharlse.R
\name{vhar_lm}
\alias{vhar_lm}
\alias{print.vharlse}
\alias{knit_print.vharlse}
\title{Fit Vector HAR Model}
\usage{
vhar_lm(y, include_mean = TRUE)

\method{print}{vharlse}(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.vharlse(x, ...)
}
\arguments{
\item{y}{Time series data of which columns indicate the variables}

\item{include_mean}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Add constant term (Default: \code{TRUE}) or not (\code{FALSE})}

\item{x}{\code{vharlse} object}

\item{digits}{digit option to print}

\item{...}{not used}
}
\value{
\code{vhar_lm} returns an object named \code{vharlse} \link{class}.
It is a list with the following components:

\describe{
\item{coefficients}{Coefficient Matrix}
\item{fitted.values}{Fitted response values}
\item{residuals}{Residuals}
\item{covmat}{LS estimate for covariance matrix}
\item{df}{Numer of Coefficients: 3m + 1 or 3m}
\item{p}{3 (The number of terms. \code{vharlse} contains this element for usage in other functions.)}
\item{m}{Dimension of the data}
\item{obs}{Sample size used when training = \code{totobs} - \code{p}}
\item{totobs}{Total number of the observation}
\item{call}{Matched call}
\item{process}{Process: VHAR}
\item{type}{include constant term (\code{const}) or not (\code{none})}
\item{y0}{\eqn{Y_0}}
\item{design}{\eqn{X_0}}
\item{y}{Raw input}
}
}
\description{
This function fits VHAR using OLS method.
}
\details{
For VHAR model
\deqn{Y_{t} = \Phi^{(d)} Y_{t - 1} + \Phi^{(w)} Y_{t - 1}^{(w)} + \Phi^{(m)} Y_{t - 1}^{(m)} + \epsilon_t}
the function gives basic values.
}
\examples{
# Perform the function using etf_vix dataset
\dontrun{
  fit <- vhar_lm(y = etf_vix)
  class(fit)
  str(fit)
}

# Extract coef, fitted values, and residuals
\dontrun{
  coef(fit)
  residuals(fit)
  fitted(fit)
}

}
\references{
Lütkepohl, H. (2007). \emph{New Introduction to Multiple Time Series Analysis}. Springer Publishing. \url{https://doi.org/10.1007/978-3-540-27752-1}

Corsi, F. (2008). \emph{A Simple Approximate Long-Memory Model of Realized Volatility}. Journal of Financial Econometrics, 7(2), 174–196. \url{https://doi:10.1093/jjfinec/nbp001}

Baek, C. and Park, M. (2021). \emph{Sparse vector heterogeneous autoregressive modeling for realized volatility}. J. Korean Stat. Soc. 50, 495–510. \url{https://doi.org/10.1007/s42952-020-00090-5}
}
\seealso{
\code{\link[=scale_har]{scale_har()}} to linear transformation for VHAR,

\code{\link[=estimate_har]{estimate_har()}} to compute coefficient VHAR matrix,

and \code{\link[=estimate_har_none]{estimate_har_none()}} to compute coefficient VHAR matrix without constant term.
}
